### 2024-03-21 #6520 @editorialbot
[PRE REVIEW]: Aurora: An open-source python implementation of the EMTF package for magnetotelluric data processing using MTH5 and mt_metadata
**Submitting author:** @kkappler (<a href="http://orcid.org/0000-0002-1877-1255">karl kappler</a>)
**Repository:** https://github.com/simpeg/aurora
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.3.12
