+++ _sources/api/transfer_function.rst.txt
@@ -1,0 +1,77 @@
+.. _transfer_function:
+
+Transfer Function
+-----------------
+
+Plot
+****
+
+.. automodule:: aurora.transfer_function.plot
+    :show-inheritance:
+    :members:
+    :undoc-members:
+
+
+Regression
+**********
+
+.. automodule:: aurora.transfer_function.regression
+    :show-inheritance:
+    :members:
+    :undoc-members:
+
+
+Base
+****
+
+.. automodule:: aurora.transfer_function.base
+    :show-inheritance:
+    :members:
+    :undoc-members:
+
+
+EMTF Z File Helpers
+*******************
+
+.. automodule:: aurora.transfer_function.emtf_z_file_helpers
+    :show-inheritance:
+    :members:
+    :undoc-members:
+
+
+Iter Control
+************
+
+.. automodule:: aurora.transfer_function.regression.iter_control
+    :show-inheritance:
+    :members:
+    :undoc-members:
+
+
+Transfer Function Collection
+****************************
+
+.. automodule:: aurora.transfer_function.transfer_function_collection
+    :show-inheritance:
+    :members:
+    :undoc-members:
+
+
+.. Transfer Function Header
+.. ************************
+
+.. .. automodule:: aurora.transfer_function.transfer_function_header
+..     :show-inheritance:
+..     :members:
+..     :undoc-members:
+
+
+TTFZ
+****
+
+.. automodule:: aurora.transfer_function.TTFZ
+    :show-inheritance:
+    :members:
+    :undoc-members:
+
+
