+++ _sources/api/time_series.rst.txt
@@ -1,0 +1,95 @@
+.. _time-series:
+
+Time Series
+-----------
+
+Filters
+*******
+
+.. automodule:: aurora.time_series.filters
+    :show-inheritance:
+    :members:
+    :undoc-members:
+
+
+Apodization Window
+******************
+
+.. automodule:: aurora.time_series.apodization_window
+    :show-inheritance:
+    :members:
+    :undoc-members:
+
+
+Decorators
+**********
+
+.. automodule:: aurora.time_series.decorators
+    :show-inheritance:
+    :members:
+    :undoc-members:
+
+
+.. Frequency Band
+.. **************
+
+.. .. automodule:: aurora.time_series.frequency_band
+..     :show-inheritance:
+..     :members:
+..     :undoc-members:
+
+
+Frequency Band Helpers
+**********************
+
+.. automodule:: aurora.time_series.frequency_band_helpers
+    :show-inheritance:
+    :members:
+    :undoc-members:
+
+
+.. Frequency Domain Helpers
+.. ************************
+
+.. .. automodule:: aurora.time_series.frequency_domain_helpers
+..     :show-inheritance:
+..     :members:
+..     :undoc-members:
+
+
+Time Axis Helpers
+*****************
+
+.. automodule:: aurora.time_series.time_axis_helpers
+    :show-inheritance:
+    :members:
+    :undoc-members:
+
+
+Window Helpers
+**************
+
+.. automodule:: aurora.time_series.window_helpers
+    :show-inheritance:
+    :members:
+    :undoc-members:
+
+
+Windowed Time Series
+********************
+
+.. automodule:: aurora.time_series.windowed_time_series
+    :show-inheritance:
+    :members:
+    :undoc-members:
+
+
+Windowing Scheme
+****************
+
+.. automodule:: aurora.time_series.windowing_scheme
+    :show-inheritance:
+    :members:
+    :undoc-members:
+
+
