+++ _sources/examples/dataset_definition.ipynb.txt
@@ -1,0 +1,2255 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "id": "e73ab089-fc34-4d5a-a88c-4e43d5874ab7",
+   "metadata": {},
+   "source": [
+    "# Dataset Definition DataFrame\n",
+    "\n",
+    "For easier interaction a user can input a dataframe of runs to process, this is a simple example of how to do that."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "id": "af5fa6bb-7f64-47fb-a6a7-8cd6d116546f",
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "import pandas as pd\n",
+    "from mt_metadata.transfer_functions.processing.aurora import Processing"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "id": "03cab6f9-6d03-4483-9199-663d3fda922c",
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "starts = [\"2020-01-01T00:00:00\", \"2020-02-02T00:00:00\"]\n",
+    "ends = [\"2020-01-31T12:00:00\", \"2020-02-28T12:00:00\"]\n",
+    "\n",
+    "data_list = []\n",
+    "\n",
+    "for ii in range(3):\n",
+    "    for start, end in zip(starts, ends):\n",
+    "        entry = {\n",
+    "            \"station_id\": \"mt01\",\n",
+    "            \"run_id\": f\"{ii:03}\",\n",
+    "            \"start\": start,\n",
+    "            \"end\": end,\n",
+    "            \"mth5_path\": r\"/home/mth5_path.h5\" ,\n",
+    "            \"sample_rate\": 10,\n",
+    "            \"input_channels\": [\"hx\", \"hy\"],\n",
+    "            \"output_channels\": [\"hz\", \"ex\", \"ey\"],\n",
+    "            \"remote\": False\n",
+    "        }\n",
+    "        \n",
+    "        data_list.append(entry)\n",
+    "        \n",
+    "        rr_entry_01 = {\n",
+    "            \"station_id\": \"rr01\",\n",
+    "            \"run_id\": f\"{ii:03}\",\n",
+    "            \"start\": start,\n",
+    "            \"end\": end,\n",
+    "            \"mth5_path\": r\"/home/mth5_path.h5\" ,\n",
+    "            \"sample_rate\": 10,\n",
+    "            \"input_channels\": [\"hx\", \"hy\"],\n",
+    "            \"output_channels\": [\"hz\", \"ex\", \"ey\"],\n",
+    "            \"remote\": True\n",
+    "        }\n",
+    "        data_list.append(rr_entry_01)\n",
+    "        \n",
+    "        rr_entry_02 = {\n",
+    "            \"station_id\": \"rr02\",\n",
+    "            \"run_id\": f\"{ii:03}\",\n",
+    "            \"start\": start,\n",
+    "            \"end\": end,\n",
+    "            \"mth5_path\": r\"/home/mth5_path.h5\" ,\n",
+    "            \"sample_rate\": 10,\n",
+    "            \"input_channels\": [\"hx\", \"hy\"],\n",
+    "            \"output_channels\": [\"hz\", \"ex\", \"ey\"],\n",
+    "            \"remote\": True\n",
+    "        }\n",
+    "        data_list.append(rr_entry_02)\n",
+    "\n",
+    "\n",
+    "sdf = pd.DataFrame(data_list)\n",
+    "sdf.start = pd.to_datetime(sdf.start)\n",
[…]
