+++ .github/workflows/tests.yml
@@ -50,8 +50,8 @@ jobs:
         python -m ipykernel install --user --name aurora-test
         # Install any other dependencies you need
 
-        - name: Execute Jupyter Notebooks
-        run: |
+    - name: Execute Jupyter Notebooks
+      run: |
         jupyter nbconvert --to notebook --execute docs/examples/dataset_definition.ipynb
         jupyter nbconvert --to notebook --execute docs/examples/make_cas04_single_station_h5.ipynb
         jupyter nbconvert --to notebook --execute docs/examples/operate_aurora.ipynb
