+++ aurora/transfer_function/regression/base.py
@@ -135,11 +135,21 @@ class RegressionEstimator(object):
         """
         20210806
         This method was originally in TRME.m, but it does not depend in
-        general on using RME method so I am putting it in the base class.
+        general on using RME method so putting it in the base class.
 
-        We basically never get here and when we do we dont trust the results
+        We basically never get here and when we do, we don't trust the results
         https://docs.scipy.org/doc/numpy-1.9.2/reference/generated/numpy.linalg.svd.html
         https://www.mathworks.com/help/matlab/ref/double.svd.html
+
+        Note that the svd outputs are different in matlab and numpy
+        https://stackoverflow.com/questions/50930899/svd-command-in-python-v-s-matlab
+        "The SVD of a matrix can be written as
+
+        A = U S V^H
+
+        Where the ^H signifies the conjugate transpose. Matlab's svd command returns U, S and V,
+        while numpy.linalg.svd returns U, the diagonal of S, and V^H.
+        Thus, to get the same S and V as in Matlab you need to reconstruct the S and also get the V:
 
         <ORIGINAL MATLAB>
             <COMMENT>
@@ -167,7 +177,7 @@ class RegressionEstimator(object):
         """
         U, s, V = np.linalg.svd(self.X, full_matrices=False)
         S_inv = np.diag(1.0 / s)
-        self.b = (V.T @ S_inv @ U.T) * self.Y
+        self.b = (V.T.conj() @ S_inv @ U.T.conj()) * self.Y
         if self.iter_control.return_covariance:
             logger.warning(
                 "Warning covariances are not xarray, may break things downstream"
