### 2024-03-26 567963a
Pretty crufty commit with updates to coherence weights
- add estimate_simple_coherence method to spectral features
- add cross_powers attr and cross_power() method to Spectrogram
- uncomment coherence_weights

This is messy not only because of uncommenting -- there are also
some complications with converting the coherences to weights.
Namely, that
1. the weights may want to have different values depending on the output
   channel being estimated, and the current logic is set up to apply the
   weights before the loop over output channel
2. Moving the weighting inside the output loop channel is
[…]
