+++ aurora/transfer_function/weights/coherence_weights.py
@@ -1,710 +1,736 @@
-# """
-# There are several ways to go about this.  One reasonably straightforward one is:
-#
-# Take all the Fourier coefficients right before they go into regression (right before or
-# after we apply the effective degrees of freedom weights).  Data are strictly 2D at
-# this point, you have each channel of data corresponding to a column, and each row
-# to an observation.
-# Following Jones and Jodicke, lets take the coherence of Ex and Hy
-# and that of Ey and Hx.  This yields one single number.
-#
-# Then, iteratively remove each observation in turn and count whether that partial
-# coherence is bigger or smaller than the baseline and by how much
-#
-# Deflection downward means naughty data and deflection upward means well-behaved data.
-#
-# This leads to a coherence score for each observation.  We can boot out all data that
-#  say move the estimate down (worse than average)
-# """
-#
-# import numpy as np
-#
-# from aurora.time_series.frequency_band_helpers import adjust_band_for_coherence_sorting
-# from aurora.time_series.spectrogram import Spectrogram
-# from aurora.transfer_function.weights.spectral_features import (
-#     estimate_multiple_coherence,
-# )
-# from aurora.transfer_function.weights.spectral_features import (
-#     estimate_time_series_of_impedances,
-# )
-# from collections import namedtuple
-# from loguru import logger
-#
-# # convenience class used in coherence calculations
-# # not to be confused with Channel classes in mt_metdata and mth5
-# CoherenceChannel = namedtuple("Channel", ["local_or_remote", "component"])
-#
-#
-# def simple_coherence_channel_pairs():
-#     # define conjugate channel pairs
-#     paired_channels = {}
-#     paired_channels["local"] = {}
-#     paired_channels["local"]["ex"] = CoherenceChannel("local", "hy")
-#     paired_channels["local"]["ey"] = CoherenceChannel("local", "hx")
-#     paired_channels["remote"] = {}
-#     paired_channels["remote"]["hx"] = CoherenceChannel("local", "hx")
-#     paired_channels["remote"]["hy"] = CoherenceChannel("local", "hy")
-#     return paired_channels
-#
-#
-# def coherence_from_fc_series(c_xy, c_xx, c_yy):
-#     # n_obs = len(c_xy)
-#     num = np.sum(c_xy) ** 2
-#     den = np.sum(c_xx) * np.sum(c_yy)
-#     coh = num / den
-#     return coh
-#
-#
-# def jackknife_coherence_weights(
-#     x, y, lower_quantile_cutoff=0.3, upper_quantile_cutoff=0.95
-# ):  # 975):#0.98
-#     """
-#     Follow the convention of Bendat & Perisol (1980):
-#     "The two-sided spectral density function between two random processes is defined
-#     using X*Y and _not_ XY*.
-#
-#
-#     Note 1: Extremely high coherence can be due to noise
-#     Consider ways to pre-filter those events before this is called.
-#     - That may need a min_fraction_to_keep=0.2-0.8, which guards against
-#     generally poor data being completely discarded.
-#     - Consider variations on coh-sorting threshold_min, and threshold_max
-#     near perfect coherency can be expected when coherent noise is present
-#     - 95% seems extreme for a threshold, maybe not yielding enogh
-#     observations for robust statistics.
-#
-#     Note2: 2024-09-10: vectorization
[…]
