+++ aurora/time_series/spectrogram.py
@@ -1,4 +1,7 @@
+import xarray as xr
+
 from aurora.time_series.frequency_band_helpers import extract_band
+from loguru import logger
 
 
 class Spectrogram(object):
@@ -11,12 +14,26 @@ class Spectrogram(object):
     """
 
     def __init__(self, dataset=None):
+        """
+
+        Parameters
+        ----------
+        dataset: xarray.core.dataset.Dataset
+            Chunk of a STFT, having dimensions time and frequency
+        """
         self._dataset = dataset
         self._frequency_increment = None
+        self._cross_powers = None
 
     @property
     def dataset(self):
         return self._dataset
+
+    @property
+    def cross_powers(self):
+        if self._cross_powers is None:
+            self._cross_powers = xr.Dataset(coords={"time": self.dataset.time.data})
+        return self._cross_powers
 
     @property
     def time_axis(self):
@@ -34,12 +51,14 @@ class Spectrogram(object):
 
         Parameters
         ----------
-        band
-        stft_obj
-
+        band: mt_metadata.transfer_functions.processing.aurora.band.Band
+            The frequency band object
+        epsilon: float
+            Prevents numeric round-off errors checking harmonics are in band
         Returns
         -------
-
+        num_harmonics: int
+            The number of frequency bins within the frequency band
         """
         cond1 = self._dataset.frequency >= frequency_band.lower_bound - epsilon
         cond2 = self._dataset.frequency <= frequency_band.upper_bound + epsilon
@@ -67,5 +86,41 @@ class Spectrogram(object):
         spectrogram = Spectrogram(dataset=extracted_band_dataset)
         return spectrogram
 
-    def cross_powers(self, ch1, ch2, band=None):
-        pass
+    def cross_power(self, ch1, ch2, augment=True):
+        """
+        TODO: This method should normally be used on a Spectrogram that has
+        been extracted (a tf_estimation band).  Consider extending
+        the class or somehow marking it that the band extraction has happened
+        TODO: The cross power should get put into its own xr.Dataset, not into the spectrogram
+
+        Convention:
+        "The two-sided spectral density function between two random processes is defined
+        using X*Y and _not_ YX*" -- Bendat & Perisol (1980, p54, Eqn 3.45).
+
+        Parameters
+        ----------
+        ch1: str
+            The channel name of the first channel in the cross-spectrum
+        ch2: str
+            The channel name of the second channel in the cross-spectrum
+
+
+        Returns
+        -------
+
+        """
+        name = f"{ch1}_{ch2}"
+        try:
+            return self.cross_powers[name]
+        except KeyError:
+            msg = f"did not find stored cross power {name}; Computing {name}"
+            logger.info(msg)
+        X = self.dataset[ch1]
+        Y = self.dataset[ch2]
+        xHy = (X.conj().transpose() * Y).sum(dim="frequency")
+        if ch1 == ch2:
+            xHy = xHy.__abs__()
+        if augment:
+            name = f"{ch1}_{ch2}"
+            self.cross_powers[name] = xHy
+        return xHy
