### 2024-03-26 f80296a
Move get_regression_estimator into regression
- This was previously declared in transfer_function_helpers
