### 2024-03-27 579ed8a
merge remote channels into STFT object
- Now use a single xarray as input to RR tf estimation
- move check_time_axes_synched into xarray_helpers
- deprecate get_band_for_tf_estimate
