### 2024-03-27 6ca0a73
suppress jupyter notebooks
