+++ .github/workflows/tests.yml
@@ -43,23 +43,23 @@ jobs:
         pip install -e .
         conda list
         
-    - name: Install Jupyter and dependencies
-      run: |
-        pip install jupyter
-        pip install ipykernel
-        python -m ipykernel install --user --name aurora-test
-        # Install any other dependencies you need
-
-    - name: Execute Jupyter Notebooks
-      run: |
-        jupyter nbconvert --to notebook --execute docs/examples/dataset_definition.ipynb
-        jupyter nbconvert --to notebook --execute docs/examples/make_cas04_single_station_h5.ipynb
-        jupyter nbconvert --to notebook --execute docs/examples/operate_aurora.ipynb
-        jupyter nbconvert --to notebook --execute tests/test_run_on_commit.ipynb
-        jupyter nbconvert --to notebook --execute tutorials/pole_zero_fitting/lemi_pole_zero_fitting_example.ipynb
-        jupyter nbconvert --to notebook --execute tutorials/processing_configuration.ipynb
-        jupyter nbconvert --to notebook --execute tutorials/synthetic_data_processing.ipynb
-        # Replace "notebook.ipynb" with your notebook's filename
+#    - name: Install Jupyter and dependencies
+#      run: |
+#        pip install jupyter
+#        pip install ipykernel
+#        python -m ipykernel install --user --name aurora-test
+#        # Install any other dependencies you need
+#
+#    - name: Execute Jupyter Notebooks
+#      run: |
+#        jupyter nbconvert --to notebook --execute docs/examples/dataset_definition.ipynb
+#        jupyter nbconvert --to notebook --execute docs/examples/make_cas04_single_station_h5.ipynb
+#        jupyter nbconvert --to notebook --execute docs/examples/operate_aurora.ipynb
+#        jupyter nbconvert --to notebook --execute tests/test_run_on_commit.ipynb
+#        jupyter nbconvert --to notebook --execute tutorials/pole_zero_fitting/lemi_pole_zero_fitting_example.ipynb
+#        jupyter nbconvert --to notebook --execute tutorials/processing_configuration.ipynb
+#        jupyter nbconvert --to notebook --execute tutorials/synthetic_data_processing.ipynb
+#        # Replace "notebook.ipynb" with your notebook's filename
 
 #    - name: Commit changes (if any)
 #      run: |
