### 2024-03-28 #6520 @kthyng
@kkappler some comments/questions:
* Please look through the DOI notes above to see if anything needs to be addressed
* Also look at the error generating the paper, to see if you can fix it
* I recommend getting your tutorials into your docs so everything is easier to find
