+++ paper/paper.bib
@@ -5,6 +5,7 @@
   volume={171},
   pages={105270},
   year={2023},
+  doi={10.1016/j.cageo.2022.105270},
   publisher={Elsevier}
 }
 
@@ -23,6 +24,7 @@
   author={Cockett, Rowan and Kang, Seogi and Heagy, Lindsey J and Pidlisecky, Adam and Oldenburg, Douglas W},
   journal={Computers \& Geosciences},
   year={2015},
+  doi={10.1016/j.cageo.2015.09.015},
   publisher={Elsevier}
 }
 
@@ -41,6 +43,7 @@
   number={2},
   pages={475--496},
   year={1997},
+  doi={10.1111/j.1365-246x.1997.tb05663.x},
   publisher={Blackwell Publishing Ltd Oxford, UK}
 }
 
@@ -61,6 +64,7 @@
   number={2-3},
   pages={207--249},
   year={2002},
+  doi={10.1023/A:1015012821040},
   publisher={Springer}
 }
 
@@ -72,6 +76,7 @@
   number={3},
   pages={555--605},
   year={2018},
+  doi={10.1111/rssa.12315},
   publisher={Oxford University Press}
 }
 
@@ -83,6 +88,7 @@
   number={1},
   pages={F1--F17},
   year={2020},
+  doi={10.1190/geo2018-0679.1},
   publisher={Society of Exploration Geophysicists}
 }
 
@@ -94,6 +100,7 @@
   number={3},
   pages={1853--1872},
   year={2019},
+  doi={10.1093/gji/ggz197},
   publisher={Oxford University Press}
 }
 
@@ -104,6 +111,7 @@
   volume={8},
   pages={IGY6--IGY20},
   year={2010},
+  doi={10.2481/dsj.ss_igy-009},
   publisher={CODATA}
 }
 
@@ -121,67 +129,19 @@
   volume={8},
   pages={296},
   year={2020},
+  doi={10.3389/feart.2020.00296},
   publisher={Frontiers Media SA}
 }
 
 @article{Vozoff:1991,
   	url = {https://pubs.geoscienceworld.org/seg/books/book/2087/chapter-abstract/114406941/THE-MAGNETOTELLURIC-METHOD?redirectedFrom=fulltext},
   	Author = {{Vozoff}, K.},
-    Booktitle = {Galactic Dynamics: Second Edition, by James Binney and Scott Tremaine.~ISBN 978-0-691-13026-2 (HB).~Published by Princeton University Press, Princeton, NJ USA, 2008.},
+    Booktitle = {Electromagnetic Methods in Applied Geophysics: Volume 2, Application, Parts A and B. Editted by MIsac N. Nabighian  ~ISBN 1-560800-22-4. ~Published by SEG, Tulsa, OK USA, 1991.},
   	Title = {{The Magnetotelluric Method}},
   	Year = 1991
 }
 
-@article{Pearson:2017,
-  	url = {http://adsabs.harvard.edu/abs/2017arXiv170304627P},
-  	Archiveprefix = {arXiv},
-  	Author = {{Pearson}, S. and {Price-Whelan}, A.~M. and {Johnston}, K.~V.},
-  	Eprint = {1703.04627},
-  	Journal = {ArXiv e-prints},
-  	Keywords = {Astrophysics - Astrophysics of Galaxies},
-  	Month = mar,
-  	Title = {{Gaps in Globular Cluster Streams: Pal 5 and the Galactic Bar}},
-  	Year = 2017
-}
-
-@book{Binney:2008,
-  	url = {http://adsabs.harvard.edu/abs/2008gady.book.....B},
-  	Author = {{Binney}, J. and {Tremaine}, S.},
[…]
