### 2024-03-30 1b74a68
Move tutorials under docs
    - suggested by JOSS editor
    - uncomment notebook tests to see if this passes
