+++ .github/workflows/tests.yml
@@ -56,9 +56,9 @@ jobs:
         jupyter nbconvert --to notebook --execute docs/examples/make_cas04_single_station_h5.ipynb
         jupyter nbconvert --to notebook --execute docs/examples/operate_aurora.ipynb
         jupyter nbconvert --to notebook --execute tests/test_run_on_commit.ipynb
-        jupyter nbconvert --to notebook --execute tutorials/pole_zero_fitting/lemi_pole_zero_fitting_example.ipynb
-        jupyter nbconvert --to notebook --execute tutorials/processing_configuration.ipynb
-        jupyter nbconvert --to notebook --execute tutorials/synthetic_data_processing.ipynb
+        jupyter nbconvert --to notebook --execute docs/tutorials/pole_zero_fitting/lemi_pole_zero_fitting_example.ipynb
+        jupyter nbconvert --to notebook --execute docs/tutorials/processing_configuration.ipynb
+        jupyter nbconvert --to notebook --execute docs/tutorials/synthetic_data_processing.ipynb
         # Replace "notebook.ipynb" with your notebook's filename
 
 #    - name: Commit changes (if any)
