### 2024-03-30 #6520 @kkappler
Hi Kristen,

Thank you for the prompts,

I pushed on the `joss` branch with the following updates:

   - DOIs updated -- I put some notes in the postscript of this message.
   - Error in pdf fixed
   - moved tutorials into doc (on PR branch `patches` and merged `patches`
   into joss/ branch)

Thanks,

Karl
P.S. DOI Notes below:
__________________
MISSING DOIs

- 10.1016/j.cageo.2022.105270 may be a valid DOI for title: SigMT: An
open-source Python package for magnetote...
added

- No DOI given, and none found for title: BIRRP: Bounded influence, remote
reference process...
DOI also not found for this legacy software

- 10.1016/j.cageo.2015.09.015 may be a valid DOI for title: SimPEG: An open
source framework for simulation an...
added

- No DOI given, and none found for title: Continuous integration: improving
software quality...
DOI also not found for this book

- 10.1111/j.1365-246x.1997.tb05663.x may be a valid DOI for title: Robust
multiple-station magnetotelluric data proce...
added

- No DOI given, and none found for title: Mod3DMT and EMTF: Free Software
for MT Data Proces...
DOI also not found for this conference proceedings abstract

- No DOI given, and none found for title: Processing and interpretation of
electromagnetic i...
DOI found: 10.1023/A:1015012821040 may be a valid DOI for this review paper

- 10.1111/rssa.12315 may be a valid DOI for title: Statistical challenges
of administrative and trans...
added

- 10.1190/geo2018-0679.1 may be a valid DOI for title: EMTF XML: New data
interchange format and conversi...
added

- 10.1093/gji/ggz197 may be a valid DOI for title: An automated new
pre-selection tool for noisy Magn...
added

- 10.2481/dsj.ss_igy-009 may be a valid DOI for title: EMScope: A
continental scale magnetotelluric obser...
added

- No DOI given, and none found for title: Resistics: A versatile native
Python 3 package for...
DOI also not found for this Conference proceedings

- 10.3389/feart.2020.00296 may be a valid DOI for title: Razorback, an open
source Python library for robus...
added

- No DOI given, and none found for title: The Magnetotelluric Method
DOI found: 10.1190/1.9781560802686.ch8

- No DOI given, and none found for title: Gaps in Globular Cluster Streams:
Pal 5 and the Ga...
removed lingering reference from template

- No DOI given, and none found for title: Galactic Dynamics: Second Edition
removed lingering reference from template

- No DOI given, and none found for title: MTHotel
DOI also not found

On Thu, Mar 28, 2024 at 3:19 PM Kristen Thyng ***@***.***>
wrote:

> @kkappler <https://github.com/kkappler> some comments/questions:
>
>    - Please look through the DOI notes above to see if anything needs to
>    be addressed
>    - Also look at the error generating the paper, to see if you can fix it
>    - I recommend getting your tutorials into your docs so everything is
>    easier to find
>
