### 2024-04-01 #6520 @kthyng
@kkappler Ok
* DOIs look good
* maybe a new error in PDF — can you check it out above?
* Is there a place I can see readthedocs of your joss branch? Can't remember if those links are public.
