### 2024-04-01 #6520 @danielskatz
The PDF error is an invalid ORCID. ORCID lines in the paper.md should either be valid ORCIDs or should be removed.
