### 2024-04-01 7be0cd2
remove invalid orcid 0000-0000-0000-0000 from template
