### 2024-04-02 #6520 @kkappler
Hi Kristen,

The pdf error should be fixed on the latest push.

There are not prepared docs for the joss branch, but the latest publicly
available docs are here: https://simpeg.xyz/aurora/ .  Please note also
that a new release (3.13) is scheduled for this month, (PR already merged
into main (https://github.com/simpeg/aurora/pull/322) -- just awaiting one
PR/release on a dependency before we push to PyPI and condaforge.

best,
Karl

On Mon, Apr 1, 2024 at 1:49 PM Kristen Thyng ***@***.***>
wrote:

> @kkappler <https://github.com/kkappler> Ok
>
>    - DOIs look good
>    - maybe a new error in PDF — can you check it out above?
>    - Is there a place I can see readthedocs of your joss branch? Can't
>    remember if those links are public.
>
