### 2024-04-05 #6520 @kkappler
Thanks for the update and feedback.

We filtered for earth science and python processing and found these names:

- [Matteo Ravasi](https://reviewers.joss.theoj.org/reviewers/3831)	[@mrava87](https://github.com/mrava87)
- [Matt Hall](https://reviewers.joss.theoj.org/reviewers/3266)	[@kwinkunks](https://github.com/kwinkunks)
- [Moritz Makowski](https://reviewers.joss.theoj.org/reviewers/2706)	[@dostuffthatmatters](https://github.com/dostuffthatmatters)
- [@blsqr](https://reviewers.joss.theoj.org/reviewers/1611)	[@blsqr](https://github.com/blsqr)
- [@krischer](https://reviewers.joss.theoj.org/reviewers/169)	[@krischer](https://github.com/krischer)

And backup list:
- [Griffin Daniel Chure](https://reviewers.joss.theoj.org/reviewers/3235)	[@gchure](https://github.com/gchure)
- [Simon Filhol](https://reviewers.joss.theoj.org/reviewers/2883)	[@ArcticSnow](https://github.com/ArcticSnow)
- [@rmsare](https://reviewers.joss.theoj.org/reviewers/306)	[@rmsare](https://github.com/rmsare)
