+++ _sources/tutorials/processing_configuration.ipynb.txt
@@ -1,0 +1,1365 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "id": "8148a4bd-32e9-4551-b4bc-f45e27a51b31",
+   "metadata": {},
+   "source": [
+    "## Processing Configuration"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "226f7d10-dd15-4061-adc6-c68b448d60c0",
+   "metadata": {},
+   "source": [
+    "The processing\\_config (\"config\") together with the kernel dataset represent the key inputs to the processing pipeline.  The processing config is based on the `mt_metadata.base.Base` class.  This means it is a container that has a JSON or dictionary representation.\n",
+    "\n",
+    "It is the purpose of the config to encapsulate all the parameters required for processing.  There are many parameters and effort has been given to selecting some \"reasonable\" default values so that users do not need to worry about all of these parameters if they don't want to.\n",
+    "\n",
+    "The ProcessingConfig is expected to evolve with aurora as new functionalities become available.  This is one reason why such a genreic data srtuctrue was selected \n",
+    "In this tutorial, we will use the synthetic dataset example to show some of the features of the config object.\n",
+    "\n",
+    "Hopefully, it will be fairly easy to add other parameters to the config, such as:\n",
+    "- coherencey sorting \n",
+    "- polarization sorting\n",
+    "- ARMA prewhitening \n",
+    "- Other tools from the community\n",
+    "\n",
+    "\n",
+    "\n",
+    "There are two main ways one would normally build the object:\n",
+    "1. Use the ConfigCreator class\n",
+    "2. Edit existing config.json files\n",
+    "\n",
+    "But one can also initialize a Processing object, which is what is done inside ConfigCreator\n",
+    "\n"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "id": "fb9a2cb3-662b-4d5f-86cc-db54e98466f4",
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "from mt_metadata.transfer_functions.processing.aurora import Processing"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "id": "777182c2-b77f-44b8-a663-1c7afe00796e",
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "p = Processing()"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 3,
+   "id": "d9069a6b-b13e-4afb-8f11-12546eb2d035",
+   "metadata": {},
+   "outputs": [
+    {
+     "data": {
+      "text/plain": [
+       "{\n",
+       "    \"processing\": {\n",
+       "        \"channel_nomenclature.ex\": \"ex\",\n",
+       "        \"channel_nomenclature.ey\": \"ey\",\n",
+       "        \"channel_nomenclature.hx\": \"hx\",\n",
+       "        \"channel_nomenclature.hy\": \"hy\",\n",
+       "        \"channel_nomenclature.hz\": \"hz\",\n",
+       "        \"decimations\": [],\n",
+       "        \"id\": null,\n",
+       "        \"stations.local.id\": null,\n",
+       "        \"stations.local.mth5_path\": null,\n",
+       "        \"stations.local.remote\": false,\n",
[…]
