+++ _sources/user-docs/installing.rst.txt
@@ -1,0 +1,44 @@
+.. _installing:
+
+Installing
+----------
+
+We recommend installing Python (>= 3.6) using
+`Anaconda <https://www.anaconda.com/products/individual>`_ or
+`Miniconda <https://docs.conda.io/en/latest/miniconda.html>`_
+
+
+Conda forge
+***********
+
+The simplest way to install `aurora` is using
+`conda-forge <https://anaconda.org/conda-forge/aurora>`_::
+
+    conda install -c conda-forge aurora
+
+
+Pypi
+****
+
+You can also use `pypi <https://pypi.org/project/aurora/>`_ to install `aurora`::
+
+    pip install aurora
+
+
+From source
+***********
+
+If you would like to install `aurora` from `GitHub <https://github.com/simpeg/aurora>`_,
+you can clone it::
+
+    git clone https://github.com/simpeg/aurora.git
+
+and install it::
+
+    cd aurora
+    pip install -e .
+
+Issues
+******
+
+If you run into installation challenges, please `post an issue on GitHub <https://github.com/simpeg/aurora/issues/new>`_.
