+++ _sources/api/index.rst.txt
@@ -1,0 +1,17 @@
+.. _api::
+
+API Reference
+=============
+
+.. automodule:: aurora
+    :show-inheritance:
+
+
+.. toctree::
+    :maxdepth: 2
+    :caption: API documentation
+
+    time_series
+    transfer_function
+    pipelines
+    general_helper_functions
