+++ _sources/index.rst.txt
@@ -1,0 +1,24 @@
+.. aurora documentation master file, created by
+   sphinx-quickstart on Sun Aug 22 14:43:40 2021.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+.. include:: ../README.rst
+
+.. toctree::
+   :maxdepth: 2
+   :caption: User Documentation
+
+   user-docs/installing
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Examples
+
+   examples/operate_aurora.ipynb
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Reference documentation
+
+   api/index.rst
