+++ _sources/tutorials/synthetic_data_processing.nbconvert.ipynb.txt
@@ -1,0 +1,1524 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Process Synthetic Data with Aurora"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "This notebook shows how to process MTH5 data from a synthetic dataset."
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "Steps\n",
+    "1. Create the synthetic mth5\n",
+    "2. Get a Run Summary from the mth5\n",
+    "3. Select the station to process and optionally the remote reference station\n",
+    "4. Create a processing config\n",
+    "5. Generate TFs\n",
+    "6. Archive the TFs (in emtf_xml or z-file)"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "### Here are the modules we will need to import "
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {
+    "execution": {
+     "iopub.execute_input": "2024-04-12T23:10:19.933372Z",
+     "iopub.status.busy": "2024-04-12T23:10:19.933192Z",
+     "iopub.status.idle": "2024-04-12T23:10:21.162037Z",
+     "shell.execute_reply": "2024-04-12T23:10:21.161354Z"
+    }
+   },
+   "outputs": [],
+   "source": [
+    "import pathlib\n",
+    "import warnings\n",
+    "\n",
+    "from aurora.config.config_creator import ConfigCreator\n",
+    "from aurora.pipelines.process_mth5 import process_mth5\n",
+    "from aurora.pipelines.run_summary import RunSummary\n",
+    "from aurora.test_utils.synthetic.make_mth5_from_asc import create_test12rr_h5\n",
+    "from aurora.test_utils.synthetic.paths import DATA_PATH\n",
+    "from aurora.transfer_function.kernel_dataset import KernelDataset\n",
+    "\n",
+    "warnings.filterwarnings('ignore')"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Define target folder and mth5 path\n",
+    "\n",
+    "By default, the synthetic mth5 file is used for testing in `aurora/tests/synthetic/` and probably already exists on your system if you have run the tests. In the code below, we check if the file exists already, and if not we make it.\n",
+    "\n",
+    "**NOTE:** If using a read-only HPC installation, you may not be able to write to the directory where aurora is installed.  In that case, defining your target path as somewhere you have write permission.  In that case, uncommment the READ ONLY INSTALLATION block below."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "metadata": {
+    "execution": {
+     "iopub.execute_input": "2024-04-12T23:10:21.165048Z",
+     "iopub.status.busy": "2024-04-12T23:10:21.164651Z",
+     "iopub.status.idle": "2024-04-12T23:10:21.168227Z",
+     "shell.execute_reply": "2024-04-12T23:10:21.167616Z"
+    }
+   },
+   "outputs": [],
+   "source": [
+    "target_folder = DATA_PATH\n",
+    "\n",
+    "## READ ONLY INSTALLATION\n",
+    "# home = pathlib.Path.home()\n",
+    "# target_folder = home.joinpath(\"aurora_test_folder\")\n",
+    "# target_folder.mkdir(parents=True, exist_ok=True)\n",
+    "\n",
[…]

[…]
