+++ .github/workflows/tests.yml
@@ -36,7 +36,7 @@ jobs:
         conda install -c conda-forge pytest pytest-cov certifi">=2017.4.17" pandoc
         pip install -r requirements-dev.txt
         pip install git+https://github.com/kujaku11/mt_metadata.git@main
-        pip install git+https://github.com/kujaku11/mth5.git@master
+        pip install git+https://github.com/kujaku11/mth5.git@patches #master
 
     - name: Install Our Package
       run: |
