### 2024-04-27 7fb1fff
Try fix issue #323
- modify some functions decorated with path_or_mth5 so that
they operate in read-only mode, so that the channel summary is not
recomputed when the file is closed
