+++ .github/workflows/tests.yml
@@ -75,7 +75,7 @@ jobs:
         # pytest -s -v tests/synthetic/test_fourier_coefficients.py --cov=./ --cov-report=xml --cov=aurora
         pytest -s -v --cov=./ --cov-report=xml --cov=aurora
 
-  - name: Upload coverage reports to Codecov
+    - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v4
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
