### 2024-05-08 #6520 @danielmk
> Hi @danielmk do you have time to review this paper?

Unfortunately I have no experience with magnetotelluric data. I briefly looked at the paper and I don't think I am qualified to review.
