### 2024-05-08 #6520 @alejandrojuria
> Hi @alejandrojuria do you have time to review this paper?

I am sorry but I have to decline, I do not think I have the expertise nor the time to do a proper review of the paper.
