### 2024-05-11 c861e84
Minor housekeeping
- add example of restricting TF calaculation to subset of a run in operate aurora
