### 2024-05-11 #6520 @mrava87
@diehlpk thanks for the invite, however I’m not really into magnetotellurics, so I don’t think I can provide a valuable feedback here 😔
