### 2024-05-28 5be0bd5
add check for 'exact' join fail
