### 2024-05-29 afd62da
add validate channels, comment out drop_nans
