### 2024-05-29 103c554
added prototype_decimation_4 to use xr.sps_signal accessors
