+++ aurora/pipelines/transfer_function_kernel.py
@@ -3,7 +3,10 @@ import pandas as pd
 import psutil
 
 from aurora.pipelines.helpers import initialize_config
-from aurora.pipelines.time_series_helpers import prototype_decimate
+from aurora.pipelines.time_series_helpers import (
+    prototype_decimate,
+    prototype_decimate_4,
+)
 from mth5.utils.exceptions import MTH5Error
 from mth5.utils.helpers import initialize_mth5
 from mth5.utils.helpers import path_or_mth5_object
@@ -132,21 +135,23 @@ class TransferFunctionKernel(object):
                     continue
                 run_xrds = row["run_dataarray"].to_dataset("channel")
                 decimation = self.config.decimations[i_dec_level].decimation
-                decimated_xrds = prototype_decimate(decimation, run_xrds)
-                self.dataset_df["run_dataarray"].at[i] = decimated_xrds.to_array(
+                decimated_xrds = prototype_decimate_4(decimation, run_xrds)
+                self.dataset_df["run_dataarray"].at[
+                    i
+                ] = decimated_xrds.to_array(
                     "channel"
                 )  # See Note 1 above
 
-        msg = (
-            f"Dataset Dataframe Updated for decimation level {i_dec_level} Successfully"
-        )
+        msg = f"Dataset Dataframe Updated for decimation level {i_dec_level} Successfully"
         logger.info(msg)
         return
 
     def apply_clock_zero(self, dec_level_config):
         """get clock-zero from data if needed"""
         if dec_level_config.window.clock_zero_type == "data start":
-            dec_level_config.window.clock_zero = str(self.dataset_df.start.min())
+            dec_level_config.window.clock_zero = str(
+                self.dataset_df.start.min()
+            )
         return dec_level_config
 
     @property
@@ -212,7 +217,12 @@ class TransferFunctionKernel(object):
             remote = run_sub_df.remote.iloc[0]
             mth5_path = run_sub_df.mth5_path.iloc[0]
             fcs_present = mth5_has_fcs(
-                mth5_path, survey_id, station_id, run_id, remote, self.processing_config
+                mth5_path,
+                survey_id,
+                station_id,
+                run_id,
+                remote,
+                self.processing_config,
             )
             self.dataset_df.loc[dataset_df_indices, "fc"] = fcs_present
 
@@ -245,7 +255,9 @@ class TransferFunctionKernel(object):
         columns_to_show = self.processing_summary.columns
         columns_to_show = [x for x in columns_to_show if x not in omit_columns]
         logger.info("Processing Summary Dataframe:")
-        logger.info(f"\n{self.processing_summary[columns_to_show].to_string()}")
+        logger.info(
+            f"\n{self.processing_summary[columns_to_show].to_string()}"
+        )
 
     def make_processing_summary(self):
         """
@@ -265,11 +277,15 @@ class TransferFunctionKernel(object):
         decimation_info = self.config.decimation_info()
         for i_dec, dec_factor in decimation_info.items():
             tmp[i_dec] = dec_factor
-        tmp = tmp.melt(id_vars=id_vars, value_name="dec_factor", var_name="dec_level")
+        tmp = tmp.melt(
+            id_vars=id_vars, value_name="dec_factor", var_name="dec_level"
+        )
         sortby = ["survey", "station_id", "run_id", "start", "dec_level"]
         tmp.sort_values(by=sortby, inplace=True)
         tmp.reset_index(drop=True, inplace=True)
-        tmp.drop("sample_rate", axis=1, inplace=True)  # not valid for decimated data
+        tmp.drop(
+            "sample_rate", axis=1, inplace=True
+        )  # not valid for decimated data
 
         # Add window info
         group_by = [
@@ -305,7 +321,9 @@ class TransferFunctionKernel(object):
                     num_samples_window=row.num_samples_window,
[…]
