### 2024-05-30 c4d0018
updating expected rms values for using the reample_poly method
