### 2024-05-30 1a891c7
if survey_id is not in survey_metadata dictionary add one
