### 2024-05-30 9009d72
Update test_define_frequency_bands.py
