### 2024-05-30 58369e2
Update test_fourier_coefficients.py
