### 2024-06-01 #6520 @kkappler
Hi @diehlpk, this paper was reviewed internally at USGS before submission by [Michael Mitchell](https://www.usgs.gov/staff-profiles/michael-albert-mitchell) submission his github handle is @micmitch  -- perhaps JOSS can use this review as well?
