### 2024-06-04 #6520 @diehlpk
> Hi @diehlpk, this paper was reviewed internally at USGS before submission by [Michael Mitchell](https://www.usgs.gov/staff-profiles/michael-albert-mitchell) --  github handle is @micmitch  -- perhaps JOSS can use this review as well?

That would be a conflict of interest. I found a second reviewer and started the review.
