### 2024-06-06 d0ec55a
Remove legacy ascii data from EMTF
- now get this from MTH5
- relates to issue #320
