+++ tests/test_general_helper_functions.py
@@ -4,6 +4,7 @@ import unittest
 
 from aurora.general_helper_functions import count_lines
 from aurora.general_helper_functions import DotDict
+from aurora.general_helper_functions import get_mth5_ascii_data_path
 from aurora.general_helper_functions import get_test_path
 
 TEST_PATH = get_test_path()
@@ -35,6 +36,19 @@ class TestGeneralHelperFunctions(unittest.TestCase):
         dot_dict = DotDict(tmp)
         assert dot_dict.a == tmp["a"]
         assert dot_dict.b == "bb"
+
+    def test_get_mth5_ascii_data_path(self):
+        """
+        Make sure that the ascii data are where we think they are.
+        Returns
+        -------
+
+        """
+        mth5_data_path = get_mth5_ascii_data_path()
+        ascii_file_paths = list(mth5_data_path.glob("*asc"))
+        file_names = [x.name for x in ascii_file_paths]
+        assert "test1.asc" in file_names
+        assert "test2.asc" in file_names
 
 
 def main():
