### 2024-06-06 be3e7dc
debug access to mth5 ascii data
