### 2024-06-06 88c8d3f
add logging messages
