### 2024-06-06 908b0dd
debug ascii file locations
