+++ .github/workflows/tests.yml
@@ -36,12 +36,16 @@ jobs:
         conda install -c conda-forge pytest pytest-cov certifi">=2017.4.17" pandoc
         pip install -r requirements-dev.txt
         pip install git+https://github.com/kujaku11/mt_metadata.git@main
-        pip install git+https://github.com/kujaku11/mth5.git@patches #master
+        pip install git+https://github.com/kujaku11/mth5.git@fix_issue_191 #master
+        ls /usr/share/miniconda/envs/aurora-test/lib/python3.10/site-packages/mth5/*
+        ls /usr/share/miniconda/envs/aurora-test/lib/python3.10/site-packages/mth5/data/*
 
     - name: Install Our Package
       run: |
         pip install -e .
         conda list
+        ls /usr/share/miniconda/envs/aurora-test/lib/python3.10/site-packages/mth5/*
+        ls /usr/share/miniconda/envs/aurora-test/lib/python3.10/site-packages/mth5/data/*
         
 #    - name: Install Jupyter and dependencies
 #      run: |
