### 2024-06-06 76a2be2
rm ls commands
