### 2024-06-06 84156d9
add jupyter tests back in
