### 2024-06-07 c441b4e
minor changes
- add some doc
- switch branch of mt_metadata to have channel_nomenclature.hx_hy_hz
  available
- use channel_nomenclature.hx_hy_hz in station_config
