### 2024-06-07 6415571
add type hint, doc
