### 2024-06-08 35cbdf3
Added a degnerate fc for decimated data
- if you store decimations in mth5, add_fcs_to_mth5 should not decimate
  by default
- added a tool for this
- also some docstrings
