### 2024-06-08 030dddb
Update synthetic data making tools
- working on updating synthetic run concept so it uses the actual Run
  class from mt_metadata, ditto for synthetic station
