+++ aurora/test_utils/synthetic/make_mth5_from_asc.py
@@ -21,7 +21,7 @@ data/test1_LEMI12.h5
 so the band between the old and new Nyquist frequencies is bogus.
 
 """
-
+# import inspect
 import numpy as np
 import pandas as pd
 import pathlib
@@ -36,7 +36,11 @@ from aurora.test_utils.synthetic.station_config import make_station_04
 from loguru import logger
 from mth5.mth5 import MTH5
 from mth5.timeseries import ChannelTS, RunTS
+import mt_metadata.timeseries as metadata_ts
 from mt_metadata.transfer_functions.processing.aurora import ChannelNomenclature
+from mt_metadata.timeseries import Channel
+from mt_metadata.timeseries import Electric
+from mt_metadata.timeseries import Magnetic
 from mth5.utils.helpers import add_filters
 
 np.random.seed(0)
@@ -67,29 +71,33 @@ def create_run_ts_from_synthetic_run(run, df, channel_nomenclature="default"):
     -------
 
     """
+    #    meta_classes = dict(inspect.getmembers(metadata_ts, inspect.isclass))
+
     channel_nomenclature_obj = ChannelNomenclature()
     channel_nomenclature_obj.keyword = channel_nomenclature
     EX, EY, HX, HY, HZ = channel_nomenclature_obj.unpack()
     ch_list = []
     for col in df.columns:
+
         data = df[col].values
-        meta_dict = {
-            "component": col,
-            "sample_rate": run.sample_rate,
-            "filter.name": run.filters[col],
-            "filter.applied": len(run.filters[col])
-            * [
-                True,
-            ],
-            "time_period.start": run.start,
-        }
+        # meta_dict = {
+        #     "component": col,
+        #     "sample_rate": run.run_metadata.sample_rate,
+        #     "filter.name": run.filters[col],
+        #     "filter.applied": len(run.filters[col])
+        #     * [
+        #         True,
+        #     ],
+        #     "time_period.start": run.start,
+        # }
         if col in [EX, EY]:
-            meta_dict["units"] = "millivolts per kilometer"
-            channel_metadata = {"electric": meta_dict}
+            channel_metadata = Electric()
+            channel_metadata.component = col
+            channel_metadata.units = "millivolts per kilometer"
             chts = ChannelTS(
                 channel_type="electric",
                 data=data,
-                channel_metadata=channel_metadata,
+                channel_metadata=channel_metadata.to_dict(),  # ["electric"],
             )
             # add metadata to the channel here
             chts.channel_metadata.dipole_length = 50
@@ -97,15 +105,31 @@ def create_run_ts_from_synthetic_run(run, df, channel_nomenclature="default"):
                 chts.channel_metadata.measurement_azimuth = 90.0
 
         elif col in [HX, HY, HZ]:
-            meta_dict["units"] = "nanotesla"
-            channel_metadata = {"magnetic": meta_dict}
+            channel_metadata = Magnetic()
+            channel_metadata.units = "nanotesla"
+            channel_metadata.component = col
+            channel_metadata.channel_number = 0
+            channel_metadata.sample_rate = 1.0
+            # channel_obj.time_period.start = str(ats_metadata["start_date_datetime"])
+            # channel_obj.time_period.end = str(ats_metadata["end_date_datetime"])
+            # channel_metadata = {"magnetic": channel_metadata.to_dict()}
             chts = ChannelTS(
-                channel_type="magnetic",
+                channel_type=channel_metadata.type,
                 data=data,
-                channel_metadata=channel_metadata,
[…]
