### 2024-06-08 0f8bdb1
cleanup and bug fix
- Now that the actual Run class is being used, needed to update  test
  metadat set correctly to use correct syntax.
