+++ aurora/test_utils/synthetic/make_mth5_from_asc.py
@@ -97,9 +97,6 @@ def create_run_ts_from_synthetic_run(run, df, channel_nomenclature="default"):
             channel_metadata.component = col
             channel_metadata.channel_number = 0
             channel_metadata.sample_rate = 1.0
-            # channel_obj.time_period.start = str(ats_metadata["start_date_datetime"])
-            # channel_obj.time_period.end = str(ats_metadata["end_date_datetime"])
-            # channel_metadata = {"magnetic": channel_metadata.to_dict()}
             chts = ChannelTS(
                 channel_type=channel_metadata.type,
                 data=data,
