+++ .github/workflows/tests.yml
@@ -35,8 +35,8 @@ jobs:
         python --version
         conda install -c conda-forge pytest pytest-cov certifi">=2017.4.17" pandoc
         pip install -r requirements-dev.txt
-        pip install git+https://github.com/kujaku11/mt_metadata.git@fix_mth5_issue_207 # main
-        pip install git+https://github.com/kujaku11/mth5.git@fix_issue_191 #master
+        pip install git+https://github.com/kujaku11/mt_metadata.git #@fix_mth5_issue_207 # main
+        pip install git+https://github.com/kujaku11/mth5.git@patches 
 
     - name: Install Our Package
       run: |
