### 2024-06-12 d4551dd
switch to mth5 synthetic config
