### 2024-06-21 simpeg/aurora#333 @blsqr
[JOSS Review] Suggestions and improvements for software paper
(This issue is part of [the review process](https://github.com/openjournals/joss-reviews/issues/6832) for a submission to JOSS.)

Overall, the software paper (in the version linked [here](https://github.com/openjournals/joss-reviews/issues/6832#issuecomment-2146362715)) is well-written and structured and provides a good starting point for assessing the scope of aurora and how to use it. However, there are a number of typos, ambiguities and inconsistencies that I feel hinder readability and understanding.

First, regarding the content:
- L 87: "results are similar": What does this mean? If they are "only" similar, where does the difference come from? Can you link to the results of this test?

Typos and errors that should be corrected:
- L 21-22: Any reason why "Tabular Data" and "Processing" is capitalised?
- L 29: missing comma after "orthogonal"
- L 33: superfluous "involves" or "requires"
- L 34: superfluous comma after "timestamps"
- L 36: missing end of sentence
- L 60: superfluous period after "parameters"
- L 71: "the flow of Fig. 2" -> "the flow described in Fig. 2" (or equivalent wording)
- Fig. 2 caption: "transfer functions can similarly _be exported to the_ most common…", right?
- L 80: "ongoing … is ongoing"
- L 95: "Aurora'a" -> "Aurora's"
- L 103: "to co-develop" -> "to be co-developed"
- L 120: reference should be in parentheses, period missing

Other (editorial) remarks, questions, suggestions:
- L 46: "a _sort of_ continuity"; what is a "sort of" continuity? (Specify or rephrase)
- Fig. 1: Is the screenshot from the config file distorted?
- Fig. 2: It's nice to include example plots here, but I think these are far too small. If you want to keep the small size, I would suggest to strongly simplify the figures or even replace them by sketches. If you want the actual output, I would suggest letting the flow go from top to bottom and use larger figures (and ideally avoid overlapping labels etc). Also, the regression figure is lacking labels altogether.
- Fig. 2 & 4: These have strongly varying font sizes, which should be adjusted to be more consistent across the whole paper.
- L 61: the "(supporting the R…)" remark is, I imagine, mysterious for people who don't know about FAIR. Perhaps add a reference? Currently, this seems more like a footnote and might as well be removed; if you want to highlight reproducibility aspects of Aurora, perhaps make this a full sentence?
- L 73/74: Perhaps rephrase the first two sentences to improve readability? (avoid "This… This…")
- L 74/75: "from the EarthScope Schultz (2010)" – I think this would be clearer if you would directly say "EMScope dataset".
- Fig. 3: Any chance to embed this as Markdown Code with Python Syntax Highlighting? I would find this preferable over a screenshot.
- Fig. 3 caption: Use `monotype` for code snippets like `RunSummary()`. Also, was this intended to be a list? The — seem to suggest so … but I'm not sure. Would propose to delete those.
- Fig. 4 caption: Perhaps it would be good to mention that this is the example output resulting from the code shown above?
- L 89: "also run via GitHub actions." … perhaps append "to assert functionality" to that sentence?
- L 102: "(most test data is sampled at 1Hz)", to me (not a domain expert), this remark seems overly specific here.
- L 115ff: The appendix feels a bit unnecessary. I think the installation instructions are best taken from the repository page. Information about documentation and license should ideally be included into the main text at a suitable point (don't need to be separate sections).
- L 128: A `## References` header should be added before the list of references starts.

And some more subjective suggestions (not affecting approval of the manuscript):
- L 35: Consider adding a [citation for xarray](https://docs.xarray.dev/en/stable/getting-started-guide/faq.html#how-should-i-cite-xarray)
- L 44: "_internal_ development" … the "internal" is perhaps not needed?
- L 55ff: I would suggest to denote code, e.g. class names, with monotype font (`KernelDataset`, `Processing`, `TransferFunctionKernel`) to make clear that you are referring to parts of the aurora code. This should be done consistently throughout the text.
