### 2024-06-21 #6832 @blsqr
@diehlpk Thank you for checking in. I continued my review today and opened two issues in the Aurora repository:

* [ ] https://github.com/simpeg/aurora/issues/332
* [ ] https://github.com/simpeg/aurora/issues/333
