### 2024-06-26 e9694fd
Add covariance matrix method
- removing tests while in dev
