+++ .github/workflows/tests.yml
@@ -1,103 +1,0 @@
-name: Testing
-
-on:
-  push:
-    branches:
-      - '*'
-  pull_request:
-    branches:
-      - '*'
-jobs:
-  setup-build:
-    name: Ex1 (${{ matrix.python-version }}, ${{ matrix.os }})
-    runs-on: ${{ matrix.os }}
-    defaults:
-      run:
-        shell: bash -l {0}
-    strategy:
-      fail-fast: false
-      matrix:
-        os: ["ubuntu-latest"]
-        python-version: [3.8, 3.9, "3.10", "3.11"]
-#        python-version: ["3.10", ] 
-
-    steps:
-    - uses: actions/checkout@v2
-
-    - name: Setup Miniconda
-      uses: conda-incubator/setup-miniconda@v2.1.1
-      with:
-        activate-environment: aurora-test
-        python-version: ${{ matrix.python-version }}
-
-    - name: Install Env
-      run: |
-        python --version
-        conda install -c conda-forge pytest pytest-cov certifi">=2017.4.17" pandoc
-        pip install -r requirements-dev.txt
-        pip install git+https://github.com/kujaku11/mt_metadata.git #@fix_mth5_issue_207 # main
-        pip install git+https://github.com/kujaku11/mth5.git@patches 
-
-    - name: Install Our Package
-      run: |
-        pip install -e .
-        conda list
-        
-    - name: Install Jupyter and dependencies
-      run: |
-        pip install jupyter
-        pip install ipykernel
-        python -m ipykernel install --user --name aurora-test
-        # Install any other dependencies you need
-
-    - name: Execute Jupyter Notebooks
-      run: |
-        jupyter nbconvert --to notebook --execute docs/examples/dataset_definition.ipynb
-        jupyter nbconvert --to notebook --execute docs/examples/make_cas04_single_station_h5.ipynb
-        jupyter nbconvert --to notebook --execute docs/examples/operate_aurora.ipynb
-        jupyter nbconvert --to notebook --execute tests/test_run_on_commit.ipynb
-        jupyter nbconvert --to notebook --execute docs/tutorials/pole_zero_fitting/lemi_pole_zero_fitting_example.ipynb
-        jupyter nbconvert --to notebook --execute docs/tutorials/processing_configuration.ipynb
-        jupyter nbconvert --to notebook --execute docs/tutorials/synthetic_data_processing.ipynb
-        # Replace "notebook.ipynb" with your notebook's filename
-
-#    - name: Commit changes (if any)
-#      run: |
-#        git config --local user.email "action@github.com"
-#        git config --local user.name "GitHub Action"
-#        git commit -a -m "Execute Jupyter notebook"
-#        git push
-#      if: ${{ success() }}
-
-
-    - name: Run Tests
-      run: |
-        # pytest -s -v tests/synthetic/test_fourier_coefficients.py
-        # pytest -s -v tests/test_general_helper_functions.py 
-        pytest -s -v --cov=./ --cov-report=xml --cov=aurora
-
-    - name: Upload coverage reports to Codecov
-      uses: codecov/codecov-action@v4
-      with:
-        token: ${{ secrets.CODECOV_TOKEN }}
-        fail_ci_if_error: false
-
-
-    - name: Build Doc
-      if: ${{ (github.ref == 'refs/heads/main') && (matrix.python-version == '3.8')}}
-      run: |
-        cd docs
-        make html
-        cd ..
-
-    - name: GitHub Pages
-      if: ${{ (github.ref == 'refs/heads/main') && (matrix.python-version == '3.8')}}
-      uses: crazy-max/ghaction-github-pages@v2.5.0
-      with:
-        build_dir: docs/_build/html
[…]
