### 2024-06-28 simpeg/aurora#332 @diehlpk
Hi @kkappler could you please have a look?
