### 2024-06-29 be24971
Add CONTRIBUTING.rst
- put a minimal description and links to the more detailed simpeg pages
