+++ CONTRIBUTING.rst
@@ -1,0 +1,14 @@
+Contributing to Aurora
+----------------------
+
+
+If you'd like to contribute, start by searching through the `issues <https://github.com/simpeg/aurora/issues>`_ to see whether someone else has raised a similar idea or question.  If you don't see your idea listed, open an issue, and if there are code changes, also a pull request.
+
+
+There are many ways to contribute to the aurora, such as:
+
+* Report/fix bugs or issues encountered when using the software
+* Suggest additional features or functionalities
+* Fix editorial inconsistencies or inaccuracies
+
+Aurora is hosted by simpeg, please refer to their  `contributing guidelines page <https://docs.simpeg.xyz/latest/content/getting_started/contributing/index.html>`_. for ore details.
