### 2024-06-29 simpeg/aurora#332 @kkappler
Hi @blsqr, thanks for opening this issue.

Aurora is hosted by the [simpeg](https://simpeg.xyz/) project, and they have a fairly extensive [contributing guidelines](https://docs.simpeg.xyz/latest/content/getting_started/contributing/index.html).

To address this issue, I added a minimal CONTRIBUTING.rst borrowing some text from the opensourceguide link you shared, and then added a link to the simpeg guidelines in the .rst.

There is a [pull request](https://github.com/simpeg/aurora/pull/334) staged to merge these changes into JOSS branch.  I tried to add @blsqr as a reviewer on the PR, but I think I would need to add you formally to the project -- Happy to do that if you request.
