+++ paper/paper.md
@@ -54,19 +54,19 @@ bibliography: paper.bib
 
 # Summary
 
-The Aurora software package robustly estimates single station and remote reference electromagnetic transfer functions (TFs) from magnetotelluric (MT) time series.  Aurora is part of an open-source processing workflow that leverages the self-describing data container MTH5, which in turn leverages the general mt\_metadata framework to manage metadata.  These pre-existing packages simplify the processing by providing managed data structures, allowing for transfer functions to be generated with only a few lines of code.  The processing depends on two inputs -- a table defining the data to use for TF estimation, and a JSON file specifying the processing parameters, both of which are generated automatically, and can be modified if desired.  Output TFs are returned as mt\_metadata objects, and can be exported to a variety of common formats for plotting, modeling and inversion.  
+The Aurora software package robustly estimates single station and remote reference electromagnetic transfer functions (TFs) from magnetotelluric (MT) time series.  Aurora is part of an open-source processing workflow that leverages the self-describing data container MTH5, which in turn leverages the general mt\_metadata framework to manage metadata.  These pre-existing packages simplify the processing by providing managed data structures, transfer functions to be generated with only a few lines of code.  The processing depends on two inputs -- a table defining the data to use for TF estimation, and a JSON file specifying the processing parameters, both of which are generated automatically, and can be modified if desired.  Output TFs are returned as mt\_metadata objects, and can be exported to a variety of common formats for plotting, modeling and inversion.  
 
 ## Key Features
 
-- Tabular Data indexing and management (pandas data frames), 
-- Dictionary-like Processing parameters configuration
+- Tabular ata indexing and management (pandas data frames), 
+- Dictionary-like processing parameters configuration
 - Programmatic or manual editing of inputs
 - Largely automated workflow
 
 
 # Introduction
 
-MT is a geophysical technique for probing subsurface electrical conductivity using co-located electric and magnetic field measurements.  After data collection, standard practice is to estimate the time invariant  (frequency domain) transfer function (TF) between electric and magnetic channels before proceeding to interpretation and modeling. If measurements are orthogonal the TF is equivalent to the electrical impedance tensor (Z) [@Vozoff:1991].  
+MT is a geophysical technique for probing subsurface electrical conductivity using co-located electric and magnetic field measurements.  After data collection, standard practice is to estimate the time invariant  (frequency domain) transfer function (TF) between electric and magnetic channels before proceeding to interpretation and modeling. If measurements are orthogonal, the TF is equivalent to the electrical impedance tensor (Z) [@Vozoff:1991].  
 
 
 $\begin{bmatrix} E_x \\ E_y \end{bmatrix}
@@ -77,46 +77,46 @@ Z_{yx} & Z_{yy}
 \end{bmatrix}
 \begin{bmatrix} H_x \\ H_y \end{bmatrix}$
  
-where ($E_x$, $E_y$), ($H_x$, $H_y$) denote orthogonal electric and magnetic fields respectively.  TF estimation involves requires the E and H time series _and_ metadata (locations, orientations, timestamps,) and uses a collection of signal processing and statistical techniques (@egbert1997robust and references therein).  MTH5 supplies time series as xarray objects for efficient, lazy access to data and easy application of scientific computing libraries available in the python.
[…]
