### 2024-06-30 2169179
Update Figure 1
- remove vertical exaggeration distortion by making figure into two panels
- add some text to caption to clarify the new format
