+++ paper/paper.md
@@ -92,7 +92,7 @@ A TF instance depends on two key prior decisions: a) The data input to the TF co
 
 Generation of robust TFs can be done in only a few lines starting from an MTH5 archive (Figure \ref{minimal_example}).  Simplicity of workflow is due to the MTH5 container already storing comprehensive metadata, including a channel summary table describing all time series stored in the archive including start/end times and sample rates.  Users can easily view a tabular summary of available data and select station pairs to process.  Once a station -- and optionally a remote reference station -- are defined, the simultaneous time intervals of data coverage at both stations are identified automatically, providing the Kernel Dataset.  Reasonable starting processing parameters are automatically generated for a given Kernel Dataset, and can be modified with code or via manual changes to a JSON file. Once the TFK is defined, the processing automatically follows the flow described by Figure \ref{FLOW}.
 
-![TF Kernel concept diagram: A box representing the TF Kernel with two inlay boxes representing the dataset (pandas DataFrame) and processing config (JSON).  Processing config image is clipped to show only a few lines.  \label{TFK}](TFKernelApproach.png)
+![TF Kernel concept diagram: Upper panel represents the TF Kernel with two inlay boxes representing the dataset (pandas DataFrame) and processing config (JSON). Lower panel illustrates example instances of these structures. Processing config image is clipped to show only a few lines.  \label{TFK}](TFKernelApproach.png)
 
 ![The main interfaces of Aurora TF processing. \label{FLOW}  Example time series from MTH5 archive in linked notebook (using MTH5 built-in time series plotting), spectrogram from Fourier coeffcient (FC) data structure, regression cartoon from @hand2018statistical and TF from [SPUD](https://ds.iris.edu/spud/emtf/18633652).  Input time series are from MTH5, these can initially be drawn from Phoenix, LEMI, FDSN, Metronix, Zonge, systems etc. and the resultant transfer functions can be exported to the most common TF formats such as .edi, .zmm, ,j, .avg, .xml etc.](aurora_workflow.png)
 
