### 2024-06-30 0c00e5d
Update caption to simplified figure 2
