### 2024-06-30 a9f005c
Add control for axis labels and tick labels font sizes
- relates to #333
