+++ aurora/transfer_function/plot/rho_phi_helpers.py
@@ -5,8 +5,34 @@ axis.
 
 
 def plot_rho(
-    ax, periods, rho, marker="o", color="k", linestyle="None", label="", markersize=10
+    ax,
+    periods,
+    rho,
+    marker="o",
+    color="k",
+    linestyle="None",
+    label="",
+    markersize=10,
+    ax_label_size=16,
 ):
+    """
+
+    Parameters
+    ----------
+    ax
+    periods
+    rho
+    marker
+    color
+    linestyle
+    label
+    markersize
+    ax_label_size
+
+    Returns
+    -------
+
+    """
     ax.loglog(
         periods,
         rho,
@@ -16,13 +42,21 @@ def plot_rho(
         label=label,
         markersize=markersize,
     )
-    ax.tick_params(axis="both", which="major", labelsize=16)
+    ax.tick_params(axis="both", which="major", labelsize=ax_label_size)
     ax.tick_params(axis="x", which="minor", bottom=True)
     return
 
 
 def plot_phi(
-    ax, periods, phi, marker="o", color="k", linestyle="None", label="", markersize=10
+    ax,
+    periods,
+    phi,
+    marker="o",
+    color="k",
+    linestyle="None",
+    label="",
+    markersize=10,
+    ax_label_size=16,
 ):
     ax.semilogx(
         periods,
@@ -33,6 +67,6 @@ def plot_phi(
         label=label,
         markersize=markersize,
     )
-    ax.tick_params(axis="both", which="major", labelsize=16)
+    ax.tick_params(axis="both", which="major", labelsize=ax_label_size)
     ax.minorticks_on()  # (axis="x", which="minor", bottom=True)
     return
