### 2024-06-30 ddcaba3
Replace figure 3 screenshot with embedded code
