### 2024-06-30 08ff310
add xarray reference
