+++ paper/paper.bib
@@ -80,6 +80,18 @@
   publisher={Oxford University Press}
 }
 
+@article{hoyer2017xarray,
+  title     = {xarray: {N-D} labeled arrays and datasets in {Python}},
+  author    = {Hoyer, S. and J. Hamman},
+  journal   = {Journal of Open Research Software},
+  volume    = {5},
+  number    = {1},
+  year      = {2017},
+  publisher = {Ubiquity Press},
+  doi       = {10.5334/jors.148},
+  url       = {https://doi.org/10.5334/jors.148}
+}
+
 @article{kelbert2020emtf,
   title={EMTF XML: New data interchange format and conversion tools for electromagnetic transfer functions},
   author={Kelbert, Anna},
