### 2024-06-30 e6bd620
Add monotypes to all class references
also update figure 1 removing abbreviations
