### 2024-06-30 9ca055a
remove remaining references to old figure 3
