+++ paper/paper.md
@@ -90,7 +90,7 @@ This manuscript describes the high-level concepts of the software – for inform
 
 A TF instance depends on two key prior decisions: a) The data input to the TF computation algorithm, b) The algorithm itself including the specific values of the processing parameters.  Aurora formalizes these concepts as classes (`KernelDataset` and `Processing`, respectively), and a third class `TransferFunctionKernel` (Figure \ref{TFK}), a composition of the `Processing`, and `KernelDataset`.  `TransferFunctionKernel` provides a place for validating consistency between selected data and processing parameters and specifies all information needed to make the calculation of a TF reproducible.
 
-Generation of robust TFs can be done in only a few lines starting from an MTH5 archive (Figure \ref{minimal_example}).  Simplicity of workflow is due to the MTH5 container already storing comprehensive metadata, including a channel summary table describing all time series stored in the archive including start/end times and sample rates.  Users can easily view a tabular summary of available data and select station pairs to process.  Once a station -- and optionally a remote reference station -- are defined, the simultaneous time intervals of data coverage at both stations are identified automatically, providing the `KernelDataset`.  Reasonable starting processing parameters are automatically generated for a given `KernelDataset`, and can be modified with code or via manual changes to a JSON file. Once the `TransferFunctionKernel` is defined, the processing automatically follows the flow described by Figure \ref{FLOW}.
+Generation of robust TFs can be done in only a few lines starting from an MTH5 archive.  Simplicity of workflow is due to the MTH5 container already storing comprehensive metadata, including a channel summary table describing all time series stored in the archive including start/end times and sample rates.  Users can easily view a tabular summary of available data and select station pairs to process.  Once a station -- and optionally a remote reference station -- are defined, the simultaneous time intervals of data coverage at both stations are identified automatically, providing the `KernelDataset`.  Reasonable starting processing parameters are automatically generated for a given `KernelDataset`, and can be modified with code or via manual changes to a JSON file. Once the `TransferFunctionKernel` is defined, the processing automatically follows the flow described by Figure \ref{FLOW}.
 
 ![TF Kernel concept diagram: Upper panel represents the TF Kernel with two inlay boxes representing the dataset (pandas DataFrame) and processing config (JSON). Lower panel illustrates example instances of these structures. Processing config image is clipped to show only a few lines.  \label{TFK}](TFKernelApproach.png)
 
@@ -98,7 +98,7 @@ Generation of robust TFs can be done in only a few lines starting from an MTH5 a
 
 # Example
 
-This section refers to a Jupyter notebook companion to this paper (archived on GitHub: [process_cas04_mulitple_station](https://github.com/simpeg/aurora/blob/joss/docs/examples/process_cas04_mulitple_station.ipynb)).  The companion notebook builds an MTH5 dataset from the EMscope dataset (@schultz2010emscope) and executes data processing -- a condensed version of which is shown in \ref{minimal_example} below.  Apparent resistivities are plotted in Figure \ref{compareTFs} along with the EMTF-generated results hosted at EarthScope.  
+This section refers to a Jupyter notebook companion to this paper (archived on GitHub: [process_cas04_mulitple_station](https://github.com/simpeg/aurora/blob/joss/docs/examples/process_cas04_mulitple_station.ipynb)).  The companion notebook builds an MTH5 dataset from the EMscope dataset (@schultz2010emscope) and executes data processing -- a minimal_example is shown below.  Apparent resistivities are plotted in Figure \ref{compareTFs} along with the EMTF-generated results hosted at EarthScope.  
 
 
 ```python
