+++ paper/paper.bib
@@ -104,6 +104,18 @@
   publisher={Society of Exploration Geophysicists}
 }
 
+@article{peacock2022mth5,
+  title={MTH5: An archive and exchangeable data format for magnetotelluric time series data},
+  author={Peacock, Jared and Kappler, Karl and Heagy, Lindsey and Ronan, Timothy and Kelbert, Anna and Frassetto, Andrew},
+  journal={Computers & Geosciences},
+  volume={162},
+  pages = {105102},
+  year = {2022},
+  issn = {0098-3004},
+  doi = {https://doi.org/10.1016/j.cageo.2022.105102},
+  publisher={Elsevier}
+}
+
 @article{platz2019automated,
   title={An automated new pre-selection tool for noisy Magnetotelluric data using the Mahalanobis distance and magnetic field constraints},
   author={Platz, Anna and Weckmann, Ute},
