+++ paper/paper.bib
@@ -109,10 +109,10 @@
   author={Peacock, Jared and Kappler, Karl and Heagy, Lindsey and Ronan, Timothy and Kelbert, Anna and Frassetto, Andrew},
   journal={Computers & Geosciences},
   volume={162},
-  pages = {105102},
-  year = {2022},
-  issn = {0098-3004},
-  doi = {https://doi.org/10.1016/j.cageo.2022.105102},
+  pages={105102},
+  year={2022},
+  issn={0098-3004},
+  doi={https://doi.org/10.1016/j.cageo.2022.105102},
   publisher={Elsevier}
 }
 
