### 2024-07-01 cd0c888
Added minor edits
- Added reference to MTH5 paper
- Capitalized `Python`
- Added some to acknowledgements a statement needed by USGS
