+++ paper/paper.md
@@ -58,7 +58,7 @@ The Aurora software package robustly estimates single station and remote referen
 
 ## Key Features
 
-- Tabular ata indexing and management (pandas data frames), 
+- Tabular data indexing and management (Pandas dataframes), 
 - Dictionary-like processing parameters configuration
 - Programmatic or manual editing of inputs
 - Largely automated workflow
@@ -66,7 +66,7 @@ The Aurora software package robustly estimates single station and remote referen
 
 # Introduction
 
-MT is a geophysical technique for probing subsurface electrical conductivity using co-located electric and magnetic field measurements.  After data collection, standard practice is to estimate the time invariant  (frequency domain) transfer function (TF) between electric and magnetic channels before proceeding to interpretation and modeling. If measurements are orthogonal, the TF is equivalent to the electrical impedance tensor (Z) [@Vozoff:1991].  
+MT is a geophysical technique for probing subsurface electrical conductivity using collocated electric and magnetic field measurements.  Field data is collected in the time domain, however the Earth can be approximated as a linear system in the frequency domain.  Therefore, common practice is to estimate the time invariant  (frequency domain) transfer function (TF) between electric and magnetic channels to get information of the Earth's resistivity structure. If measurements are orthogonal, the TF is equivalent to the electrical impedance tensor (Z) [@Vozoff:1991].  
 
 
 $\begin{bmatrix} E_x \\ E_y \end{bmatrix}
@@ -77,11 +77,11 @@ Z_{yx} & Z_{yy}
 \end{bmatrix}
 \begin{bmatrix} H_x \\ H_y \end{bmatrix}$
  
-where ($E_x$, $E_y$), ($H_x$, $H_y$) denote orthogonal electric and magnetic fields respectively.  TF estimation requires the E and H time series _and_ metadata (locations, orientations, timestamps) and uses a collection of signal processing and statistical techniques (@egbert1997robust and references therein).  MTH5 archives the metadata _with_ the data, and supplies time series as xarray (@hoyer2017xarray) objects for efficient, lazy access to data and easy application of scientific computing libraries available in the python.  
+where ($E_x$, $E_y$), ($H_x$, $H_y$) denote orthogonal electric and magnetic fields respectively.  TF estimation requires the E and H time series _and_ metadata (locations, orientations, timestamps) along with a collection of signal processing and statistical techniques (@egbert1997robust and references therein).  MTH5 archives the metadata _with_ the data (@peacock2022mth5), and supplies time series as xarray (@hoyer2017xarray) objects for efficient, lazy access to data and easy application of scientific computing libraries available in the Python.  
 
 # Statement of Need
 
[…]
