### 2024-07-12 557ffdd
tidy some docstrings
