+++ aurora/transfer_function/regression/helper_functions.py
@@ -1,29 +1,26 @@
 import numpy as np
 
 
-def rme_beta(r0):
+def rme_beta(r0: float) -> float:
     """
-    This is an RME specific property.
-    It represents a bias in the calculation of residual_variance
-    which we correct for in TRME and TRME_RR.
+    This is an RME specific property.  It represents a bias in the calculation of
+    residual_variance which we correct for in TRME and TRME_RR. The implemented
+    formula is an approximation.  This is approximately equal to 1/beta where beta is
+    defined by Equation A3 in Egbert & Booker 1986.
 
-    The implemented formula is an approximation.  This is approximately equal to 1/beta
-    where beta is defined by Equation A3 in Egbert & Booker 1986.
-
-    for r0=1.5  the correction factor is ~1.28 and beta ~0.78
+    For r0=1.5 the correction factor is ~1.28 and beta ~0.78
 
     Some notes from a discussion with Gary:
-    In the regression esimtate you downweight things with large errors, but
-    you need to define what's large.  You estimate the standard devation
-    (sigma) of the errors from the residuals BUT with this cleaned data
-    approach (Yc) sigma is smaller than it should be, you need to
-    compensate for this by using a correction_factor. It's basically the
-    expectation, if the data really were Gaussian, and you estimated from
+    In the regression estimate you down-weight things with large errors, but you need
+    to define what is large.  You estimate the standard deviation (sigma) of the errors
+    from the residuals BUT with this cleaned data approach (Yc) sigma is smaller than
+    it should be.  You need to compensate for this by using a correction_factor. It's
+    basically the expectation, if the data really were Gaussian, and you estimated from
     the corrected data. This is how much too small the estimate would be.
 
-    If you change the penalty functional you may need a pencil, paper and
-    some calculus.  The relationship between the corrected-data-residuals
-    and the gaussin residauls could change if you change the penalty
+    If you change the penalty functional you may need pencil, paper and some calculus.
+    The relationship between the corrected-data-residuals and the Gaussian residuals
+    could change if you change the penalty.
 
     Parameters
     ----------
@@ -80,7 +77,9 @@ def simple_solve_tf(Y, X, R=None):
 
 def direct_solve_tf(Y, X, R=None):
     """
-    Cast problem Y = Xb
+    Solve problem Y = Xb for b.  This function can be used for testing.  It is not as
+    stable as using simple_solve_tf, but it is instructive to have an example of
+    regression using the crudest approach.
 
     Parameters
     ----------
@@ -97,14 +96,6 @@ def direct_solve_tf(Y, X, R=None):
     -------
     z: numpy.ndarray
         The TF estimate -- Usually two complex numbers, (Zxx, Zxy), or (Zyx, Zyy)
-    Parameters
-    ----------
-    Y
-    X
-    R
-
-    Returns
-    -------
 
     """
     if R is None:
