### 2024-07-12 bc60ffa
tidy doc strings
