### 2024-07-12 792c859
minor cleanup to IterControl
