### 2024-07-13 0ca7b70
remove setter for number_of_redescending_iterations
- use increment() and reset() functions instead
