### 2024-07-13 5bf9df9
Add some doc
- move some previously hard coded values (r0,u0, tolerance)
to be args.  This will eventually get populated from the config
- relates to #107
