### 2024-07-14 9783db7
Minor updates to regression class
- rename check_number_of_observations_xy_consistent to be an underscore function
- make X, Y explicit arguments to RegressionEstimator, not kwargs
- add typehinting for their current xr.Dataset dtypes
- factor the assignement to numpy arrays into its own method _set_up_regression_variables
