### 2024-07-14 750a3d9
Minor changes to regression
- move _check_number_of_observations_xy_consistent under _set_up_regression_variables()
- remove unneeded extra call to that in TRME_RR
