### 2024-07-14 6d711eb
Make iter_control an explicit argument, not a kwarg
