### 2024-07-14 8edfc57
stop allowing X, Y inputs to regression to be None
- can't see any reason to allow it, and it adds a bunch of checks
later in the code
