### 2024-07-14 3f4a9fc
Factor _xarray_to_numpy_array into a helper function
- add support for dataarray in regression
- add a test to confirm dataarray OK
