### 2024-07-14 2f8aa4d
update regression estimator to allow numpy arrays
- add a test
