### 2024-07-14 f2d5300
Minor fix for 1D output arrays
- if we pass a dataarray made from .sel() of a single channel the
underlying numpy is 1D.  Force this to be 2D to make things smooth
