### 2024-07-14 3a9b9b9
Hotfix for channel names for non xr.Dataset regression
- add attrs _input_channel_names, _output_channel_names to base regression
- allow these to be passed if desired, otherwise deduce from xarray, or assign to numpy array
- update m_estimator to use these attrs instead of assuming self._X, self._Y are xr.datasets
