### 2024-07-17 simpeg/aurora#333 @diehlpk
Hi @kkappler how are things going?
