### 2024-07-17 simpeg/aurora#333 @kkappler
Hi @diehlpk

Thanks for checking in.  This specific issue 333 has been addressed and is ready to merge pending feedback blsqr.

Before requesting that feedback, we had thought to wait for the second review and incorporate those changes as well.

What do you think? Should we request blsqrs' review on this issue now, or wait for the second review?
