### 2024-07-17 cee2d2c
add typehints and doc strings to _get_channel_names
