### 2024-07-17 d59d62b
add note to docstrings
