### 2024-07-17 07b6283
add optional verbosity argument to IterControl
- also somt typehints/doc
