### 2024-07-18 #6832 @sinanozaydin
@diehlpk I have opened an issue for my review:

https://github.com/simpeg/aurora/issues/337
