### 2024-07-20 8aacebb
add some type hints
