### 2024-07-20 533fe4f
add optional verbose output
