### 2024-07-20 211e556
add optional verbose output
