### 2024-07-21 6d7ed23
Update caption of Figure 2 per issue #337
