+++ paper/paper.md
@@ -90,11 +90,11 @@ This manuscript describes the high-level concepts of the software – for inform
 
 A TF instance depends on two key prior decisions: a) The data input to the TF computation algorithm, b) The algorithm itself including the specific values of the processing parameters.  Aurora formalizes these concepts as classes (`KernelDataset` and `Processing`, respectively), and a third class `TransferFunctionKernel` (Figure \ref{TFK}), a composition of the `Processing`, and `KernelDataset`.  `TransferFunctionKernel` provides a place for validating consistency between selected data and processing parameters and specifies all information needed to make the calculation of a TF reproducible.
 
-Generation of robust TFs can be done in only a few lines starting from an MTH5 archive.  Simplicity of workflow is due to the MTH5 container already storing comprehensive metadata, including a channel summary table describing all time series stored in the archive including start/end times and sample rates.  Users can easily view a tabular summary of available data and select station pairs to process.  Once a station -- and optionally a remote reference station -- are defined, the simultaneous time intervals of data coverage at both stations are identified automatically, providing the `KernelDataset`.  Reasonable starting processing parameters are automatically generated for a given `KernelDataset`, and can be modified with code or via manual changes to a JSON file. Once the `TransferFunctionKernel` is defined, the processing automatically follows the flow described by Figure \ref{FLOW}.
+Generation of robust TFs can be done in only a few lines starting from an MTH5 archive.  Simplicity of workflow is due to the MTH5 container already storing comprehensive metadata, including a channel summary table describing all time series stored in the archive including start/end times and sample rates.  Users can easily view a tabular summary of available data and select station pairs to process.  Once a station -- and optionally a remote reference station -- are defined, the simultaneous time intervals of data coverage at both stations are identified automatically, providing the `KernelDataset`.  Reasonable starting processing parameters are automatically generated for a given `KernelDataset`, and can be modified with code or via manual changes to a JSON file. Once the `TransferFunctionKernel` is defined, the processing automatically follows the flow described by Figure \ref{FLOW}.  Input time series are from MTH5, these can initially be drawn from Phoenix, LEMI, FDSN, Metronix, Zonge, systems etc. and the resultant transfer functions can be exported to the most common TF formats such as .edi, .zmm, ,j, .avg, .xml etc.  The images of Figure  \ref{FLOW} are conceptual -- in reality the time series can have data from more than one station, and the spectrograms are also multivariate (not single channel as shown).  The regression is also multivariate, and applied on complex-valued data from the spectrograms, this illustration however conveys the key idea of regression in the presence of outliers and mixed clusters.
 
 ![TF Kernel concept diagram: Upper panel represents the TF Kernel with two inlay boxes representing the dataset (pandas DataFrame) and processing config (JSON). Lower panel illustrates example instances of these structures. Processing config image is clipped to show only a few lines.  \label{TFK}](TFKernelApproach.png)
 
[…]
