### 2024-07-21 ed1dba2
Add some docstrings re issue #337
This is a partial commit.
I want to merge patches into this branch before continuing to avoid conflicts in future
