+++ aurora/pipelines/time_series_helpers.py
@@ -1,3 +1,7 @@
+"""
+    Collection of modules used in processing pipeline that operate on time series
+"""
+import mt_metadata
 import numpy as np
 import pandas as pd
 import scipy.signal as ssig
@@ -11,11 +15,12 @@ from aurora.time_series.windowing_scheme import window_scheme_from_decimation
 
 def validate_sample_rate(run_ts, expected_sample_rate, tol=1e-4):
     """
+    Check that the sample rate of a run_ts is the expected value, and warn if not.
 
     Parameters
     ----------
     run_ts: mth5.timeseries.run_ts.RunTS
-        Time series object
+        Time series object with data and metadata.
     expected_sample_rate: float
         The sample rate the time series is expected to have. Normally taken from
         the processing config
@@ -69,6 +74,8 @@ def apply_prewhitening(decimation_obj, run_xrds_input):
 
 def apply_recoloring(decimation_obj, stft_obj):
     """
+    Inverts the pre-whitening operation in frequency domain.
+
     Parameters
     ----------
     decimation_obj : mt_metadata.transfer_functions.processing.fourier_coefficients.decimation.Decimation
@@ -113,6 +120,9 @@ def apply_recoloring(decimation_obj, stft_obj):
 
 def run_ts_to_stft_scipy(decimation_obj, run_xrds_orig):
     """
+    Converts a runts object into a time series of Fourier coefficients.
+    This method uses scipy.signal.spectrogram.
+
     Parameters
     ----------
     decimation_obj : mt_metadata.transfer_functions.processing.aurora.DecimationLevel
@@ -209,22 +219,24 @@ def truncate_to_clock_zero(decimation_obj, run_xrds):
     return run_xrds
 
 
-def nan_to_mean(xrds):
+def nan_to_mean(xrds: xr.Dataset) -> xr.Dataset:
     """
-    Set Nan values to mean value
+    Set Nan values in xr.Dataset to the mean value per channel.
 
-    :param xrds: DESCRIPTION
-    :type xrds: TYPE
-    :return: DESCRIPTION
-    :rtype: TYPE
+    xrds: xr.Dataset
+        Time series data
+
+    Returns
+    -------
+    run_xrds : xr.Dataset
+        The same as the input time series but NaN values are replaced by the mean of the time series (per channel).
 
     """
     for ch in xrds.keys():
         null_values_present = xrds[ch].isnull().any()
         if null_values_present:
-            logger.info(
-                "Null values detected in xrds -- this is not expected and should be examined"
-            )
+            msg = "Unexpected Null values in xrds -- this should be examined"
+            logger.warning(msg)
             value = np.nan_to_num(np.nanmean(xrds[ch].data))
             xrds[ch] = xrds[ch].fillna(value)
     return xrds
@@ -232,6 +244,10 @@ def nan_to_mean(xrds):
 
 def run_ts_to_stft(decimation_obj, run_xrds_orig):
     """
+    Converts a runts object into a time series of Fourier coefficients.
+    Similar to run_ts_to_stft_scipy, but in this implementation operations on individual
+    windows are possible (for example pre-whitening per time window via ARMA filtering).
+
 
     Parameters
     ----------
@@ -271,6 +287,7 @@ def run_ts_to_stft(decimation_obj, run_xrds_orig):
 
 def calibrate_stft_obj(stft_obj, run_obj, units="MT", channel_scale_factors=None):
     """
+    Calibrates frequency domain data into MT units.
 
     Parameters
     ----------
@@ -326,12 +343,21 @@ def calibrate_stft_obj(stft_obj, run_obj, units="MT", channel_scale_factors=None
     return stft_obj
 
 
[…]
