+++ aurora/general_helper_functions.py
@@ -1,5 +1,9 @@
+"""
+This module contains from miscellaneous functions and some global paths.
+"""
 import inspect
-import os
+
+# import os
 import pathlib
 import scipy.io as sio
 import subprocess
@@ -20,6 +24,26 @@ CONFIG_PATH = AURORA_PATH.joinpath("aurora", "config")
 BAND_SETUP_PATH = CONFIG_PATH.joinpath("emtf_band_setup")
 
 
+def get_test_path() -> pathlib.Path:
+    """
+    Gets the path to where the test are.
+
+    Returns
+    -------
+    test_path: pathlib.Path
+        Object that points to where aurora's tests are.
+    """
+    test_path = AURORA_PATH.joinpath("tests")
+    if not test_path.exists():
+        msg = (
+            f"Could not locate test directory {TEST_PATH}\n "
+            f"This is most likely because aurora was installed from pypi or conda forge\n"
+            f"TEST_PATH should be replaced with DATA_PATH"
+        )
+        logger.warning(msg)
+    return test_path
+
+
 def get_mth5_ascii_data_path():
     """
     Get the path to the
@@ -30,18 +54,6 @@ def get_mth5_ascii_data_path():
     """
     mth5_data_path = pathlib.Path(mth5.__file__).parent.joinpath("data")
     return mth5_data_path
-
-
-def get_test_path():
-    test_path = AURORA_PATH.joinpath("tests")
-    if not test_path.exists():
-        msg = (
-            f"Could not locate test directory {TEST_PATH}\n "
-            f"This is most likely because aurora was installed from pypi or conda forge\n"
-            f"TEST_PATH should be replaced with DATA_PATH"
-        )
-        logger.warning(msg)
-    return test_path
 
 
 try:
@@ -79,19 +91,14 @@ def count_lines(file_name):
 
 def execute_subprocess(cmd, **kwargs):
     """
+    A wrapper for subprocess.call
 
     Parameters
     ----------
     cmd : string
         command as it would be typed in a terminal
-    kwargs
+    kwargs: denotes keyword arguments that would be passed to subprocess
 
-    Returns
-    -------
-
-    """
-    """
-    A wrapper for subprocess.call
     """
     exit_status = subprocess.call([cmd], shell=True, **kwargs)
     if exit_status != 0:
@@ -99,38 +106,43 @@ def execute_subprocess(cmd, **kwargs):
     return
 
 
-def execute_command(cmd, **kwargs):
-    """
-    Executes command in terminal from script.
-
-    Parameters:
-        cmd (str): command to exectute from a terminal
-        kwargs: exec_dir (str): the directory from which to execute
-        kwargs: no_exception: suppress output if exception
-
-    Other Parameters:
-        exit_status: :code:`0` is good, otherwise there is some problem
-
-    .. note:: When executing :code:`rm *` this crashes if the directory we are removing
-        from is empty
-
-    .. note:: if you can you should probably use execute_subprocess() instead
-    """
-    exec_dir = kwargs.get("exec_dir", os.path.expanduser("~/"))
-    allow_exception = kwargs.get("allow_exception", True)
-    logger.info("executing from {}".format(exec_dir))
-    cwd = os.getcwd()
-    os.chdir(exec_dir)
-    exit_status = os.system(cmd)
-    if exit_status != 0:
-        logger.info(f"exit_status of {cmd} = {exit_status}")
-        if allow_exception:
-            raise Exception(f"Failed to successfully execute \n {cmd}")
-    os.chdir(cwd)
+# TODO: Add test for execute_command or delete.
+# def execute_command(cmd, **kwargs):
+#     """
+#     Executes command in terminal from script.
+#
+#     Parameters:
+#     ----------
+#     cmd : str
[…]
