### 2024-07-21 1dba995
Add doc strings and some typehints (#337)
