+++ aurora/config/config_creator.py
@@ -1,29 +1,47 @@
 """
-Helper class to make config files.
+This module contains a Helper class to make config files.
 
-Note: the config is still evolving and this class and its methods are expected to
-change.
-
+The processing config is still evolving and this class and its methods may change.
 
 """
-from loguru import logger
-
 from aurora.config.metadata.processing import Processing
 from aurora.config import BANDS_DEFAULT_FILE
-from mt_metadata.transfer_functions.processing.aurora.window import Window
 from aurora.sandbox.io_helpers.emtf_band_setup import EMTFBandSetupFile
+from loguru import logger
+from mt_metadata.transfer_functions.processing.aurora.window import Window
+from typing import Optional, Union
+import pathlib
 
 SUPPORTED_BAND_SPECIFICATION_STYLES = ["EMTF", "band_edges"]
 
 
 class ConfigCreator:
-    def __init__(self, **kwargs):
-        self._emtf_band_file = kwargs.get("emtf_band_file", None)
-        self._band_edges = kwargs.get("band_edges", None)
+    def __init__(
+        self,
+        emtf_band_file: Optional[Union[str, pathlib.Path, None]] = None,
+        band_edges: Optional[Union[dict, None]] = None,
+    ):
+        """
+        Constructor
+
+        Parameters
+        ----------
+        emtf_band_file: Optional[Union[str, pathlib.Path, None]]
+            Allows the specification of an EMTF "band setup file" for defining frequency bands.
+        band_edges: dict
+            Keys are integers corresponding to decimation level.  Values are numpy arrays.
+            Numpy arrays are one-row-per-band.  Array shape n_bands x 2. array[i_band,0] is the
+            lower edge of band, and array[i_band,0] is the upper.
+
+        """
+        self._emtf_band_file = emtf_band_file
+        self._band_edges = band_edges
         self._band_specification_style = None
 
     def processing_id(self, kernel_dataset):
         """
+        Generates a string id label for the processing config: WIP.
+
         In the past, we used f"{local}-{remote}" or  f"{local}-{run_id}"
         Neither of these is sufficiently unique.  In fact, they only describe the
         dataset, and not the processing config.  It is difficult to see how to make a
@@ -37,31 +55,47 @@ class ConfigCreator:
 
         Parameters
         ----------
-        kernel_dataset
+        kernel_dataset: aurora.transfer_function.kernel_dataset.KernelDataset
+            An object that defines the data to be processed.
 
         Returns
         -------
-
+        id: str
+            A label for the processing config.
         """
         id = f"{kernel_dataset.local_station_id}-{kernel_dataset.remote_station_id}"
         return id
 
     @property
     def band_specification_style(self):
+        """return a description of the scheme used to define the bands."""
         return self._band_specification_style
 
     @band_specification_style.setter
-    def band_specification_style(self, value):
+    def band_specification_style(self, value: str) -> None:
+        """
+        Sets the band_specification_style
+
+        Parameters
+        ----------
+        value: str
+            The label for the scheme used to define the bands.
+
+        Returns
+        -------
+
+        """
         if value not in SUPPORTED_BAND_SPECIFICATION_STYLES:
             msg = f"Won't set band specification style to unrecognized value {value}"
             logger.warning(msg)
             raise NotImplementedError(msg)
-            # return
         else:
             self._band_specification_style = value
 
-    def determine_band_specification_style(self):
+    def determine_band_specification_style(self) -> None:
         """
+        Try to identify which scheme was used to define the bands
+
         TODO: Should emtf_band_file path be stored in config to support reproducibility?
 
         """
@@ -85,43 +119,55 @@ class ConfigCreator:
         kernel_dataset,
         input_channels=["hx", "hy"],
         output_channels=["hz", "ex", "ey"],
-        estimator=None,
-        **kwargs,
[…]
