+++ aurora/pipelines/transfer_function_kernel.py
@@ -1,19 +1,26 @@
-import numpy as np
-import pandas as pd
-import psutil
+"""
+    This module contains the TrasnferFunctionKernel class which is the main object that
+    links the KernelDataset to Processing configuration.
+
+"""
 
 from aurora.pipelines.helpers import initialize_config
 from aurora.pipelines.time_series_helpers import prototype_decimate
+from loguru import logger
 from mth5.utils.exceptions import MTH5Error
 from mth5.utils.helpers import initialize_mth5
 from mth5.utils.helpers import path_or_mth5_object
 from mt_metadata.transfer_functions.core import TF
-from loguru import logger
+
+import numpy as np
+import pandas as pd
+import psutil
 
 
 class TransferFunctionKernel(object):
     def __init__(self, dataset=None, config=None):
         """
+        Constructor
 
         Parameters
         ----------
@@ -27,32 +34,39 @@ class TransferFunctionKernel(object):
 
     @property
     def dataset(self):
+        """returns the KernelDataset object"""
         return self._dataset
 
     @property
     def kernel_dataset(self):
+        """returns the KernelDataset object"""
         return self._dataset
 
     @property
-    def dataset_df(self):
+    def dataset_df(self) -> pd.DataFrame:
+        """returns the KernelDataset dataframe"""
         return self._dataset.df
 
     @property
     def processing_config(self):
+        """Returns the processing config object"""
         return self._config
 
     @property
     def config(self):
+        """Returns the processing config object"""
         return self._config
 
     @property
     def processing_summary(self):
+        """Returns the processing summary object -- creates if it doesn't yet exist."""
         if self._processing_summary is None:
             self.make_processing_summary()
         return self._processing_summary
 
     @property
-    def mth5_objs(self):
+    def mth5_objs(self) -> dict:
+        """Returns self._mth5_objs"""
         if self._mth5_objs is None:
             self.initialize_mth5s()
         return self._mth5_objs
@@ -144,13 +158,25 @@ class TransferFunctionKernel(object):
         return
 
     def apply_clock_zero(self, dec_level_config):
-        """get clock-zero from data if needed"""
+        """
+        get clock-zero from data if needed
+
+        Parameters
+        ----------
+        dec_level_config: mt_metadata.transfer_functions.processing.aurora.decimation_level.DecimationLevel
+
+        Returns
+        -------
+        dec_level_config: mt_metadata.transfer_functions.processing.aurora.decimation_level.DecimationLevel
+            The modified DecimationLevel with clock-zero information set.
+        """
         if dec_level_config.window.clock_zero_type == "data start":
             dec_level_config.window.clock_zero = str(self.dataset_df.start.min())
         return dec_level_config
 
     @property
-    def all_fcs_already_exist(self):
+    def all_fcs_already_exist(self) -> bool:
+        """Return true of all FCs needed to process data already exist in the mth5s"""
         if self.kernel_dataset.df["fc"].isna().any():
             self.check_if_fcs_already_exist()
 
@@ -248,6 +274,15 @@ class TransferFunctionKernel(object):
             "run_dataarray",
         ),
     ):
+        """
+        Prints the processing summary table via logger.
+
+        Parameters
+        ----------
+        omit_columns: tuple
+            List of columns to omit when showing channel summary (used to keep table small).
+
+        """
         columns_to_show = self.processing_summary.columns
         columns_to_show = [x for x in columns_to_show if x not in omit_columns]
         logger.info("Processing Summary Dataframe:")
@@ -255,8 +290,9 @@ class TransferFunctionKernel(object):
 
     def make_processing_summary(self):
         """
-        Melt the decimation levels over the run summary.  Add columns to estimate
-        the number of FFT windows for each row
[…]
