+++ aurora/pipelines/run_summary.py
@@ -1,6 +1,12 @@
 """
 
-Note 1: Functionality of RunSummary()
+This module contains the RunSummary class.
+
+This is a helper class that summarizes the Runs in an mth5.
+
+TODO: This class and methods could be replaced by methods in MTH5.
+
+Functionality of RunSummary()
 1. User can get a list of local_station options, which correspond to unique pairs
 of values: (survey_id,  station_id)
 
@@ -18,6 +24,13 @@ maximize coverage of the local station runs is generated
 7. Time interval endpoints can be changed
 
 
+Development Notes:
+    TODO: consider adding methods:
+     "drop_runs_shorter_than": removes short runs from summary
+     "fill_gaps_by_time_interval": allows runs to be merged if gaps between are short
+     "fill_gaps_by_run_names": allows runs to be merged if gaps between are short
+    TODO: Consider whether this should return a copy or modify in-place when querying the df.
+
 """
 
 import copy
@@ -33,6 +46,7 @@ from mt_metadata.transfer_functions.processing.aurora.channel_nomenclature impor
 import mth5
 from mth5.utils.helpers import initialize_mth5
 from loguru import logger
+from typing import Optional, Union
 
 
 RUN_SUMMARY_COLUMNS = [
@@ -51,28 +65,29 @@ RUN_SUMMARY_COLUMNS = [
 
 class RunSummary:
     """
-    The dependencies aren't clear yet.
-    Maybe still Dataset:
-        Could have methods
-            "drop_runs_shorter_than"
-            "fill_gaps_by_time_interval"
-            "fill_gaps_by_run_names"
+    Class to contain a run-summary table from one or more mth5s.
             "
+    WIP: For the full MMT case this may need modification to a channel based summary.
 
-    For the full MMT case this may need modification to a channel based summary.
-
-    Question: To return a copy or modify in-place when querying.  Need to decide on
-    standards and syntax.  Handling this in general could use a decorator that allows
-    df as kwarg, and if it is not passed the modification is done in place.
-    The user who doesn't want to modify in place can work with a clone.
-    Could also try the @staticmethod decorator so that it returns a modified df.
 
     """
 
-    def __init__(self, **kwargs):
+    def __init__(
+        self,
+        input_dict: Optional[Union[dict, None]] = None,
+        df: Optional[Union[pd.DataFrame, None]] = None,
+    ):
+
+        """
+        Constructor
+
+        Parameters
+        ----------
+        kwargs
+        """
         self.column_dtypes = [str, str, pd.Timestamp, pd.Timestamp]
-        self._input_dict = kwargs.get("input_dict", None)
-        self.df = kwargs.get("df", None)
+        self._input_dict = input_dict
+        self.df = df
         self._mini_summary_columns = ["survey", "station_id", "run_id", "start", "end"]
 
     def clone(self):
@@ -84,23 +99,28 @@ class RunSummary:
         return copy.deepcopy(self)
 
     def from_mth5s(self, mth5_list):
+        """Iterates over mth5s in list and creates one big dataframe summarizing the runs"""
         run_summary_df = extract_run_summaries_from_mth5s(mth5_list)
         self.df = run_summary_df
 
     @property
     def mini_summary(self):
+        """shows the dataframe with only a few columns for readbility"""
         return self.df[self._mini_summary_columns]
 
     @property
     def print_mini_summary(self):
+        """Calls minisummary through logger so it is formatted."""
         logger.info(self.mini_summary)
 
-    def add_duration(self, df=None):
+    def add_duration(self, df: Optional[Union[pd.DataFrame, None]] = None) -> None:
         """
+        Adds a column called "duration" to the dataframe
 
         Parameters
         ----------
-        df
[…]
