+++ aurora/sandbox/io_helpers/emtf_band_setup.py
@@ -1,12 +1,22 @@
-from loguru import logger
+"""
+    In this module is a class that emulates the old EMTF Band Setup File
 
+"""
+from loguru import logger
+from typing import Union, Optional
 import numpy as np
 import pandas as pd
+import pathlib
 
 
 class EMTFBandSetupFile:
-    def __init__(self, **kwargs):
+    def __init__(
+        self,
+        filepath: Optional[Union[str, pathlib.Path, None]] = None,
+        sample_rate: Optional[Union[float, None]] = None,
+    ):
         """
+        Constructor.
 
         Parameters
         ----------
@@ -14,31 +24,43 @@ class EMTFBandSetupFile:
             this is the sampling rate of the un-decimated data,
             aka decimation level 1 in EMTF (and probably zero in the
             aurora nomenclature)
-        kwargs
         """
-        self.filepath = kwargs.get("filepath", None)
+        self.filepath = filepath
         self.df = None
         self._num_bands = None
-        self.sample_rate = kwargs.get("sample_rate", None)
+        self.sample_rate = sample_rate
         if self.filepath:
             self.load()
 
     @property
-    def num_bands(self):
+    def num_bands(self) -> int:
+        """return the number of bands in the band setup file"""
         return self._num_bands
 
     @property
-    def num_decimation_levels(self):
+    def num_decimation_levels(self) -> int:
+        """Return the number of decimation levels"""
         num_dec = len(self.decimation_levels)
         return num_dec
 
     @property
-    def decimation_levels(self):
+    def decimation_levels(self) -> np.ndarray:
+        """Return the decimation level names (they are integers) in an array"""
         dec_levels = self.df.decimation_level.unique()
         dec_levels.sort()
         return dec_levels
 
-    def load(self, filepath=None):
+    def load(self, filepath: Optional[Union[str, pathlib.Path, None]] = None) -> None:
+        """
+        Loads an EMTF band setup file and casts it to a dataframe.
+        - populates self.df with the EMTF band setup dataframe.
+
+        Parameters
+        ----------
+        filepath: Optional[Union[str, pathlib.Path, None]]
+            The path to the emtf band setup (.cfrg) file.
+
+        """
         if filepath is None:
             filepath = self.filepath
         msg = f"loading band setup file {filepath}"
@@ -59,7 +81,23 @@ class EMTFBandSetupFile:
             raise Exception(msg)
         self.df = df
 
-    def get_decimation_level(self, decimation_level, order="ascending_frequency"):
+    def get_decimation_level(
+        self, decimation_level: int, order: str = "ascending_frequency"
+    ) -> pd.DataFrame:
+        """
+        Return a sub-dataframe with only the rows that correspond to the requested decimation level
+
+        Parameters
+        ----------
+        decimation_level: int
+            The id of the decimation level
+        order: str
+
+        Returns
+        -------
+        decimation_level_df: pd.DataFrame
+            A sub-dataframe with only the rows that correspond to the requested decimation level
+        """
         if self.df is None:
             self.load()
         decimation_level_df = self.df[self.df["decimation_level"] == decimation_level]
@@ -67,28 +105,43 @@ class EMTFBandSetupFile:
             decimation_level_df = decimation_level_df.sort_values(
                 by="lower_bound_index"
             )
-
+        else:
+            msg = f"order of decimation dataframe {order} not supported -- returning default order"
+            logger.warning(msg)
         return decimation_level_df
 
-    def compute_band_edges(self, decimation_factors, num_samples_window):
+    def compute_band_edges(
+        self, decimation_factors: list, num_samples_window: int
+    ) -> dict:
         """
-        Adds columns to df defining some useful numbers
-        sample_rate
+        Adds columns to df defining the upper and lower bounds of the frequency bands in Hz.
+
[…]
