+++ aurora/test_utils/synthetic/make_mth5_from_asc.py
@@ -17,8 +17,10 @@ Also created are some files with the same data but other channel_nomenclature sc
 data/test12rr_LEMI34.h5
 data/test1_LEMI12.h5
 
-- 20231103: Added an 8Hz upsampled version of test1.  No spectral content was added
+- 20231103: Added an 8Hz up-sampled version of test1.  No spectral content was added
 so the band between the old and new Nyquist frequencies is bogus.
+
+Notes: Work in progress -- this module is being migrated to MTH5.
 
 """
 # import inspect
@@ -47,11 +49,9 @@ synthetic_test_paths = SyntheticTestPaths()
 MTH5_PATH = synthetic_test_paths.mth5_path
 
 
-def create_run_ts_from_synthetic_run(run, df, channel_nomenclature="default"):
+def create_run_ts_from_synthetic_run(run, df, channel_nomenclature="default") -> RunTS:
     """
-    Loop over stations and make ChannelTS objects.
-    Need to add a tag in the channels
-    so that when you call a run it will get all the filters with it.
+    Loop over channels of synthetic data in df and make ChannelTS objects.
 
     Parameters
     ----------
@@ -67,7 +67,8 @@ def create_run_ts_from_synthetic_run(run, df, channel_nomenclature="default"):
 
     Returns
     -------
-
+    runts: RunTS
+        MTH5 run time series object, data and metadata bound into one.
     """
 
     channel_nomenclature_obj = ChannelNomenclature()
@@ -127,12 +128,16 @@ def create_run_ts_from_synthetic_run(run, df, channel_nomenclature="default"):
 
 def get_time_series_dataframe(run, source_folder, add_nan_values):
     """
+    Returns time series data in a dataframe with columns named for EM field component.
+
     Parameters
     ----------
     run: aurora.test_utils.synthetic.station_config.SyntheticRun
         Information needed to define/create the run
-
     source_folder: pathlib.Path, or null
+        Where to load the ascii time series from
+    add_nan_values: bool
+        If True, add some NaN, if False, do not add Nan.
 
     Up-samples data to run.sample_rate, which is treated as in integer.
     Only tested for 8, to make 8Hz data for testing.  If run.sample_rate is default (1.0)
@@ -141,7 +146,7 @@ def get_time_series_dataframe(run, source_folder, add_nan_values):
     Returns
     -------
     df: pandas.DataFrame
-    The time series data for the synthetic run
+        The time series data for the synthetic run
     """
     # point to the ascii time series
     if source_folder:
@@ -186,6 +191,7 @@ def create_mth5_synthetic_file(
     force_make_mth5=True,
 ):
     """
+    Creates an MTH5 from synthetic data
 
     Parameters
     ----------
@@ -284,6 +290,21 @@ def create_test1_h5(
     source_folder="",
     force_make_mth5=True,
 ):
+    """
+    Creates an MTH5 file for a single station named "test1".
+
+    Parameters
+    ----------
+    file_version
+    channel_nomenclature
+    target_folder
+    source_folder
+    force_make_mth5
+
+    Returns
+    -------
+
+    """
     station_01_params = make_station_01(channel_nomenclature=channel_nomenclature)
     mth5_name = station_01_params.mth5_name
     station_params = [
@@ -309,6 +330,9 @@ def create_test2_h5(
     target_folder=MTH5_PATH,
     source_folder="",
 ):
+    """
+    Creates an MTH5 file for a single station named "test2".
+    """
     station_02_params = make_station_02(channel_nomenclature=channel_nomenclature)
     mth5_name = station_02_params.mth5_name
     station_params = [
@@ -332,6 +356,9 @@ def create_test1_h5_with_nan(
[…]
