+++ aurora/time_series/decorators.py
@@ -1,4 +1,9 @@
 """
+This module is a place to put decorators used by methods in aurora.
+It is a work in progress.
+
+Development notes:
+
 | Here is the decorator pattern
 | def decorator(func):
 |    @functools.wraps(func)
@@ -8,6 +13,7 @@
 |        # Do something after
 |        return value
 |    return wrapper_decorator
+
 """
 
 import functools
@@ -16,7 +22,11 @@ import xarray as xr
 
 def can_use_xr_dataarray(func):
     """
-    Intended as a decorator.  Most of the windowed time series methods are
+    Decorator to allow functions that operate on xr.Dataset to also operate
+    on xr.DataArray, by converting xr.DataArray to xr.Dataset, calling the
+    function, and then  converting back to xr.DataArray
+
+    Most of the windowed time series methods are
     intended to work with xarray.Dataset class.  But I would like to be able
     to pass them xarray.DataArray objects.  This class casts a DataArray to a
     Dataset, runs it through func and casts back to a DataArray.
