+++ aurora/time_series/windowed_time_series.py
@@ -1,62 +1,21 @@
+"""
+
+    This module contains methods associated with operating on windowed time series.
+    i.e. Arrays that have been chunked so that an operator can operate chunk-wise.
+
+    Development Notes:
+     Many of the methods in this module are not currently in use.
+     It looks like I once thought we should have a class to handle windowed time series, but then
+     decided that static methods applied to xr.Dataset would be better.
+
+
+"""
+from aurora.time_series.decorators import can_use_xr_dataarray
+from typing import Optional, Union
+from loguru import logger
 import numpy as np
 import scipy.signal as ssig
-
-from scipy.interpolate import interp1d
-from loguru import logger
-
-from aurora.time_series.decorators import can_use_xr_dataarray
-
-
-def validate_coordinate_ordering_time_domain(dataset):
-    """
-    Check that the data dimensions are what you expect.  This may evolve, but for now,
-    just want to make sure that we are operating along the correct axes when we
-    demean, detrend, taper, etc.
-    Parameters
-    ----------
-    dataset : xarray.Dataset
-
-    Returns
-    -------
-
-    """
-    coordinate_labels = list(dataset.coords.keys())
-    cond1 = coordinate_labels[0] == "within-window time"
-    cond2 = coordinate_labels[1] == "time"
-    if cond1 & cond2:
-        return True
-    else:
-        logger.error("Uncertain that xarray coordinates are correctly ordered")
-        raise Exception
-
-
-def get_time_coordinate_axis(dataset):
-    """
-    It is common to pass an argument to scipy.signal methods axis=int
-    where that integer specifies along which dimension we are applying the
-    operator.  This method helps ensure that we have the correct axis.
-    Parameters
-    ----------
-    dataset : xarray.Dataset
-
-    Returns
-    -------
-
-    """
-    coordinate_labels = list(dataset.coords.keys())
-
-    if len(coordinate_labels) != 2:
-        logger.warning("Warning - Expected two distinct coordinates")
-        # raise Exception
-
-    return coordinate_labels.index("time")
-    # time_coord_indices = [ndx for x, ndx in enumerate(coordinate_labels) if
-    #                                                   x=="time"]
-    # if len(time_coord_indices) == 1:
-    #     return time_coord_indices[0]
-    # else:
-    #     print("expected only one universal time coordinate")
-    #     raise Exception
+import xarray as xr
 
 
 class WindowedTimeSeries(object):
@@ -66,7 +25,7 @@ class WindowedTimeSeries(object):
     This is a place where we can put methods that operate on these sorts of
     objects.
 
-    The assumption is that we take xarrays keyed by "channel"
+    Assumes xr.Datasets keyed by "channel"
 
     Specific methods:
         Demean
@@ -75,138 +34,200 @@ class WindowedTimeSeries(object):
         stft
         invert_prewhitening
 
-        probably make these @staticmethod s so we import WindowedTimeSeries
-        and then call the static methods
+        TODO: Consider making these @staticmethod so import WindowedTimeSeries
+         and then call the static methods
     """
 
     def __init__(self):
+        """Constructor"""
         pass
 
-    @can_use_xr_dataarray
-    @staticmethod
-    def apply_taper(data=None, taper=None, in_place=True):
-        """
-        Point by point multiplication of taper against time series.
-        xarray handles this very cleanly as a direct multiply operation.
-        tapered_obj = windowed_obj * windowing_scheme.taper
-        """
-        data = data * taper
-        return data
+    # Not used 20240723: Commenting out.
+    # @can_use_xr_dataarray
+    # @staticmethod
+    # def apply_taper(data=None, taper=None):  #, in_place=True):
+    #     """
+    #     Point by point multiplication of taper against time series.
+    #
+    #     Not Currently Used.
+    #
[…]
