+++ aurora/time_series/windowing_scheme.py
@@ -1,9 +1,15 @@
 #!/usr/bin/env python2
 # -*- coding: utf-8 -*-
 """
+
+This module is concerned with windowing time series.
+
+
+Development Notes:
+
 The windowing scheme defines the chunking and chopping of the time series for
 the Short Time Fourier Transform.  Often referred to as a "sliding window" or
-a "striding window".  Iin its most basic form it is a taper with a rule to
+a "striding window".  In its most basic form it is a taper with a rule to
 say how far to advance at each stride (or step).
 
 To generate an array of data-windows from a data series we only need the
@@ -39,8 +45,8 @@ When 2D arrays are generated how should we index them?
 |    [12 13 14]]
 
 In this example the rows are indexing the individual windows ... and so they
-should be associated with the time of each window.  We will need to set a
-standard for this.  Obvious options are center_time of window and time_of_first
+should be associated with the time of each window.  Will need a standard for
+this.  Obvious options are center_time of window and time_of_first
 sample. I prefer time_of_first sample.  This can always be transformed to
 center time or another standard later.  We can call this the "window time
 axis".  The columns are indexing "steps of delta-t".  The actual times are
@@ -73,22 +79,31 @@ from mt_metadata.transfer_functions.processing.aurora.decimation_level import (
     get_fft_harmonics,
 )
 from loguru import logger
+from typing import Optional, Union
 
 
 class WindowingScheme(ApodizationWindow):
     """
-    20210415: Casting window length, overlap, advance, etc. in terms of number
-    of samples or "points" here as this is common signal processing the
-    nomenclature.  We may provide an interface to define these things in terms
-    of percent, duration in seconds etc. in a supporting module.
 
-    Note that sample_rate is actually a property of the data and not of the
-    window ... still not sure if we want to make sample_rate an attr here
-    or if its better to put properties like window_duration() as a method of
-    some composition of time series and windowing scheme.
+    Development notes
+    20210415: Casting window length, overlap, advance, etc. in terms of number
+    of samples ("taps", "points") here.  May allow defining these in terms
+    of percent, duration in seconds etc. in future.
+
+    Note: Technically, sample_rate is a property of the data, and not of the
+    window ... but once the window is applied to data, the sample rate is defined.
+    Sample rate is defined here because this window will operate on time series
+    with a defined time axis.
     """
 
     def __init__(self, **kwargs):
+        """
+        Constructor
+
+        Parameters
+        ----------
+        kwargs
+        """
         super(WindowingScheme, self).__init__(**kwargs)
         self.num_samples_overlap = kwargs.get(
             "num_samples_overlap", None
@@ -98,28 +113,36 @@ class WindowingScheme(ApodizationWindow):
         self.sample_rate = kwargs.get("sample_rate", None)
 
     def clone(cls):
+        """return a deepcopy of self"""
         return copy.deepcopy(cls)
 
-    def __str__(self):
+    def __str__(self) -> str:
+        """return a descriptive string"""
         info_string = (
             f"Window of {self.num_samples_window} samples with "
             f"overlap {self.num_samples_overlap}"
         )
-        # add taper summary here?
         return info_string
 
     @property
     def num_samples_advance(self):
         """
-        A derived property.  If we made this a fundamental defined property
-        then overlap would become a derived property.  Overlap is more
-        conventional than advance in the literature however so we choose it as
-        our property label.
+        Returns the number of samples the window advances at each step.
+
+        Development Note:
[…]
