+++ aurora/time_series/window_helpers.py
@@ -2,7 +2,12 @@
 
 This module contains some helper functions that are used when working with sliding windows.
 
-TODO: Not all the functions here are needed, some of them are just examples and tests.
+Development Notes:
+ - Not all the functions here are needed, some of them are just examples and tests.
+ - for example there are three sliding window functions that were considered.
+   The idea was to compare their performance, but currently we use `sliding_window_crude`.
+
+
 
 Notes in google doc:
 https://docs.google.com/document/d/1CsRhSLXsRG8HQxM4lKNqVj-V9KA9iUQAvCOtouVzFs0/edit?usp=sharing
@@ -17,7 +22,6 @@ from numba import jit
 from numpy.lib.stride_tricks import as_strided
 
 
-# Window-to-timeseries relationship
 def available_number_of_windows_in_array(
     n_samples_array: int, n_samples_window: int, n_advance: int
 ) -> int:
@@ -56,6 +60,7 @@ def sliding_window_crude(
     data, num_samples_window, num_samples_advance, num_windows=None
 ):
     """
+    Reshapes input data with a sliding window.
 
     Parameters
     ----------
@@ -90,6 +95,7 @@ def sliding_window_crude(
 @jit(nopython=True)
 def sliding_window_numba(data, num_samples_window, num_samples_advance, num_windows):
     """
+    Reshapes input data with a sliding window.
 
     Parameters
     ----------
@@ -118,24 +124,29 @@ def sliding_window_numba(data, num_samples_window, num_samples_advance, num_wind
 
 def striding_window(data, num_samples_window, num_samples_advance, num_windows=None):
     """
+    Reshapes input data with a sliding window.
+
+    Not currently used.
+
+    Development notes
     Applies a striding window to an array.  We use 1D arrays here.
     Note that this method is extendable to N-dimensional arrays as was once shown
     at  http://www.johnvinyard.com/blog/?p=268
 
-    Karl has an implementation of this code but chose to restict to 1D here.
-    This is becuase of several warnings encountered, on the notes of stride_tricks.py,
+    Karl has an implementation of this code but chose to resrtict to 1D here.
+    This is because of several warnings encountered, on the notes of stride_tricks.py,
     as well as for example here:
     https://stackoverflow.com/questions/4936620/using-strides-for-an-efficient-moving-average-filter
 
-    While we can possibly setup Aurora so that no copies of the strided window are made
+    While we can possibly set up Aurora so that no copies of the strided window are made
     downstream, we cannot guarantee that another user may not add methods that require
-    copies.  For robustness we will use 1d implementation only for now.
+    copies.  To avoid this, use 1d implementation only for now.
 
     Another clean example of this method can be found in the razorback codes from brgm.
 
     result is 2d: result[i] is the i-th window
 
-    >>> sliding_window(np.arange(15), 4, 3, 2)
+    > sliding_window(np.arange(15), 4, 3, 2)
     array([[0, 1, 2],
            [2, 3, 4],
            [4, 5, 6],
@@ -163,13 +174,10 @@ def striding_window(data, num_samples_window, num_samples_advance, num_windows=N
         num_windows = available_number_of_windows_in_array(
             len(data), num_samples_window, num_samples_advance
         )
-    # min_ = (num_windows - 1) * num_samples_advance + num_samples_window
     bytes_per_element = data.itemsize
     output_shape = (num_windows, num_samples_window)
-    # print("output_shape", output_shape)
     strides_shape = (num_samples_advance * bytes_per_element, bytes_per_element)
-    # strides_shape = None
-    logger.info("strides_shape", strides_shape)
+    logger.debug("strides_shape", strides_shape)
[…]
