### 2024-07-24 2c6b383
uncomment used function apply_taper
