### 2024-07-24 2e1c023
uncomment apply_taper
