### 2024-07-24 26873f8
Cruft removal -- towards fix #337
- updated joss ipynb after confirming comparison plot works
