### 2024-07-24 1c9bb2e
partial add doc to plotting functions
- many of these may be unused
- running tests with some deletions
