### 2024-07-24 fec58c2
factor set_rho_limits out of set_lims
- embedded test is passing
