### 2024-07-24 471218e
remove cruft
- embedded sanity check removed
