### 2024-07-24 304cda1
add doc to plotter
- deprecate unused `orient` return arg from get_lims
