### 2024-07-24 eb53f6a
factor legacy methods for plot axes - towards #337 (testing)
