### 2024-07-25 f4b2c88
add doc (#337)
- add doc, typehints,
- some variable rnaming for clarity
- delete/comment some unused functions
