+++ aurora/transfer_function/plot/rho_plot.py
@@ -3,6 +3,8 @@
 
     This is based on Gary's RhoPlot.m in the matlab EMTF version.
 iris_mt_scratch/egbert_codes-20210121T193218Z-001/egbert_codes/matlabPrototype_10-13-20/TF/classes
+
+TODO: replace with calls to mtpy
 """
 import matplotlib.pyplot as plt
 import numpy as np
@@ -19,48 +21,40 @@ class RhoPlot(object):
 
     Development Notes:
         This should be deprecated and replaced with MTpy
+        The only place this class is used is in aurora/sandbox/plot_helpers.py in the
+        plot_tf_obj method.
 
     """
 
     def __init__(self, tf_obj):
+        """
+        Constructor
+
+        TODO: Replace tf_obj with mt_metadata tf if this method not replaced with mtpy.
+
+        Parameters
+        ----------
+        tf_obj: aurora.transfer_function.TTFZ.TTFZ
+            Object with TF information
+
+
+        """
         self.tf = tf_obj
 
-    def rho_phi_plot(self, pred=None):
+    def phase_sub_plot(self, ax, ttl_str="", pred=None, linewidth=2):
         """
+        place a phase subplot on given figure axis
+
+        Development notes:
+         Originally this took an optional input argument `axRect`
+         but it was never used. It looks as it it was intended to be able to set the
+         position of the figure.  There was also some hardcoded control of linewidth
+         and markersize which has been removed for readability.
+
 
         Parameters
         ----------
-        pred
-
-        Returns
-        -------
-
-        """
-        rects = self.set_figure_size()
-        fig, ax = plt.subplots(2, 1, 1)
-        # hfig = figure('Position', rects.Screen, 'PaperPosition',
-        #                rects.Paper, 'Tag', 'rho-phi plot');
-
-        if pred is not None:
-            h_phi = self.phase_sub_plot(ax, rects.phi, pred)
-            h_rho = self.rho_sub_plot(ax, rects.rho, pred)
-        else:
-            h_phi = self.phase_sub_plot(ax, rects.phi)
-            h_rho = self.rho_sub_plot(ax, rects.rho)
-
-        H = {"fig": fig, "rho": h_rho, "phi": h_phi}
-        return H
-
-    def phase_sub_plot(self, ax, ttl_str="", axRect=None, pred=None):
-        """
-        place a phase subplot on a figure, given figure handle and axis
-        postion
-
-        Parameters
-        ----------
-        self
         ax
-        axRect
         pred
 
         Returns
@@ -75,36 +69,31 @@ class RhoPlot(object):
 
         Tmin, Tmax = self.set_period_limits()
         axis_limits = [Tmin, Tmax, 0, 90]
-        # xticks = self.get_xticks()
 
         [xb, yb] = err_log(
             np.transpose(self.tf.periods),
             self.tf.phi[:, 0],
             self.tf.phi_se[:, 0],
-            "XLOG",
             axis_limits,
+            log_x_axis=True,
         )
-        # figure(ax); #need this?
-        # set current axes
-        # hax = plt.axes('Position', axRect);
+
         ax.semilogx(xb, yb, "b-")
         ax.semilogx(self.tf.periods, phi[:, 0], "bo")
-        # print("OK, now set linewidth and markersize")
-        # set(lines, 'LineWidth', 1, 'MarkerSize', 7);
-        # hold on;
+
         xb, yb = err_log(
[…]
