### 2024-07-26 e09cb2e
update doc, add typehints, reduce kwargs (#337)
