+++ aurora/transfer_function/regression/base.py
@@ -1,11 +1,13 @@
 """
-follows Gary's TRegression.m in
+
+This module contains the base class for regression functions.
+ It follows Gary Egbert's EMTF Matlab code TRegression.m in
+ which can be found in
 iris_mt_scratch/egbert_codes-20210121T193218Z-001/egbert_codes/matlabPrototype_10-13-20/TF/classes
 
+This class originally used numpy arrays to make adapting the Matlab easier, but
+experimental support for xarray is now added (2024).
 
-There are some high-level decisions to make about usage of xarray.Dataset,
-xarray.DataArray and numpy arrays.  For now I am going to cast X,Y to numpy
-arrays to follow Gary's codes more easily since his are matlab arrays.
 """
 import numpy as np
 import xarray as xr
@@ -96,11 +98,25 @@ class RegressionEstimator(object):
         X: Union[xr.Dataset, xr.DataArray, np.ndarray],
         Y: Union[xr.Dataset, xr.DataArray, np.ndarray],
         iter_control: IterControl = IterControl(),
-        **kwargs,
+        input_channel_names: Optional[Union[list, None]] = None,
+        output_channel_names: Optional[Union[list, None]] = None,
+        **kwargs,  # n
     ):
         """
+        Constructor
 
         Parameters
+        ----------
+        X: Union[xr.Dataset, xr.DataArray, np.ndarray]
+            The input channels for regression
+        Y: Union[xr.Dataset, xr.DataArray, np.ndarray]
+            The output channels for regression
+        iter_control: IterControl
+            Contains parameters controlling the regression, e.g. convergence criteria, thresholds, etc.
+        input_channel_names: Optional[Union[list, None]]
+            If X is np.ndarray, this allows associating channel names to X's columns
+        output_channel_names: Optional[Union[list, None]]
+            If Y is np.ndarray, this allows associating channel names to Y's columns
 
         ----------
         kwargs
@@ -121,13 +137,14 @@ class RegressionEstimator(object):
         self._QHY = None  #
         self._QHYc = None
         self._n_channels_out = None
-        self._input_channel_names = kwargs.get("input_channel_names", None)
-        self._output_channel_names = kwargs.get("output_channel_names", None)
+        self._input_channel_names = input_channel_names
+        self._output_channel_names = output_channel_names
 
-    def _set_up_regression_variables(self):
+    def _set_up_regression_variables(self) -> None:
         """
-        Placeholder for making this method more general. Currently, the input arguments X, Y are xr.Dataset
-        Here we could add logic for supporting dataarrays as well.
+        Initialize arrays needed for regression and cast any xarray to numpy
+
+        Development Notes:
 
         When xr.Datasets are X, Y we cast to array (num channels x num observations) and then transpose them
         When xr.DataArrays are X, Y extract the array -- but how do we know whether or not to transpose?
@@ -143,26 +160,31 @@ class RegressionEstimator(object):
         self._check_number_of_observations_xy_consistent()
 
     @property
-    def input_channel_names(self):
+    def input_channel_names(self) -> list:
+        """returns the list of channel names associated with X"""
         if self._input_channel_names is None:
             self._input_channel_names = _get_channel_names(self._X, label="IN")
         return self._input_channel_names
 
     @property
-    def output_channel_names(self):
+    def output_channel_names(self) -> list:
+        """returns the list of channel names associated with Y"""
         if self._output_channel_names is None:
             self._output_channel_names = _get_channel_names(self._Y, label="OUT")
         return self._output_channel_names
 
     @property
-    def inverse_signal_covariance(self):
+    def inverse_signal_covariance(self) -> np.ndarray:
[…]
