### 2024-07-26 996cbbd
add doc to regression (#337)
