+++ aurora/transfer_function/regression/m_estimator.py
@@ -1,12 +1,14 @@
 """
-An extension of RegressionEstimator with properties common to Gary's TRME and TRME_RR
-see Notes in RME, RME_RR
+This module contains the MEstimator class - an extension of RegressionEstimator.
+
+ MEstimator has the class methods that are common to both TRME and TRME_RR.
+ See Notes in RME, RME_RR for more details.
+
 """
 import numpy as np
 import xarray as xr
 
 from aurora.transfer_function.regression.base import RegressionEstimator
-from aurora.transfer_function.regression.helper_functions import rme_beta
 from copy import deepcopy
 from loguru import logger
 
@@ -14,6 +16,7 @@ from loguru import logger
 class MEstimator(RegressionEstimator):
     def __init__(self, **kwargs):
         """
+        Constructor.
 
         Parameters
         ----------
@@ -41,26 +44,40 @@ class MEstimator(RegressionEstimator):
         self._residual_variance = None
 
     @property
-    def QHYc(self):
+    def QHYc(self) -> np.ndarray:
+        """
+        Returns the matrix QH @ Yc
+        This is Q.conj().T from the QR decomposition multiplied with the cleaned data.
+
+
+        Returns
+        -------
+        QHYc: np.ndarray
+            A convenience matrix that makes computing the predicted data (QQHYc) more efficient.
+        """
         if self._QHYc is None:
             self.update_QHYc()
         return self._QHYc
 
-    def update_QHYc(self):
+    def update_QHYc(self) -> np.ndarray:
+        """Updates the QHYc matrix with the most recent cleaned data"""
         self._QHYc = self.QH @ self.Yc
 
     @property
-    def Y_hat(self):
+    def Y_hat(self) -> np.ndarray:
+        """returns the most recent predicted data"""
         if self._Y_hat is None:
             self.update_y_hat()
         return self._Y_hat
 
     def update_y_hat(self):
+        """updates the predicted data from the most recent cleaned data"""
         logger.error("Y_hat update method is not defined for abstract MEstimator class")
         logger.error("Try using RME or RME_RR class instead")
         raise Exception
 
     def update_residual_variance(self, correction_factor=1):
+        """updates residual variance from most recent cleaned and predicted data"""
         logger.error(
             "update_residual_variance method not defined in abstract MEstimator"
         )
@@ -68,27 +85,26 @@ class MEstimator(RegressionEstimator):
         raise Exception
 
     @property
-    def residual_variance(self):
+    def residual_variance(self) -> np.ndarray:
+        """returns the residual variance"""
         if self._residual_variance is None:
             self.update_residual_variance()
         return self._residual_variance
 
     @property
-    def r0(self):
+    def r0(self) -> float:
+        """returns the Huber r0 value"""
         return self.iter_control.r0
 
     @property
-    def u0(self):
+    def u0(self) -> float:
+        """returns the u0 threshold for redescending regression step"""
         return self.iter_control.u0
 
-    # @property
-    # def beta(self):
-    #     return rme_beta(self.r0)
-
     @property
-    def correction_factor(self):
+    def correction_factor(self) -> float:
         """
-        *May want to move this out of iter_control, it is really only controlled by r0.
+        Return teh correction factor for the residual variance.
 
         Returns
         -------
@@ -97,18 +113,20 @@ class MEstimator(RegressionEstimator):
         """
         return self.iter_control.correction_factor
 
-    def residual_variance_method1(self):
+    def residual_variance_method1(self) -> np.ndarray:
         """
+        returns the residual variance of the output channels.
+
         This is the method that was originally in TRME_RR.m.  It seems more correct
         than the one in TRME, but also has more computational overhead.
         """
[…]
