### 2024-07-26 c3743b9
Remove dev coherence weights from joss branch
- also comment out reference to it in tf helpers
