### 2024-07-26 simpeg/aurora#337 @kkappler
**WORK IN PROGRESS:**
Below are pasted sinanozaydin's review bullets. I will tag sinanozaydin when all items are addressed and a PR is ready. I will put my responses one level of indentation below the original comments. Comments marked with TODO have not been addressed yet.  Note that because some of the errors/improvements were already addressed in issue #333, the branch to address these comments `fix_issue_337` was forked from `fix_issue_333`.

The authors of this study completed a great deal of work standardizing the MT-data formats and creating a Python version of one of the most used MT processing codes, EMTF. I am sure that it will be a valuable tool for the MT community. However, I think if the authors could reach a bigger user count if they had notebooks/workflows handling other different data formats (Phoenix or regular ASCII, handled through mth5 and mtmetadata). The state of the notebooks is also a bit messy right now and could be improved with some organization. My comments are:

**TODO**: add a general response here

-----------------------------------------------

***Comments on the article:***

Line 60: Small error with dot.
 - Fixed
Line 71: Suggestion: I feel like lot of the captions in Figure 2 can be put into text at this location instead.
- **TODO**
Line 95: Small typo
- **TODO**
Figure 3: This would be better if it was inserted as a markdown code snippet.
 - Done -- see updated manuscript.
 TODO: add link to `updated manuscript` text above

 -----------------------------------------------

**Comments on the code**   **TODO**: address these comments

There are no clear references in the documentation about each file or module does. There are a lot of docstrings, I assume these will be uploaded later into the readthedocs.

Throughout the code, there are a lot of lines that has not been cleared up. For instance there are three versions of prototype_decimate in piplines/time_series_helpers.py. If these are not cleared up, they should be indicated at the docstring about how they are different. Alternatively, these could be moved to a development branch to not confuse the users.

It would also be good to have more informative docstrings in general. They should be clearly indicating what each function does in a sentence or so. A lot of the defs starts with ‘Parameters’ directly without any initial description.

The main page can be decorated with more information (installation, requirements, contact, etc.). It would be good to have information on where to access the examples or tutorials with clear guidelines. Also, the Aurora logo is not rendered at the main page (check the source perhaps).

pandas >v2.0 has to be set as a requirement. Otherwise the datetime conversions will not work properly. Other requirements may needed to be put in config file or setup.py.

At the end of the notebook operate_aurora.ipynb overwrites all files as .xml for different data types at the last block. This needs to be fixed.

edi writer prints out the main header with tab indentation ('\t'), which might be problematic for other people's edi reading scripts. The other header sections of the edi file are created with spaces instead of tabs, which should not be happening for consistency in the format.

A minor suggestion rather than a fix: At the end of operate_aurora.ipynb, if possible, it might be nice to plot the response plots. I know they are generated in the 'process_mth5 scheme', but they do not look very nice. It would perhaps be also nice to call mtpy-v2 here to plot them. This will also showcase the utilities of the ecosystem of codes authors are building together.

Having both a /tutorials and /examples folders is a bit confusing. It would be good to merge these notebook destinations unless there is a good reason not to. It would also be good to have an information text on the main page, readme.txt, or somewhere to describe them briefly one by one.

There are no examples on how to perform the remote reference method.

-----------------------------------------------

**Other notebooks**    **TODO**: address these comments

make_cas04_single_station_h5.ipynb: At the cell https://github.com/simpeg/aurora/issues/12, the code fails to get the data from the servers.

earthscope_magnetic_data_tutorial: At the cell https://github.com/simpeg/aurora/issues/55, the code cannot find the filename “frd20150622vsec.sec” The block that checks whether the file is downloaded or not seem to be not working properly.

pkd_units_check.ipynb: Not clear what this notebook does.

Suggestions that are not neccesary for publication but could be good for further development:  **TODO**: address these comments

It would be really good to have a workflow for Phoenix or other format time-series files rather than just pulling data from the IRIS servers. For instance, having a notebook about ASCII data format or Phoenix binaries, converting that into an mth5 object, and then processing it would be nice. I know it won't be hard to figure this out, especially considering these workflows were already exemplified in mth5 notebooks. However, I feel that most people would just continue doing their work with old Fortran codes if this has not been shown incredibly easily here. At least, it should be clear on the main page that mth5 is where you would find how to convert your time series data into something usable by Aurora.

A notebook about serializing the workflow with multiple stations could be great!
