### 2024-07-26 #6832 @kkappler
Hi @diehlpk ,

I have started a branch fix_issue_337, and have a [Draft PR](https://github.com/simpeg/aurora/pull/338) in place to address #337.   Currently I am working on documentation (making each module start with "This module contains ... ", and adding a descriptive sentence at the top of each method).

This response is started in a google doc.  I will tag @sinanozaydin on issue #337 once all the comments are addressed.  Lets give this a week or two for next status update.
