+++ README.rst
@@ -13,6 +13,16 @@
 .. image:: https://img.shields.io/pypi/l/aurora.svg
     :target: https://pypi.python.org/pypi/aurora
 
-Aurora is a Python library for processing natural source electromagnetic data. It uses MTH5 formatted Magnetotelluric data and a configuration file as inputs and generates transfer functions that can be formatted as EMTF XML or other output formats. 
+Aurora is an open-source package that robustly estimates single station and remote reference electromagnetic transfer functions (TFs) from magnetotelluric (MT) time series.  Aurora is part of an open-source processing workflow that leverages the self-describing data container `MTH5 <https://github.com/kujaku11/mth5>`_, which in turn leverages the general `mt-metadata <https://github.com/kujaku11/mth5>`_ framework to manage metadata.  These pre-existing packages simplify the processing by providing managed data structures, transfer functions to be generated with only a few lines of code.  The processing depends on two inputs -- a table defining the data to use for TF estimation, and a JSON file specifying the processing parameters, both of which are generated automatically, and can be modified if desired.  Output TFs are returned as mt-metadata objects, and can be exported to a variety of common formats for plotting, modeling and inversion.  
+
+Key Features
+-------------
+
+- Tabular data indexing and management (Pandas dataframes), 
+- Dictionary-like processing parameters configuration
+- Programmatic or manual editing of inputs
+- Largely automated workflow 
 
 Documentation for the Aurora project can be found at http://simpeg.xyz/aurora/
+
+
