+++ README.rst
@@ -1,4 +1,4 @@
-.. image:: ../docs/figures/aurora_logo.png
+.. image:: docs/figures/aurora_logo.png
    :width: 900
    :alt: AURORA
 
