+++ docs/api/aurora.transfer_function.weights.rst
@@ -1,0 +1,29 @@
+aurora.transfer\_function.weights package
+=========================================
+
+Submodules
+----------
+
+aurora.transfer\_function.weights.edf\_weights module
+-----------------------------------------------------
+
+.. automodule:: aurora.transfer_function.weights.edf_weights
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+aurora.transfer\_function.weights.spectral\_features module
+-----------------------------------------------------------
+
+.. automodule:: aurora.transfer_function.weights.spectral_features
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: aurora.transfer_function.weights
+   :members:
+   :undoc-members:
+   :show-inheritance:
