+++ docs/api/index.rst
@@ -1,17 +1,0 @@
-.. _api::
-
-API Reference
-=============
-
-.. automodule:: aurora
-    :show-inheritance:
-
-
-.. toctree::
-    :maxdepth: 4
-    :caption: API documentation
-
-    time_series
-    transfer_function
-    pipelines
-    general_helper_functions
