+++ docs/index.rst
@@ -19,11 +19,22 @@ Welcome to Aurora Documentation
    :caption: Examples
 
    examples/operate_aurora.ipynb
+   examples/dataset_definition.ipynb
+   examples/synthetic_data_processing.ipynb
+     
+.. toctree::
+   :maxdepth: 2
+   :caption: Tutorials
+   
+   tutorials/processing_configuration.ipynb
+   tutorials/earthscope_magnetic_data_tutorial.ipynb
+   tutorials/process_cas04_single_station_h5.ipynb
+   tutorials/process_cas04_mulitple_station.ipynb
 
 .. toctree::
    :maxdepth: 4
    :caption: Reference documentation
 
-   api/index.rst
+   api/modules.rst
 
 .. include:: ../CONTRIBUTING.rst
