+++ aurora/transfer_function/weights/edf_weights.py
@@ -1,26 +1,39 @@
 """
-from egbert_codes-20210121T193218Z-001/egbert_codes/matlabPrototype_10-13-20/TF
-/functions/Edfwts.m
+    This module contains a class for computing so-called "Effective Degrees of Freedom" weights.
+
+Development notes:
+The code here is based on the function Edfwts.m from egbert_codes-
+20210121T193218Z-001/egbert_codes/matlabPrototype_10-13-20/TF/functions/Edfwts.m
+
 """
 
 import numpy as np
+import xarray as xr
 from loguru import logger
-
+from typing import Optional, Union
 import aurora.transfer_function.weights.edf_weights
 
 
 class EffectiveDegreesOfFreedom(object):
-    def __init__(self, **kwargs):
+    def __init__(
+        self,
+        edf_l1: Optional[float] = 20.0,
+        alpha: Optional[float] = 0.5,
+        c1: Optional[float] = 2.0,
+        c2: Optional[float] = 10.0,
+        p3: Optional[float] = 5.0,
+        n_data: Optional[int] = 0,
+    ):
         """
+        Constructor.
 
         Parameters
         ----------
-        kwargs
-        n_data : integer
+        n_data : int
             Number of "numeric" observations in the input and remote reference data.
             Any time-frequency index having nan value in any channel is subtracted
             from the total number of observations before computation of p1, p2
-        edfl1 : int or float
+        edf_l1 : float
              the effective number of degrees of freedom at which a datapoint should be
               removed from the total energy calculation in the iterative loop
         alpha : float
@@ -30,19 +43,19 @@ class EffectiveDegreesOfFreedom(object):
         c2 : float
             ad hoc parameter for scaling the p2 threshold
         p3 : float
-            parameter controling the threshold at which the ratio of a data weighting
-            associated with remote and input channel is unaccepatble.
+            parameter controlling threshold at which the ratio of a data weighting
+            associated with remote and input channel is unacceptable.
         """
 
-        self.edfl1 = kwargs.get("edf_l1", 20.0)
-        self.alpha = kwargs.get("alpha", 0.5)
-        self.c1 = kwargs.get("c1", 2.0)
-        self.c2 = kwargs.get("c2", 10.0)
-        self.p3 = kwargs.get("p3", 5)
-        self.n_data = kwargs.get("n_data", 0)
+        self.edf_l1 = edf_l1
+        self.alpha = alpha
+        self.c1 = c1
+        self.c2 = c2
+        self.p3 = p3
+        self.n_data = n_data
 
     @property
-    def p1(self):
+    def p1(self) -> float:
         """
         Threshold applied to edf.  All edf below  this value
         are set to weight=0
@@ -50,64 +63,74 @@ class EffectiveDegreesOfFreedom(object):
         return self.c1 * (self.n_data**self.alpha)
 
     @property
-    def p2(self):
+    def p2(self) -> float:
         """
         Threshold applied to edf.  All edf above th  this value
         are set to weight=0
         """
         return self.c2 * (self.n_data**self.alpha)
 
-    def compute_weights(self, X, use):
+    def compute_weights(self, X: np.ndarray, use: np.ndarray) -> np.ndarray:
         """
+        Compute the EDF Weights
+
+        Development Notes:
         The data covariance matrix s and its inverse h are iteratively recomputed using
         fewer and fewer observations. However, the edf is also computed at every
         iteration but doesn't seem to use any fewer observations.
-        Thus the edf weights change as use drops, EVEN FOR INDICES that were
-        previously computed... this seems like it could be an error.
+        Thus the edf weights change as use drops, _even for indices_ that were
[…]
