+++ aurora/transfer_function/base.py
@@ -1,18 +1,25 @@
 """
-follows Gary's TTF.m in iris_mt_scratch
+This module contains a base class for Transfer function information.
+
+Development Notes:
+ The class is based on Gary's TTF.m in iris_mt_scratch
 egbert_codes-20210121T193218Z-001/egbert_codes/matlabPrototype_10-13-20/TF/classes
+TODO:
+ This class should be updated to use mt_metadata.transfer_function.TF, or replaced by it (issue #203)
+
 """
 
 import numpy as np
 import xarray as xr
-
-# from mt_metadata.base import Base
+from aurora.config.metadata.processing import Processing
 from loguru import logger
+from mt_metadata.transfer_functions.processing.aurora.band import FrequencyBands
+from typing import Optional, Union
 
 
 class TransferFunction:
-    """Class to contain transfer function array.
-
+    """
+    Class to contain transfer function array.
 
     Parameters
     ----------
@@ -32,23 +39,21 @@ class TransferFunction:
     FullCov : boolean
         true if full covariance is provided
 
-    Returns
-    -------
-    properties (Dependent)
-    StdErr % standard errors of TF components, same size and order as TF
-    NBands
-    freqs % inverse of period
-    Nout
-    Nin
     """
 
-    def __init__(self, decimation_level_id: int, frequency_bands, **kwargs):
+    def __init__(
+        self,
+        decimation_level_id: int,
+        frequency_bands: FrequencyBands,
+        processing_config: Optional[Union[Processing, None]] = None,
+        **kwargs
+    ):
         """
+        Constructor.
+
+        Development Notes:
         change 2021-07-23 to require a frequency_bands object.  We may want
-        to just pass the band_edges.  I'm not a fan of forcing dependency of
-        TF on the FrequencyBands class, but it will simplify writting the z_file
-        interfaces.  The old TTF.py class shows an example that just accepted
-        num_bands as an initialation variable.
+        to just pass the band_edges.
 
         Parameters
         ----------
@@ -68,7 +73,7 @@ class TransferFunction:
         self.cov_nn = None
         self.R2 = None
         self.initialized = False
-        self.processing_config = kwargs.get("processing_config", None)
+        self.processing_config = processing_config
 
         if self.emtf_tf_header is not None:
             if self.num_bands is not None:
@@ -76,6 +81,13 @@ class TransferFunction:
 
     @property
     def emtf_tf_header(self):
+        """
+
+        Returns
+        -------
+        emtf_tf_header: None or OrderedDict
+            A legacy construction from the old EMTF codes with some metadata about the TF.
+        """
         if self.processing_config is None:
             logger.info("No header is available without a processing config")
             self._emtf_tf_header = None
@@ -87,6 +99,7 @@ class TransferFunction:
 
     @property
     def tf_header(self):
+        """returns the emtf_header"""
         return self.emtf_tf_header
 
     @property
@@ -110,27 +123,20 @@ class TransferFunction:
         periods = np.flipud(periods)
         return periods
 
-    def _initialize_arrays(self):
+    def _initialize_arrays(self) -> None:
         """
-        There are four separate data strucutres, indexed by period here:
+        Initialize the xarray objects that will contain the TF data.
 
-        TF (num_channels_out, num_channels_in)
-        cov_ss_inv (num_channels_in, num_channels_in, num_bands),
-        Cov_NN (num_channels_out, num_channels_out),
-        R2 num_channels_out)
+        Development Notes:
+        There are four separate data strucutres, indexed by period here:
+        - TF (num_channels_out, num_channels_in)
+        - cov_ss_inv (num_channels_in, num_channels_in, num_bands),
+        - Cov_NN (num_channels_out, num_channels_out),
+        - R2 num_channels_out)
 
[…]
