+++ aurora/transfer_function/transfer_function_collection.py
@@ -1,21 +1,36 @@
 """
-This is a container to hold:
+This module provides a TransferFunctionCollection container to hold:
 1. TransferFunctionHeader
 2. Dictionary of TransferFunction Objects
 
-Note that a single transfer function object is associated with a station,
-which we call the "local_station".  In a database of TFs we could add a column
-for local_station and one for reference station.
+Development Notes:
+ This class could use substantial updating.  There are two main issues to be aware of.
+
+ First, the transfer function objects that is "collects" are themselves things
+ that could use substantial updates.  This is because the aurora TransferFunction
+ objects to not use the mt_metadata.transfer_function.TF and they should.  This
+ is issue #203.
+
+ Second, the Collection can have more than one value for a TF at a given
+ frequency or period because the TFs can be estimated from different
+ decimation levels.
+
+ A single transfer function object is associated with a station, here called
+ the "local_station". In a database of TFs could add a column for
+ local_station and one for reference station.
 """
+import pathlib
 
 import numpy as np
 import xarray as xr
-from loguru import logger
 
+from aurora.config.metadata.processing import Processing
+from aurora.sandbox.io_helpers.zfile_murphy import ZFile
 from aurora.transfer_function.plot.rho_phi_helpers import plot_phi
 from aurora.transfer_function.plot.rho_phi_helpers import plot_rho
 from aurora.general_helper_functions import FIGURES_PATH
-
+from loguru import logger
+from typing import Optional, Union
 
 EMTF_REGRESSION_ENGINE_LABELS = {}
 EMTF_REGRESSION_ENGINE_LABELS["RME"] = "Robust Single Station"
@@ -23,8 +38,13 @@ EMTF_REGRESSION_ENGINE_LABELS["RME_RR"] = "Robust Remote Reference"
 
 
 class TransferFunctionCollection(object):
-    def __init__(self, **kwargs):
+    def __init__(
+        self,
+        tf_dict: Optional[Union[dict, None]] = None,
+        processing_config: Optional[Union[Processing, None]] = None,
+    ):
         """
+        Constructor.
 
         Parameters
         ----------
@@ -33,84 +53,116 @@ class TransferFunctionCollection(object):
         tf_dict: dict
             This is a dictionary of TTFZ objects, one per decimation level.  They are
             keyed by the decimation_level_id, usually integers 0, 1, 2... n_dec
+        processing_config: Processing
+            The processing parameters associated with the TF objects in the dict.
 
         """
-        self.tf_dict = kwargs.get("tf_dict", None)
-        self.processing_config = kwargs.get("processing_config", None)
-        self.labelled_tf = None
+        self.tf_dict = tf_dict
+        self.processing_config = processing_config
         self.merged_tf = None
         self.merged_cov_nn = None
         self.merged_cov_ss_inv = None
 
     @property
-    def header(self):
+    def header(self) -> dict:
+        """returns the tf_header (ordered dict) from the first tf in the dict"""
         return self.tf_dict[0].tf_header
 
     @property
-    def local_station_id(self):
+    def local_station_id(self) -> str:
+        """return the local station id"""
         return self.header.local_station.id
 
     @property
-    def remote_station_id(self):
+    def remote_station_id(self) -> str:
+        """return the remote station id"""
         if self.header.remote_station:
             return self.header.remote_station[0].id
         else:
             return ""
 
     @property
-    def total_number_of_frequencies(self):
+    def total_number_of_frequencies(self) -> int:
+        """
+        Returns int:
+            Count of number of frequencies over all TFs in dict
+        """
         num_frequecies = 0
         for dec_level in self.tf_dict.keys():
             num_frequecies += len(self.tf_dict[dec_level].periods)
         return num_frequecies
 
[…]
