### 2024-07-27 33c6205
remove reference to deleted module
