### 2024-07-27 496f956
merge some minor doc build warnign fixes from update_docs
