+++ aurora/pipelines/transfer_function_kernel.py
@@ -209,7 +209,7 @@ class TransferFunctionKernel(object):
         or not at all.  This assumption can be relaxed in future by using the time_period attribute of the FC layer.
         For now, we proceed with the all-or-none logic.  That is, if a ['survey', 'station_id', 'run_id',] has FCs,
         assume that the FCs are present for the entire run. We assign the "fc" column of dataset_df to have the same
-         boolean value for all rows of same  ['survey', 'station_id', 'run_id',] .
+        boolean value for all rows of same  ['survey', 'station_id', 'run_id',] .
 
         Returns: None
             Modifies self.dataset_df inplace, assigning bools to the "fc" column
