### 2024-08-02 ad6c2c9
updates to the manuscript from USGS reviewer
