+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Aurora: An open-source python implementation of the EMTF package for magnetotelluric data processing using MTH5 and mt_metadata'
+title: 'Aurora: An open-source Python implementation of the EMTF package for magnetotelluric data processing using MTH5 and mt_metadata'
 tags:
   - Python
   - Geophysics
@@ -54,7 +54,7 @@ bibliography: paper.bib
 
 # Summary
 
-The Aurora software package robustly estimates single station and remote reference electromagnetic transfer functions (TFs) from magnetotelluric (MT) time series.  Aurora is part of an open-source processing workflow that leverages the self-describing data container MTH5, which in turn leverages the general mt\_metadata framework to manage metadata.  These pre-existing packages simplify the processing by providing managed data structures, transfer functions to be generated with only a few lines of code.  The processing depends on two inputs -- a table defining the data to use for TF estimation, and a JSON file specifying the processing parameters, both of which are generated automatically, and can be modified if desired.  Output TFs are returned as mt\_metadata objects, and can be exported to a variety of common formats for plotting, modeling and inversion.  
+The Aurora software package robustly estimates single station and remote reference electromagnetic transfer functions (TFs) from magnetotelluric (MT) time series.  Aurora is part of an open-source processing workflow that leverages the self-describing data container MTH5, which in turn leverages the general mt\_metadata framework to manage metadata.  These pre-existing packages simplify the processing workflow by providing managed data structures, transfer functions to be generated with only a few lines of code.  The processing depends on two inputs -- a table defining the data to use for TF estimation and a JSON file specifying the processing parameters, both of which are generated automatically and can be modified if desired.  Output TFs are returned as mt\_metadata objects, and can be exported to a variety of common formats for plotting, modeling, and inversion.  
 
 ## Key Features
 
@@ -66,7 +66,7 @@ The Aurora software package robustly estimates single station and remote referen
 
 # Introduction
 
-MT is a geophysical technique for probing subsurface electrical conductivity using collocated electric and magnetic field measurements.  Field data is collected in the time domain, however the Earth can be approximated as a linear system in the frequency domain.  Therefore, common practice is to estimate the time invariant  (frequency domain) transfer function (TF) between electric and magnetic channels to get information of the Earth's resistivity structure. If measurements are orthogonal, the TF is equivalent to the electrical impedance tensor (Z) [@Vozoff:1991].  
+Magnetotellurics (MT) is a geophysical technique for probing subsurface electrical conductivity using collocated electric and magnetic field measurements.  Field data is collected in the time domain, however the Earth can be approximated as a linear system in the frequency domain.  Therefore, common practice is to estimate the time invariant  (frequency domain) transfer function (TF) between electric and magnetic channels to get information of the Earth's resistivity structure [@egbert2002processing]. If measurements are orthogonal, the TF is equivalent to the electrical impedance tensor (Z) [@Vozoff:1991].  
 
 
 $\begin{bmatrix} E_x \\ E_y \end{bmatrix}
@@ -77,28 +77,28 @@ Z_{yx} & Z_{yy}
 \end{bmatrix}
 \begin{bmatrix} H_x \\ H_y \end{bmatrix}$
  
[…]
