### 2024-08-02 589bb20
removed channel to run summary
