+++ aurora/pipelines/run_summary.py
@@ -36,13 +36,6 @@ Development Notes:
 import copy
 import pandas as pd
 
-
-from mt_metadata.transfer_functions import (
-    ALLOWED_INPUT_CHANNELS,
-)
-from mt_metadata.transfer_functions import (
-    ALLOWED_OUTPUT_CHANNELS,
-)
 import mth5
 from mth5.utils.helpers import initialize_mth5
 from loguru import logger
@@ -157,9 +150,7 @@ class RunSummary:
             run_obj = m.get_run(row.station_id, row.run_id, row.survey)
             runts = run_obj.to_runts()
             if runts.dataset.to_array().data.__abs__().sum() == 0:
-                logger.critical(
-                    "CRITICAL: Detected a run with all zero values"
-                )
+                logger.critical("CRITICAL: Detected a run with all zero values")
                 self.df["valid"].at[i_row] = False
             # load each run, and take the median of the sum of the absolute values
         if drop:
@@ -189,127 +180,6 @@ class RunSummary:
     #     return df
 
 
-def channel_summary_to_run_summary(
-    ch_summary,
-    allowed_input_channels=ALLOWED_INPUT_CHANNELS,
-    allowed_output_channels=ALLOWED_OUTPUT_CHANNELS,
-    sortby=["station_id", "start"],
-):
-    """
-    Method for compressing an mth5 channel_summary into a "run summary" which
-    has one row per run (not one row per channel)
-
-    Devlopment Notes:
-    TODO: replace station_id with station, and run_id with run
-    Note will need to modify: aurora/tests/config$ more test_dataset_dataframe.py
-    TODO: Add logic for handling input and output channels based on channel
-    summary.  Specifically, consider the case where there is no vertical magnetic
-    field, this information is available via ch_summary, and output channels should
-    then not include hz.
-    TODO: Just inherit all the run-level and higher el'ts of the channel_summary,
-    including n_samples?
-
-    When creating the dataset dataframe, make it have these columns:
-    [
-            "station_id",
-            "run_id",
-            "start",
-            "end",
-            "mth5_path",
-            "sample_rate",
-            "input_channels",
-            "output_channels",
-            "remote",
-            "channel_scale_factors",
-        ]
-
-    Parameters
-    ----------
-    ch_summary: mth5.tables.channel_table.ChannelSummaryTable or pandas DataFrame
-       If its a dataframe it is a representation of an mth5 channel_summary.
-        Maybe restricted to only have certain stations and runs before being passed to
-        this method
-    allowed_input_channels: list of strings
-        Normally ["hx", "hy", ]
-        These are the allowable input channel names for the processing.  See further
-        note under allowed_output_channels.
-    allowed_output_channels: list of strings
-        Normally ["ex", "ey", "hz", ]
-        These are the allowable output channel names for the processing.
-        A global list of these is kept at the top of this module.  The purpose of
-        this is to distinguish between runs that have different layouts, for example
-        some runs will have hz and some will not, and we cannot process for hz the
-        runs that do not have it.  By making this a kwarg we sort of prop the door
-        open for more general names (see issue #74).
-    sortby: bool or list
-        Default: ["station_id", "start"]
-
-    Returns
-    -------
-    run_summary_df: pd.Dataframe
-        A table with one row per "acquistion run" that was in the input channel
-        summary table
-    """
-    if isinstance(ch_summary, mth5.tables.channel_table.ChannelSummaryTable):
-        ch_summary_df = ch_summary.to_dataframe()
[…]
