### 2024-08-02 0859121
Update transfer_function_kernel.py
