+++ aurora/pipelines/transfer_function_kernel.py
@@ -75,12 +75,12 @@ class TransferFunctionKernel(object):
         """
         returns a dict of open mth5 objects, keyed by station_id
 
-        A future version of this for multiple station processing may need nested dict with [survey_id][station_id]
+        A future version of this for multiple station processing may need nested dict with [survey_id][station]
 
         Returns
         -------
         mth5_objs : dict
-            Keyed by station_ids.
+            Keyed by stations.
             local station id : mth5.mth5.MTH5
             remote station id: mth5.mth5.MTH5
         """
@@ -140,20 +140,18 @@ class TransferFunctionKernel(object):
                 if not self.is_valid_dataset(row, i_dec_level):
                     continue
                 if row.fc:
-                    row_ssr_str = f"survey: {row.survey}, station_id: {row.station_id}, run_id: {row.run_id}"
+                    row_ssr_str = f"survey: {row.survey}, station: {row.station}, run_id: {row.run_id}"
                     msg = f"FC already exists for {row_ssr_str} -- skipping decimation"
                     logger.info(msg)
                     continue
                 run_xrds = row["run_dataarray"].to_dataset("channel")
                 decimation = self.config.decimations[i_dec_level].decimation
                 decimated_xrds = prototype_decimate(decimation, run_xrds)
-                self.dataset_df["run_dataarray"].at[i] = decimated_xrds.to_array(
-                    "channel"
+                self.dataset_df["run_dataarray"].at[i] = (
+                    decimated_xrds.to_array("channel")
                 )  # See Note 1 above
 
-        msg = (
-            f"Dataset Dataframe Updated for decimation level {i_dec_level} Successfully"
-        )
+        msg = f"Dataset Dataframe Updated for decimation level {i_dec_level} Successfully"
         logger.info(msg)
         return
 
@@ -171,7 +169,9 @@ class TransferFunctionKernel(object):
             The modified DecimationLevel with clock-zero information set.
         """
         if dec_level_config.window.clock_zero_type == "data start":
-            dec_level_config.window.clock_zero = str(self.dataset_df.start.min())
+            dec_level_config.window.clock_zero = str(
+                self.dataset_df.start.min()
+            )
         return dec_level_config
 
     @property
@@ -207,9 +207,9 @@ class TransferFunctionKernel(object):
         intersection of the long run with each of the remote runs. We ignore this for now, selecting only the first
         element of the run_sub_df, under the assumption that FCs have been created for the entire run,
         or not at all.  This assumption can be relaxed in future by using the time_period attribute of the FC layer.
-        For now, we proceed with the all-or-none logic.  That is, if a ['survey', 'station_id', 'run_id',] has FCs,
+        For now, we proceed with the all-or-none logic.  That is, if a ['survey', 'station', 'run',] has FCs,
         assume that the FCs are present for the entire run. We assign the "fc" column of dataset_df to have the same
-         boolean value for all rows of same  ['survey', 'station_id', 'run_id',] .
+         boolean value for all rows of same  ['survey', 'station', 'run',] .
 
         Returns: None
             Modifies self.dataset_df inplace, assigning bools to the "fc" column
@@ -217,15 +217,15 @@ class TransferFunctionKernel(object):
         """
         groupby = [
             "survey",
-            "station_id",
-            "run_id",
+            "station",
+            "run",
         ]
         grouper = self.processing_summary.groupby(groupby)
 
         for (survey_id, station_id, run_id), df in grouper:
             cond1 = self.dataset_df.survey == survey_id
-            cond2 = self.dataset_df.station_id == station_id
[…]
