### 2024-08-02 abb3cc6
Update process_mth5.py
