+++ aurora/pipelines/process_mth5.py
@@ -28,6 +28,7 @@ Note 3: This point in the loop marks the interface between _generation_ of the F
  compute weights for the FCs.
 
 """
+
 import mth5.groups
 
 # =============================================================================
@@ -36,7 +37,9 @@ import mth5.groups
 
 from aurora.pipelines.time_series_helpers import calibrate_stft_obj
 from aurora.pipelines.time_series_helpers import run_ts_to_stft
-from aurora.pipelines.transfer_function_helpers import process_transfer_functions
+from aurora.pipelines.transfer_function_helpers import (
+    process_transfer_functions,
+)
 from aurora.pipelines.transfer_function_kernel import TransferFunctionKernel
 from aurora.pipelines.transfer_function_kernel import station_obj_from_row
 from aurora.sandbox.triage_metadata import triage_run_id
@@ -60,7 +63,9 @@ SUPPORTED_PROCESSINGS = [
 # =============================================================================
 
 
-def make_stft_objects(processing_config, i_dec_level, run_obj, run_xrds, units="MT"):
+def make_stft_objects(
+    processing_config, i_dec_level, run_obj, run_xrds, units="MT"
+):
     """
     Operates on a "per-run" basis.  Applies STFT to all time series in the input run.
 
@@ -98,7 +103,9 @@ def make_stft_objects(processing_config, i_dec_level, run_obj, run_xrds, units="
         ].channel_scale_factors
     elif run_obj.station_metadata.id == processing_config.stations.remote[0].id:
         scale_factors = (
-            processing_config.stations.remote[0].run_dict[run_id].channel_scale_factors
+            processing_config.stations.remote[0]
+            .run_dict[run_id]
+            .channel_scale_factors
         )
 
     stft_obj = calibrate_stft_obj(
@@ -145,7 +152,9 @@ def process_tf_decimation_level(
         The transfer function values packed into an object
     """
     frequency_bands = config.decimations[i_dec_level].frequency_bands_obj()
-    transfer_function_obj = TTFZ(i_dec_level, frequency_bands, processing_config=config)
+    transfer_function_obj = TTFZ(
+        i_dec_level, frequency_bands, processing_config=config
+    )
     dec_level_config = config.decimations[i_dec_level]
     # segment_weights = coherence_weights(dec_level_config, local_stft_obj, remote_stft_obj)
     transfer_function_obj = process_transfer_functions(
@@ -174,7 +183,9 @@ def triage_issue_289(local_stfts: list, remote_stfts: list):
     for i_chunk in range(n_chunks):
         ok = local_stfts[i_chunk].time.shape == remote_stfts[i_chunk].time.shape
         if not ok:
-            logger.warning("Mismatch in FC array lengths detected -- Issue #289")
+            logger.warning(
+                "Mismatch in FC array lengths detected -- Issue #289"
+            )
             glb = max(
                 local_stfts[i_chunk].time.min(),
                 remote_stfts[i_chunk].time.min(),
@@ -185,13 +196,18 @@ def triage_issue_289(local_stfts: list, remote_stfts: list):
             )
             cond1 = local_stfts[i_chunk].time >= glb
             cond2 = local_stfts[i_chunk].time <= lub
-            local_stfts[i_chunk] = local_stfts[i_chunk].where(cond1 & cond2, drop=True)
+            local_stfts[i_chunk] = local_stfts[i_chunk].where(
+                cond1 & cond2, drop=True
+            )
             cond1 = remote_stfts[i_chunk].time >= glb
             cond2 = remote_stfts[i_chunk].time <= lub
             remote_stfts[i_chunk] = remote_stfts[i_chunk].where(
[…]
