### 2024-08-02 ba102e0
reduce sphinx warnings
