+++ aurora/test_utils/synthetic/paths.py
@@ -2,10 +2,9 @@
 This module contains a class that helps manage data paths for testing aurora on synthetic data.
 
 Development Notes:
-    - The DATA_PATH from general_helper_functions has traditionally had the
-     synthetic ascii data, but this is now stored in MTH5.
-    - This class was built to handle Issue #303 (installation on read-only file system).
-     https://github.com/simpeg/aurora/issues/303
+ - The DATA_PATH from general_helper_functions has traditionally had the synthetic ascii data, but this is now stored in MTH5.
+ - This class was built to handle Issue #303 (installation on read-only file system). https://github.com/simpeg/aurora/issues/303
+
 """
 import pathlib
 
