### 2024-08-03 #6832 @kkappler
Hi @diehlpk , we have addressed @sinanozaydin 's [review](https://github.com/simpeg/aurora/issues/337) with inline responses [here](https://github.com/simpeg/aurora/issues/337#issuecomment-2253058516)
