### 2024-08-03 0911980
Update transfer_function_kernel.py
