+++ aurora/pipelines/transfer_function_kernel.py
@@ -785,8 +785,8 @@ def station_obj_from_row(row):
 
     """
     if row.mth5_obj.file_version == "0.1.0":
-        station_obj = row.mth5_obj.stations_group.get_station(row.station_id)
+        station_obj = row.mth5_obj.stations_group.get_station(row.station)
     elif row.mth5_obj.file_version == "0.2.0":
         survey_group = row.mth5_obj.surveys_group.get_survey(row.survey)
-        station_obj = survey_group.stations_group.get_station(row.station_id)
+        station_obj = survey_group.stations_group.get_station(row.station)
     return station_obj
