### 2024-08-05 simpeg/aurora#337 @sinanozaydin
Hello @kkappler,

Thanks a lot for addressing my issues. With all the updates, all seems to be good from my side! I will form an issue later this week regarding "make_cas04_single_station_h5.ipynb".

@diehlpk
