### 2024-08-05 simpeg/aurora#333 @blsqr
Thank you for implementing the changes, @kkappler!

Could you perhaps upload a revised version of the paper proof PDF here (or merge and then generate a new proof in the JOSS review issue), such that I can inspect the changes in rendered form?
