### 2024-08-05 simpeg/aurora#333 @blsqr
My issues regarding the software have been addressed and I will update the review checklist accordingly. Thank you, @kkappler!

From my side, the corresponding PR #335 can be merged. Once merged, could you please point me to the most up-to-date branch such that I can continue the review there? Is it #338 or #340 or some other branch?
