### 2024-08-05 simpeg/aurora#337 @kkappler
Thanks @sinanozaydin -- based on that feedback, merging this into the [first review JOSS branch](https://github.com/simpeg/aurora/issues/333)
