### 2024-08-06 264c5f0
Update index.rst
add doc for synthetic data processing notebook
