+++ docs/index.rst
@@ -34,6 +34,7 @@ Welcome to Aurora Documentation
    tutorials/earthscope_magnetic_data_tutorial.ipynb
    tutorials/process_cas04_single_station_h5.ipynb
    tutorials/process_cas04_mulitple_station.ipynb
+   tutorials/synthetic_data_processing.ipynb
 
 .. toctree::
    :maxdepth: 4
