### 2024-08-06 simpeg/aurora#333 @kkappler
Thanks @blsqr we have now merged all the review branches into the [`joss`](https://github.com/simpeg/aurora/tree/joss) branch.  That is the latest branch that relates to this manuscript.

@diehlpk
