### 2024-08-09 simpeg/aurora#333 @diehlpk
@blsqr could zou please have a look and check if all your concerns were addressed.
