### 2024-08-09 simpeg/aurora#333 @kkappler
> Thank you for implementing the changes, @kkappler!
>
> Could you perhaps upload a revised version of the paper proof PDF here (or merge and then generate a new proof in the JOSS review issue), such that I can inspect the changes in rendered form?
>
> EDIT: Nevermind, found it [here](https://github.com/simpeg/aurora/blob/f76f24186976811502e16caaf2bd7a071c2adc56/paper/paper.pdf).

@blsqr please note that the latest version of the draft manuscript is now merged into the `joss` branch and can be found [here](https://github.com/simpeg/aurora/blob/joss/paper/paper.pdf)
