### 2024-08-09 simpeg/aurora#333 @blsqr
Thank you, @kkappler!

I have two minor points remaining regarding the paper:

* Fig. 3 caption ends with `The plotting details are in the Jupyter notebook(.`: remove superfluous `(`
* Last reference, Vozoff K. (1991): proper citation is missing. I've found this book chapter, is that correct? https://library.seg.org/doi/10.1190/1.9781560802686.ch8 If so, please update this reference.
