### 2024-08-09 simpeg/aurora#346 @blsqr
[JOSS Review] Add a CITATION.cff file
(Part of the review process of [a submission to JOSS](https://github.com/openjournals/joss-reviews/issues/6832))

I would suggest to add a [`CITATION.cff` file](https://citation-file-format.github.io) to this repository, populated with citation information of the (upcoming) JOSS publication. Also note the [GitHub docs](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-citation-files) regarding this topic.

Of course this can only be done once the paper has been fully published and a DOI is available.
