### 2024-08-09 #6832 @blsqr
@diehlpk I have also concluded my checklist and support publication of the submitted software paper in JOSS.

The aurora package seems to address an important use case in magnetotelluric data processing, building on and extending the existing ecosystem. The submitted software paper gives a good overview of the main ideas and provides examples of how to use `aurora` for data analysis, starting from common dataset formats; while I'm not from this domain and probably can't assess shortcomings, the approaches taken seem convenient and well-implemented. Furthermore, it seems like this project is under very active development; the planned co-development of aurora with key packages of the domain seems like a very good idea.

Thanks to @kkappler for quickly addressing the issues raised during this review. If there are any other questions, feel free to reach out.
