### 2024-08-09 3ca63e2
fix typo in figure caption (#333)
