+++ paper/paper.bib
@@ -157,15 +157,18 @@
   publisher={Frontiers Media SA}
 }
 
-@article{Vozoff:1991,
-  	url = {https://pubs.geoscienceworld.org/seg/books/book/2087/chapter-abstract/114406941/THE-MAGNETOTELLURIC-METHOD?redirectedFrom=fulltext},
-  	Author = {{Vozoff}, K.},
-    Booktitle = {Electromagnetic Methods in Applied Geophysics: Volume 2, Application, Parts A and B. Editted by MIsac N. Nabighian  ~ISBN 1-560800-22-4. ~Published by SEG, Tulsa, OK USA, 1991.},
-  	Title = {{The Magnetotelluric Method}},
-  	Year = 1991
+@incollection{Vozoff:1991,
+    author = {Vozoff, K.},
+    isbn = {9781560800224},
+    title = "{THE MAGNETOTELLURIC METHOD}",
+    booktitle = "{Electromagnetic Methods in Applied Geophysics: Volume 2, Application, Parts A and B}",
+    publisher = {Society of Exploration Geophysicists},
+    year = {1991},
+    month = {01},
+    doi = {10.1190/1.9781560802686.ch8},
+    url = {https://doi.org/10.1190/1.9781560802686.ch8},
 }
 
-ls
 @misc{mthotel,
   author = {B Friedrichs},
   title = {MTHotel},
