### 2024-08-09 simpeg/aurora#333 @kkappler
@blsqr Thanks for catching those errors.  They have been fixed on `joss` branch ([here](https://github.com/simpeg/aurora/blob/joss/paper/paper.pdf)).
