+++ docs/index.rst
@@ -1,3 +1,10 @@
+Welcome to Aurora Documentation
+===================================
+
+.. image:: figures/aurora_logo.png
+   :width: 900
+   :alt: AURORA
+
 .. aurora documentation master file, created by
    sphinx-quickstart on Sun Aug 22 14:43:40 2021.
    You can adapt this file completely to your liking, but it should at least
@@ -16,9 +23,17 @@
    :caption: Examples
 
    examples/operate_aurora.ipynb
+   examples/dataset_definition.ipynb
+
 
 .. toctree::
    :maxdepth: 2
+   :caption: Tutorials
+   
+   tutorials/processing_configuration.ipynb
+
+.. toctree::
+   :maxdepth: 4
    :caption: Reference documentation
 
    api/index.rst
