### 2024-08-10 f7a5608
Add temporary workaround for RunSummary while migrating to mth5
- this allows station and run columns to be available (as well as station_id, and run_id)
