+++ aurora/pipelines/run_summary.py
@@ -26,40 +26,39 @@ maximize coverage of the local station runs is generated
 
 Development Notes:
     TODO: consider adding methods:
-     "drop_runs_shorter_than": removes short runs from summary
-     "fill_gaps_by_time_interval": allows runs to be merged if gaps between are short
-     "fill_gaps_by_run_names": allows runs to be merged if gaps between are short
-    TODO: Consider whether this should return a copy or modify in-place when querying the df.
+     - drop_runs_shorter_than": removes short runs from summary
+     - fill_gaps_by_time_interval": allows runs to be merged if gaps between 
+       are short
+     - fill_gaps_by_run_names": allows runs to be merged if gaps between are 
+       short
+    TODO: Consider whether this should return a copy or modify in-place when
+    querying the df.
 
 """
 
+# =============================================================================
+# Imports
+# =============================================================================
 import copy
+from typing import Optional, Union
+
 import pandas as pd
+from loguru import logger
+
+from aurora.pipelines import RUN_SUMMARY_COLUMNS, MINI_SUMMARY_COLUMNS
 
 import mth5
 from mth5.utils.helpers import initialize_mth5
-from loguru import logger
-from typing import Optional, Union
 
-RUN_SUMMARY_COLUMNS = [
-    "survey",
-    "station",
-    "run",
-    "start",
-    "end",
-    "sample_rate",
-    "input_channels",
-    "output_channels",
-    "remote",
-    "mth5_path",
-]
+# =============================================================================
 
 
 class RunSummary:
     """
     Class to contain a run-summary table from one or more mth5s.
-            "
-    WIP: For the full MMT case this may need modification to a channel based summary.
+
+    WIP: For the full MMT case this may need modification to a channel based
+    summary.
 
 
     """
@@ -79,13 +78,48 @@ class RunSummary:
         self.column_dtypes = [str, str, pd.Timestamp, pd.Timestamp]
         self._input_dict = input_dict
         self.df = df
-        self._mini_summary_columns = [
-            "survey",
-            "station",
-            "run",
-            "start",
-            "end",
-        ]
+        self._mini_summary_columns = MINI_SUMMARY_COLUMNS
+
+    def __str__(self):
+        return str(self.mini_summary.head())
+
+    def __repr__(self):
+        return self.__str__()
+
+    @property
+    def df(self):
+        return self._df
+
+    @df.setter
+    def df(self, value):
+        """
+        Make sure the data frame is set properly with proper column names
+
+        :param value: DESCRIPTION
+        :type value: TYPE
+        :return: DESCRIPTION
+        :rtype: TYPE
+
+        """
+        if value is None:
+            self._df = None
+            return
+
+        if not isinstance(value, pd.DataFrame):
+            msg = f"Need to set df with a Pandas.DataFrame not type({type(value)})"
+            logger.error(msg)
+
+            raise TypeError(msg)
+
+        need_columns = []
+        for col in RUN_SUMMARY_COLUMNS:
+            if not col in value.columns:
+                need_columns.append(col)
+        if need_columns:
+            msg = f"DataFrame needs columns {', '.join(need_columns)}"
+            logger.error(msg)
+            raise ValueError(msg)
+        self._df = value
 
     def clone(self):
         """
@@ -96,9 +130,20 @@ class RunSummary:
         return copy.deepcopy(self)
 
     def from_mth5s(self, mth5_list):
-        """Iterates over mth5s in list and creates one big dataframe summarizing the runs"""
+        """Iterates over mth5s in list and creates one big dataframe
+        summarizing the runs
+        """
         run_summary_df = extract_run_summaries_from_mth5s(mth5_list)
[…]
