+++ aurora/transfer_function/kernel_dataset.py
@@ -57,25 +57,29 @@ TODO: Might need to groupby survey & station, for now consider station_id  uniqu
 
 """
 
+# =============================================================================
+# Imports
+# =============================================================================
+from pathlib import Path
 import copy
-
-import mt_metadata.timeseries
-import mth5.timeseries.run_ts
-import pandas as pd
-
-from loguru import logger
-
-from aurora.pipelines.run_summary import RUN_SUMMARY_COLUMNS
-from aurora.pipelines.run_summary import RunSummary
-from mt_metadata.utils.list_dict import ListDict
 from typing import Optional, Union
 
-# TODO: Add these to a standard, so we track add/subtract columns
-KERNEL_DATASET_COLUMNS = RUN_SUMMARY_COLUMNS + [
-    "channel_scale_factors",
-    "duration",
-    "fc",
-]
+import pandas as pd
+from loguru import logger
+
+import mt_metadata.timeseries
+from mt_metadata.utils.list_dict import ListDict
+
+import mth5.timeseries.run_ts
+from mth5.utils.helpers import initialize_mth5
+
+from aurora.pipelines.run_summary import RunSummary
+from aurora.transfer_function import (
+    KERNEL_DATASET_DTYPE,
+    MINI_SUMMARY_COLUMNS,
+)
+
+# =============================================================================
 
 
 class KernelDataset:
@@ -145,15 +149,103 @@ class KernelDataset:
         self.df = df
         self.local_station_id = local_station_id
         self.remote_station_id = remote_station_id
-        self._mini_summary_columns = [
-            "survey",
-            "station",
-            "run",
-            "start",
-            "end",
-            "duration",
-        ]
+        self._mini_summary_columns = MINI_SUMMARY_COLUMNS
         self.survey_metadata = {}
+        self.initialized = False
+        self.local_mth5_obj = None
+        self.remote_mth5_obj = None
+
+    def __str__(self):
+        return str(self.mini_summary.head())
+
+    def __repr__(self):
+        return self.__str__()
+
+    # def __iter__(self):
+    #     """
+    #     Iterate over rows in the dataframe
+
+    #     :return: DESCRIPTION
+    #     :rtype: TYPE
+
+    #     """
+
+    #     return self.df.iterrows()[0]
+
+    @property
+    def df(self):
+        return self._df
+
+    @df.setter
+    def df(self, value):
+        """
+        Make sure the data frame is set properly with proper column names
+
+        :param value: DESCRIPTION
+        :type value: TYPE
+        :return: DESCRIPTION
+        :rtype: TYPE
+
+        """
+        if value is None:
+            self._df = None
+            return
+
+        if not isinstance(value, pd.DataFrame):
+            msg = f"Need to set df with a Pandas.DataFrame not type({type(value)})"
+            logger.error(msg)
+
+            raise TypeError(msg)
+
+        self._df = self._add_duration_column(
+            self._set_datetime_columns(self._add_columns(value)), inplace=False
+        )
+
+    def _has_df(self):
+        """
+        check to see if dataframe is set
+        """
+        if self._df is not None:
+            if not self._df.empty:
+                return True
+            return False
+        return False
+
+    def _df_has_local_station_id(self):
+        """
+        Check to make sure the dataframe has the local station id
+
+        :return: DESCRIPTION
+        :rtype: bool
+
+        """
+        if self._has_df():
+            return (self._df.station == self.local_station_id).any()
+
+    def _df_has_remote_station_id(self):
+        """
+        Check to make sure the dataframe has the local station id
+
+        :return: DESCRIPTION
+        :rtype: bool
+
+        """
+        if self._has_df():
+            return (self._df.station == self.remote_station_id).any()
+
[…]
