+++ .github/workflows/tests.yml
@@ -35,7 +35,7 @@ jobs:
         python --version
         conda install -c conda-forge pytest pytest-cov certifi">=2017.4.17" pandoc
         pip install -r requirements-dev.txt
-        pip install git+https://github.com/kujaku11/mt_metadata.git@anna # main
+        pip install git+https://github.com/kujaku11/mt_metadata.git@main
         pip install git+https://github.com/kujaku11/mth5.git@add_aurora_tools 
 
     - name: Install Our Package
