+++ tests/transfer_function/test_kernel_dataset.py
@@ -1,11 +1,22 @@
+# =============================================================================
+# Imports
+# =============================================================================
+from pathlib import Path
 import pandas as pd
 import unittest
 
+from mth5.data.make_mth5_from_asc import MTH5_PATH, create_test12rr_h5
+
+from aurora.transfer_function import KERNEL_DATASET_COLUMNS
 from aurora.pipelines.run_summary import RunSummary
-from aurora.test_utils.synthetic.make_mth5_from_asc import create_test12rr_h5
-from aurora.transfer_function.kernel_dataset import intervals_overlap
-from aurora.transfer_function.kernel_dataset import overlap
-from aurora.transfer_function.kernel_dataset import KernelDataset
+
+from aurora.transfer_function.kernel_dataset import (
+    intervals_overlap,
+    overlap,
+    KernelDataset,
+)
+
+# =============================================================================
 
 
 class TestKernelDataset(unittest.TestCase):
@@ -13,46 +24,75 @@ class TestKernelDataset(unittest.TestCase):
 
     @classmethod
     def setUpClass(self):
-        self._mth5_path = create_test12rr_h5()
-        self._run_summary = RunSummary()
-        self._run_summary.from_mth5s(
-            [
-                self._mth5_path,
-            ]
-        )
+        self.mth5_path = MTH5_PATH.joinpath("test12rr.h5")
+        if not self.mth5_path.exists():
+            self.mth5_path = create_test12rr_h5()
+        self.run_summary = RunSummary()
+        self.run_summary.from_mth5s([self.mth5_path])
 
     def setUp(self):
-        self.run_summary = self._run_summary.clone()
+        self.run_summary = self.run_summary.clone()
         self.kd = KernelDataset()
         self.kd.from_run_summary(self.run_summary, "test1", "test2")
 
+    def test_df_columns(self):
+        self.assertListEqual(
+            sorted(KERNEL_DATASET_COLUMNS), sorted(self.kd.df.columns)
+        )
+
+    def test_set_df_fail_bad_type(self):
+        def set_df(value):
+            self.kd.df = value
+
+        self.assertRaises(TypeError, set_df, 10)
+
+    def test_set_df_fail_bad_df(self):
+        def set_df(value):
+            self.kd.df = value
+
+        self.assertRaises(ValueError, set_df, pd.DataFrame({"test": [0]}))
+
+    def test_df_shape(self):
+        self.assertEqual((2, 20), self.kd.df.shape)
+
     def test_exception_from_empty_run_summary(self):
-        # make the run summary df empty
-        self.run_summary.df.valid = False
-        self.run_summary.drop_invalid_rows()
+        self.run_summary.df.loc[:, "has_data"] = False
+        self.run_summary.drop_no_data_rows()
         with self.assertRaises(ValueError):  # as context:
             self.kd.from_run_summary(self.run_summary, "test1", "test2")
 
     def test_clone_dataframe(self):
         cloned_df = self.kd.clone_dataframe()
 
-        # fc column is None so this wont be true
-        self.assertFalse((cloned_df == self.kd.df).all().all())
+        # need to change na to a value
+        cloned_df.fillna(0, inplace=True)
+        self.kd.df.fillna(0, inplace=True)
 
-        cloned_df["fc"] = False
-        self.kd.df["fc"] = False
-        assert (cloned_df == self.kd.df).all().all()
+        # fc column is None so this wont be true
+        self.assertTrue((cloned_df == self.kd.df).all().all())
 
     def test_clone(self):
         clone = self.kd.clone()
 
-        # fc column is None so this wont be true
-        self.assertFalse((clone.df == self.kd.df).all().all())
[…]
