### 2024-08-14 eabf5c8
removed arguments for input/output channels
KernelDataset now has properties for input/output channels that it gets from the dataframe, so should be consistent with the data.
