+++ tests/transfer_function/test_kernel_dataset.py
@@ -7,14 +7,12 @@ import unittest
 
 from mth5.data.make_mth5_from_asc import MTH5_PATH, create_test12rr_h5
 
-from aurora.transfer_function import KERNEL_DATASET_COLUMNS
-from aurora.pipelines.run_summary import RunSummary
+from mtpy.processing import KERNEL_DATASET_COLUMNS
+from mtpy.processing.run_summary import RunSummary
 
-from aurora.transfer_function.kernel_dataset import (
-    intervals_overlap,
-    overlap,
-    KernelDataset,
-)
+from mtpy.processing.kernel_dataset import intervals_overlap
+from mtpy.processing.kernel_dataset import overlap
+from mtpy.processing.kernel_dataset import KernelDataset
 
 # =============================================================================
 
@@ -90,6 +88,32 @@ class TestKernelDataset(unittest.TestCase):
         mini_df = self.kd.mini_summary
         self.assertEqual(str(mini_df.head()), str(self.kd))
 
+    def test_input_channels(self):
+        self.assertListEqual(["hx", "hy"], self.kd.input_channels)
+
+    def test_output_channels(self):
+        self.assertListEqual(["ex", "ey", "hz"], self.kd.output_channels)
+
+    def test_local_df(self):
+        with self.subTest("shape"):
+            self.assertEqual(self.kd.local_df.shape, (1, 20))
+        with self.subTest("local station only length"):
+            self.assertEqual(len(self.kd.local_df.station.unqiue()), 1)
+        with self.subTest("local station only"):
+            self.assertListEqual(
+                list(self.kd.local_df.station.unique()), ["test1"]
+            )
+
+    def test_remote_df(self):
+        with self.subTest("shape"):
+            self.assertEqual(self.kd.remote_df.shape, (1, 20))
+        with self.subTest("remote station only length"):
+            self.assertEqual(len(self.kd.remote_df.station.unqiue()), 1)
+        with self.subTest("remote station only"):
+            self.assertListEqual(
+                list(self.kd.remote_df.station.unique()), ["test1"]
+            )
+
     # @classmethod
     # def tearDownClass(self):
     #     self.mth5_path.unlink()
@@ -113,11 +137,15 @@ class TestOverlapFunctions(unittest.TestCase):
         # hours
 
         # shift the interval forward, leave it overlapping
-        self.ti2_start = self.ti1_start + pd.Timedelta(hours=self.shift_1_hours)
+        self.ti2_start = self.ti1_start + pd.Timedelta(
+            hours=self.shift_1_hours
+        )
         self.ti2_end = self.ti1_end + pd.Timedelta(hours=self.shift_1_hours)
 
         # shift the interval forward, non-verlapping
-        self.ti3_start = self.ti1_start + pd.Timedelta(hours=self.shift_2_hours)
+        self.ti3_start = self.ti1_start + pd.Timedelta(
+            hours=self.shift_2_hours
+        )
         self.ti3_end = self.ti1_end + pd.Timedelta(hours=self.shift_2_hours)
 
     def test_overlaps_boolean(self):
