### 2024-08-14 3a63d32
put back in the option to set input/output channels
propbably makes sense to keep the option to specify the input/output channels in case you don't want to process a certain channel like hz
