+++ tests/parkfield/test_process_parkfield_run.py
@@ -75,7 +75,9 @@ def test():
     test_processing(z_file_path=z_file_path, test_clock_zero="data start")
 
     # COMPARE WITH ARCHIVED Z-FILE
-    auxilliary_z_file = PARKFIELD_PATHS["emtf_results"].joinpath("PKD_272_00.zrr")
+    auxilliary_z_file = PARKFIELD_PATHS["emtf_results"].joinpath(
+        "PKD_272_00.zrr"
+    )
     if z_file_path.exists():
         compare_two_z_files(
             z_file_path,
@@ -89,7 +91,9 @@ def test():
             xlims=[0.05, 500],
         )
     else:
-        logger.error("Z-File not found - Parkfield tests failed to generate output")
+        logger.error(
+            "Z-File not found - Parkfield tests failed to generate output"
+        )
         logger.warning("NCEDC probably not returning data")
 
 
