+++ aurora/config/config_creator.py
@@ -5,11 +5,13 @@ The processing config is still evolving and this class and its methods may chang
 
 """
 
-from aurora.config.metadata.processing import Processing
 from aurora.config import BANDS_DEFAULT_FILE
 from aurora.sandbox.io_helpers.emtf_band_setup import EMTFBandSetupFile
 from loguru import logger
 from mt_metadata.transfer_functions.processing.aurora.window import Window
+from mt_metadata.transfer_functions.processing.aurora.processing import (
+    Processing,
+)
 from typing import Optional, Union
 import pathlib
 
@@ -39,33 +41,32 @@ class ConfigCreator:
         self._band_edges = band_edges
         self._band_specification_style = None
 
-    def processing_id(self, kernel_dataset):
-        """
-        Generates a string id label for the processing config: WIP.
+    # def processing_id(self, kernel_dataset):
+    #     """
+    #     Generates a string id label for the processing config: WIP.
 
-        In the past, we used f"{local}-{remote}" or  f"{local}-{run_id}"
-        Neither of these is sufficiently unique.  In fact, they only describe the
-        dataset, and not the processing config.  It is difficult to see how to make a
-        comprehensive, unique id without it being very long or involving hash
-        functions.
+    #     In the past, we used f"{local}-{remote}" or  f"{local}-{run_id}"
+    #     Neither of these is sufficiently unique.  In fact, they only describe the
+    #     dataset, and not the processing config.  It is difficult to see how to make a
+    #     comprehensive, unique id without it being very long or involving hash
+    #     functions.
 
-        For now, will try to use {local}-{local_runs}-{remote}-{remote_runs},
-        which at least describes the dataset, then a string can be generated by the
-        config and appended if needed.
+    #     For now, will try to use {local}-{local_runs}-{remote}-{remote_runs},
+    #     which at least describes the dataset, then a string can be generated by the
+    #     config and appended if needed.
 
+    #     Parameters
+    #     ----------
+    #     kernel_dataset: aurora.transfer_function.kernel_dataset.KernelDataset
+    #         An object that defines the data to be processed.
 
-        Parameters
-        ----------
-        kernel_dataset: aurora.transfer_function.kernel_dataset.KernelDataset
-            An object that defines the data to be processed.
-
-        Returns
-        -------
-        id: str
-            A label for the processing config.
-        """
-        id = f"{kernel_dataset.local_station_id}-{kernel_dataset.remote_station_id}"
-        return id
+    #     Returns
+    #     -------
+    #     id: str
+    #         A label for the processing config.
+    #     """
+    #     id = f"{kernel_dataset.local_station_id}-{kernel_dataset.remote_station_id}"
+    #     return id
 
     @property
     def band_specification_style(self):
@@ -120,6 +121,8 @@ class ConfigCreator:
     def create_from_kernel_dataset(
         self,
         kernel_dataset,
+        input_channels=None,
+        output_channels=None,
         estimator: Optional[Union[str, None]] = None,
         emtf_band_file: Optional[Union[str, pathlib.Path, None]] = None,
         band_edges: Optional[Union[dict, None]] = None,
@@ -173,8 +176,10 @@ class ConfigCreator:
         processing_obj: aurora.config.metadata.processing.Processing
             Object storing the processing parameters.
         """
-        processing_id = self.processing_id(kernel_dataset)
-        processing_obj = Processing(id=processing_id)  # , **kwargs)
+
+        processing_obj = Processing(
+            id=kernel_dataset.processing_id
+        )  # , **kwargs)
 
         # pack station and run info into processing object
         processing_obj.stations.from_dataset_dataframe(kernel_dataset.df)
@@ -224,8 +229,16 @@ class ConfigCreator:
         if self.band_specification_style == "EMTF":
[…]
