+++ aurora/transfer_function/kernel_dataset.py
@@ -381,6 +381,17 @@ class KernelDataset:
             return False
         else:
             return self.remote_mth5_path.exists()
+            
+    @property
+    def processing_id(self):
+        """its difficult to come put with unique ids without crazy long names
+        so this is a generic id of local-remote, the station metadata
+        will have run information and the config parameters.
+        """
+        if self.remote_station_id is not None:
+            return f"{self.local_station_id}-rr_{self.remote_station_id}"
+        else:
+            return self.local_station_id
 
     @property
     def input_channels(self):
