+++ aurora/transfer_function/kernel_dataset.py
@@ -381,7 +381,7 @@ class KernelDataset:
             return False
         else:
             return self.remote_mth5_path.exists()
-            
+
     @property
     def processing_id(self):
         """its difficult to come put with unique ids without crazy long names
@@ -389,9 +389,12 @@ class KernelDataset:
         will have run information and the config parameters.
         """
         if self.remote_station_id is not None:
-            return f"{self.local_station_id}-rr_{self.remote_station_id}"
+            return (
+                f"{self.local_station_id}-rr_{self.remote_station_id}_"
+                f"sr{int(self.sample_rate)}"
+            )
         else:
-            return self.local_station_id
+            return f"{self.local_station_id}_sr{int(self.sample_rate)}"
 
     @property
     def input_channels(self):
