### 2024-08-14 #6832 @diehlpk
@kkappler if you are done with [346](https://github.com/simpeg/aurora/issues/346), please work on the **Additional Author Tasks After Review is Complete** above to proceed with the publication process.
