### 2024-08-15 a98e030
add flatten method to spectrogram
