### 2024-08-15 0ee8247
add init 1d xarray method
