### 2024-08-15 e29ca89
move 2d xr init to helpers/add test
