### 2024-08-16 65b7a24
Add cross_power and some tests
