+++ aurora/transfer_function/cross_power.py
@@ -1,0 +1,487 @@
+"""
+    Computes transfer functions from covariance matrix of cross powers.
+
+    Development notes:
+    It is not uncommon to want to compute simple, "non-robust" transfer functions from matrices of cross-powers during
+    MT processing.  The preliminary estimates can be used as "features" that are in turn used to compute data weights
+    prior to robust regression.
+
+    There will probably be faster, vectorial ways to do this in future versions of the software, however, in this
+    work-in-progress module are the functions that can be used to generate TF estimates from matrices of cross-powers
+    using xarray DataArrays as the cross-power containers.
+
+    The idea here is to boil the TF estimation from cross powers down to only two equations.
+    Equations 10-15 in Sims et al. 1971 describe the 6 ways to compute the SS TF.
+    Vozoff (1991) distills this to one equation (42) with six choices for a tuple (A,B)
+    (A,B) = (Ex, Ey), (Ex, Hx), (Ex, Hy), (Ey, Hx), (Ey, Hy), (Hx, Hy)
+
+    Vozoff Notation:
+                <Ex,A*><Hy,B*> - <Ex,B*><Hy,A*>
+     Z_xx =     ------------------------------       Equation V91_42a
+                <Hx,A*><Hy,B*> - <Hx,B*><Hy,A*>
+
+
+                <Ex,A*><Hx,B*> - <Ex,B*><Hx,A*>
+     Z_xy =     ------------------------------        Equation V91_42b (modified to correct typo in publication)
+                <Hy,A*><Hx,B*> - <Hy,B*><Hx,A*>
+
+
+                <Ey,A*><Hy,B*> - <Ey,B*><Hy,A*>
+     Z_yx =     ------------------------------         Equation V91_42c
+                <Hx,A*><Hy,B*> - <Hx,B*><Hy,A*>
+
+
+                <Ey,A*><Hx,B*> - <Ey,B*><Hx,A*>
+     Z_yy =     ------------------------------          Equation V91_42d
+                <Hy,A*><Hx,B*> - <Hy,B*><Hx,A*>
+
+    Note that the Equations involving Ex (Z_xx, & Zxy) are identical to the Equations
+    involving Ey (Z_yx, & Zyy) if we swap Ex for Ey.  I.e.
+    - Equation V91_42a is the same as Equation V91_42c if we switch Ex for Ey.  Similarly,
+    - Equation V91_42b is the same as Equation V91_42d if we switch Ex for Ey.
+    - This motivates a similar formulation for the Tipper:
+
+
+                <Hz,A*><Hy,B*> - <Hz,B*><Hy,A*>
+     T_x =     ------------------------------
+                <Hx,A*><Hy,B*> - <Hx,B*><Hy,A*>
+
+                <Hz,A*><Hx,B*> - <Hz,B*><Hx,A*>
+     T_y =     ------------------------------
+                <Hy,A*><Hx,B*> - <Hy,B*><Hy,A*>
+
+    However, even V91 can be further distilled to only two equations, with an argument defining the field component
+    that is the "output" of the TF formulation Y = one of ("Ex", "Ey", "Hz")
+    I.e. by the above logic, there is one more abstraction -- instead of six functions for TF, let Ex, Ey, Hz be called say, Y.
+
+    References:
+    Sims, W.E., F.X. Bostick, and H.W. SMITH. “THE ESTIMATION OF MAGNETOTELLURIC IMPEDANCE TENSOR ELEMENTS FROM MEASURED DATA” GEOPHYSICS 36, no. 5 (1971). https://doi.org/DOI:10.1190/1.1440225.
+    @inbook{doi:10.1190/1.9781560802686.ch8,
+    author = {K. Vozoff},
[…]
