### 2024-08-16 #6832 @kkappler
Hi @diehlpk ,

Re issue [346](https://github.com/simpeg/aurora/issues/346), as blsqr states, that task is on hold until the publication process is complete, so I will get to work on the Additional author tasks.

One question about the Additional Author Task 2:

> Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

Does this mean to bump version and merge into main, or just tagging the `joss` branch? I think it is the former, but wanted to confirm.
