### 2024-08-16 63b3abb
bug fix and add test for crosspowers
