### 2024-08-16 a3a071a
add covariance_xr method and a test
