+++ aurora/time_series/xarray_helpers.py
@@ -89,6 +89,39 @@ def handle_nan(X, Y, RR, drop_dim=""):
     return X, Y, RR
 
 
+def covariance_xr(
+    X: xr.DataArray, aweights: Optional[Union[np.ndarray, None]] = None
+) -> xr.DataArray:
+    """
+    Compute the covariance matrix with numpy.cov.
+
+    Parameters
+    ----------
+    X: xarray.core.dataarray.DataArray
+        Multivariate time series as an xarray
+    aweights: array_like, optional
+        Doc taken from numpy cov follows:
+        1-D array of observation vector weights. These relative weights are
+        typically large for observations considered "important" and smaller for
+        observations considered less "important". If ``ddof=0`` the array of
+        weights can be used to assign probabilities to observation vectors.
+
+    Returns
+    -------
+    S: xarray.DataArray
+        The covariance matrix of the data in xarray form.
+    """
+
+    channels = list(X.coords["variable"].values)
+
+    S = xr.DataArray(
+        np.cov(X, aweights=aweights),
+        dims=["channel_1", "channel_2"],
+        coords={"channel_1": channels, "channel_2": channels},
+    )
+    return S
+
+
 def initialize_xrda_1d(
     channels: list,
     dtype=Optional[type],
