### 2024-08-16 145b921
update EarthScope acknowledgement
