### 2024-08-16 bc0a663
Add gh actions tests
These were suppresed during manuscript development to conserve energy.
We are re-enabling them so that the `joss` release can be shown
to pass tests with the current mt_metadata, and mth5 releases.

The tests.yml pins mt_metadata to 0.3.7, and mth5 to 0.4.4

- also updated path to ipynb moved from example to tutorial in yml
  - (and updated the filename)
