### 2024-08-16 06f417c
Update pinned mth5 version to 0.4.5
- minor changes to accomodate mt_metadata 0.3.7
