+++ .github/workflows/tests.yml
@@ -36,7 +36,7 @@ jobs:
         conda install -c conda-forge pytest pytest-cov certifi">=2017.4.17" pandoc
         pip install -r requirements-dev.txt
         pip install mt_metadata==0.3.7
-        pip install mth5==0.4.4
+        pip install mth5==0.4.5
 
     - name: Install Our Package
       run: |
