### 2024-08-17 eb774fe
pin mt_metadata and mth5 in setup.py
