+++ tests/synthetic/test_multi_run.py
@@ -2,10 +2,11 @@ import logging
 import unittest
 from aurora.config.config_creator import ConfigCreator
 from aurora.pipelines.process_mth5 import process_mth5
-from aurora.pipelines.run_summary import RunSummary
+
 from aurora.test_utils.synthetic.make_mth5_from_asc import create_test3_h5
 from aurora.test_utils.synthetic.paths import SyntheticTestPaths
-from aurora.transfer_function.kernel_dataset import KernelDataset
+from mtpy.processing.run_summary import RunSummary
+from mtpy.processing.kernel_dataset import KernelDataset
 from mth5.helpers import close_open_files
 
 
@@ -137,13 +138,8 @@ class TestMultiRunProcessing(unittest.TestCase):
         tf_cls.write(fn=xml_file_name, file_type="emtfxml")
 
 
-def main():
-    # tmp = TestMultiRunProcessing()
-    # tmp.setUp()
-    # tmp.test_works_with_truncated_run()
-    # tmp.test_all_runs()
-    unittest.main()
-
-
+# =============================================================================
+# run
+# =============================================================================
 if __name__ == "__main__":
-    main()
+    unittest.main()

[…]
