### 2024-08-18 c768d1d
removed KDS and RS
