+++ aurora/pipelines/run_summary.py
@@ -1,287 +1,0 @@
-"""
-
-This module contains the RunSummary class.
-
-This is a helper class that summarizes the Runs in an mth5.
-
-TODO: This class and methods could be replaced by methods in MTH5.
-
-Functionality of RunSummary()
-1. User can get a list of local_station options, which correspond to unique pairs
-of values: (survey,  station)
-
-2. User can see all possible ways of processing the data:
-- one list per (survey,  station) pair in the run_summary
-
-Some of the following functionalities may end up in KernelDataset:
-3. User can select local_station
--this can trigger a reduction of runs to only those that are from the local staion
-and simultaneous runs at other stations
-4. Given a local station, a list of possible reference stations can be generated
-5. Given a remote reference station, a list of all relevent runs, truncated to
-maximize coverage of the local station runs is generated
-6. Given such a "restricted run list", runs can be dropped
-7. Time interval endpoints can be changed
-
-
-Development Notes:
-    TODO: consider adding methods:
-     - drop_runs_shorter_than": removes short runs from summary
-     - fill_gaps_by_time_interval": allows runs to be merged if gaps between 
-       are short
-     - fill_gaps_by_run_names": allows runs to be merged if gaps between are 
-       short
-    TODO: Consider whether this should return a copy or modify in-place when
-    querying the df.
-
-"""
-
-# =============================================================================
-# Imports
-# =============================================================================
-import copy
-from typing import Optional, Union
-
-import pandas as pd
-from loguru import logger
-
-from aurora.pipelines import RUN_SUMMARY_COLUMNS, MINI_SUMMARY_COLUMNS
-
-import mth5
-from mth5.utils.helpers import initialize_mth5
-
-# =============================================================================
-
-
-class RunSummary:
-    """
-    Class to contain a run-summary table from one or more mth5s.
-
-    WIP: For the full MMT case this may need modification to a channel based
-    summary.
-
-
-    """
-
-    def __init__(
-        self,
-        input_dict: Optional[Union[dict, None]] = None,
-        df: Optional[Union[pd.DataFrame, None]] = None,
-    ):
-        """
-        Constructor
-
-        Parameters
-        ----------
-        kwargs
-        """
-        self.column_dtypes = [str, str, pd.Timestamp, pd.Timestamp]
-        self._input_dict = input_dict
-        self.df = df
-        self._mini_summary_columns = MINI_SUMMARY_COLUMNS
-
-    def __str__(self):
-        return str(self.mini_summary.head())
-
-    def __repr__(self):
-        return self.__str__()
-
-    @property
-    def df(self):
-        return self._df
-
-    @df.setter
-    def df(self, value):
-        """
-        Make sure the data frame is set properly with proper column names
-
-        :param value: DESCRIPTION
-        :type value: TYPE
-        :return: DESCRIPTION
-        :rtype: TYPE
-
-        """
-        if value is None:
-            self._df = None
-            return
-
-        if not isinstance(value, pd.DataFrame):
-            msg = f"Need to set df with a Pandas.DataFrame not type({type(value)})"
-            logger.error(msg)
-
-            raise TypeError(msg)
-
-        need_columns = []
-        for col in RUN_SUMMARY_COLUMNS:
-            if not col in value.columns:
-                need_columns.append(col)
-        if need_columns:
-            msg = f"DataFrame needs columns {', '.join(need_columns)}"
-            logger.error(msg)
-            raise ValueError(msg)
-        self._df = value
-
-    def clone(self):
-        """
-        2022-10-20:
[…]
