+++ aurora/transfer_function/kernel_dataset.py
@@ -1,1160 +1,0 @@
-"""
-
-This module contains a class for representing a dataset that can be processed.
-
-Development Notes:
-The KernelDataset could potentially be moved into mth5 or mtpy and used
-as the dataset description for other processing flows.
-
-Players on the stage:  One or more mth5s.
-
-Each mth5 has a mth5_obj.channel_summary dataframe which tells what data are available.
-Use a compressed view of this df with one line per acquisition run -- a "run_summary".
-
-Run_summary provides options for the local and possibly remote reference stations.
-Candidates for local station are the unique values in the station column.
-
-For any candidate station, there are some integer n runs available.
-This yields 2^n - 1 possible combinations that can be processed, neglecting any
-flagging of time intervals within any run, or any joining of runs.
-(There are actually 2**n, but we ignore the empty set, so -1)
-
-Intuition suggests default ought to be to process n runs in n+1 configurations:
-{all runs} + each run individually.  This will give a bulk answer, and bad runs can
-be flagged by comparing them.  After an initial processing, the tfs can be reviewed
-and the problematic runs can be addressed.
-
-The user can interact with the run_summary_df, selecting sub dataframes via querying,
-and in future maybe via some GUI (or a spreadsheet).
-
-The intended usage process is as follows:
- 0. Start with a list of mth5s
- 1. Extract channel_summaries from each mth5 and join them vertically
- 2. Compress to a run_summary
- 3. Stare at the run_summary_df & Select a station "S" to process
- 4. Select a non-empty set of runs for station "S"
- 5. Select a remote reference "RR", (this is allowed to be None)
- 6. Extract the sub-dataframe corresponding to acquisition_runs from "S" and "RR"
- 7. If the remote is not None:
-  - Drop the runs (rows) associated with RR that do not intersect with S
-  - Restrict start/end times of RR runs that intersect with S so overlap is complete.
-  - Restrict start/end times of S runs so that they intersect with remote
- 8. This is now a TFKernel Dataset Definition (ish). Initialize a default processing
- object and pass it this df.
- ```
-  >>> cc = ConfigCreator()
-  >>> p = cc.create_from_kernel_dataset(kernel_dataset)
-  - Optionally pass emtf_band_file=emtf_band_setup_file
- 9. Edit the Processing Config appropriately,
-
-TODO: Consider supporting a default value for 'channel_scale_factors' that is None,
-
-TODO: As of March 2023 a RunSummary is available at the station level in mth5, but
- the aurora version is still being used.  This should be merged if possible so that
- aurora uses the built-in mth5 method. -- Run Summary exists atstation level in mth5
-
-TODO: Might need to groupby survey & station, for now consider station_id  unique.
-
-"""
-
-# =============================================================================
-# Imports
-# =============================================================================
-from pathlib import Path
-import copy
-from typing import Optional, Union
-
-import pandas as pd
-from loguru import logger
-
-import mt_metadata.timeseries
-from mt_metadata.utils.list_dict import ListDict
-
-import mth5.timeseries.run_ts
-from mth5.utils.helpers import initialize_mth5
-
-from aurora.pipelines.run_summary import RunSummary
-from aurora.transfer_function import (
-    KERNEL_DATASET_DTYPE,
-    MINI_SUMMARY_COLUMNS,
-)
-
-# =============================================================================
-
-
-class KernelDataset:
-    """
-    This class is intended to work with mth5-derived channel_summary or run_summary
-    dataframes, that specify time series intervals.
-
-    Development Notes:
-    This class is closely related to (may actually be an extension of) RunSummary
-
-    The main idea is to specify one or two stations, and a list of acquisition "runs"
-    that can be merged into a "processing run". Each acquisition run can be further
-    divided into non-overlapping chunks by specifying time-intervals associated with
[…]

[…]
