+++ tests/synthetic/test_metadata_values_set_correctly.py
@@ -7,7 +7,8 @@ import logging
 import pandas as pd
 import unittest
 
-from mtpy.processing.run_summary import RunSummary
+# from mtpy-v2
+from mtpy.processing import RunSummary
 from mth5.data.make_mth5_from_asc import create_test3_h5
 from mth5.data.station_config import make_station_03
 from mth5.helpers import close_open_files
