+++ tests/synthetic/test_compare_aurora_vs_archived_emtf.py
@@ -1,5 +1,4 @@
 from aurora.pipelines.process_mth5 import process_mth5
-from mtpy.processing.run_summary import RunSummary
 from aurora.sandbox.io_helpers.zfile_murphy import read_z_file
 from mth5.data.make_mth5_from_asc import create_test1_h5
 from mth5.data.make_mth5_from_asc import create_test2_h5
@@ -14,7 +13,10 @@ from aurora.test_utils.synthetic.rms_helpers import get_expected_rms_misfit
 from aurora.transfer_function.emtf_z_file_helpers import (
     merge_tf_collection_to_match_z_file,
 )
-from mtpy.processing.kernel_dataset import KernelDataset
+
+# from mtpy-v2
+from mtpy.processing import RunSummary, KernelDataset
+
 from plot_helpers_synthetic import plot_rho_phi
 from loguru import logger
 from mth5.helpers import close_open_files
