+++ tests/synthetic/test_fourier_coefficients.py
@@ -15,8 +15,8 @@ from aurora.test_utils.synthetic.make_processing_configs import (
 )
 from aurora.test_utils.synthetic.paths import SyntheticTestPaths
 
-from mtpy.processing.run_summary import RunSummary
-from mtpy.processing.kernel_dataset import KernelDataset
+# from mtpy-v2
+from mtpy.processing import RunSummary, KernelDataset
 
 from loguru import logger
 from mth5.helpers import close_open_files
