+++ tests/parkfield/test_process_parkfield_run.py
@@ -4,8 +4,8 @@ from aurora.test_utils.parkfield.make_parkfield_mth5 import ensure_h5_exists
 from aurora.test_utils.parkfield.path_helpers import PARKFIELD_PATHS
 from aurora.transfer_function.plot.comparison_plots import compare_two_z_files
 
-from mtpy.processing.run_summary import RunSummary
-from mtpy.processing.kernel_dataset import KernelDataset
+from mtpy.processing import RunSummary, KernelDataset
+
 from loguru import logger
 from mth5.helpers import close_open_files
 
