+++ aurora/test_utils/synthetic/processing_helpers.py
@@ -20,8 +20,7 @@ def get_example_kernel_dataset():
         The kernel dataset from a synthetic, single station mth5
     """
 
-    from aurora.pipelines.run_summary import RunSummary
-    from aurora.transfer_function.kernel_dataset import KernelDataset
+    from mtpy.processing import RunSummary, KernelDataset
 
     mth5_path = create_test1_h5(force_make_mth5=False)
 
@@ -47,8 +46,7 @@ def tf_obj_from_synthetic_data(
 
     """
     from aurora.config.config_creator import ConfigCreator
-    from aurora.pipelines.run_summary import RunSummary
-    from aurora.transfer_function.kernel_dataset import KernelDataset
+    from mtpy.processing import RunSummary, KernelDataset
 
     run_summary = RunSummary()
     run_summary.from_mth5s(list((mth5_path,)))
