### 2024-08-20 #6832 @diehlpk
@kkappler please fix the invalid DOI
