+++ paper/paper.bib
@@ -112,7 +112,7 @@
   pages={105102},
   year={2022},
   issn={0098-3004},
-  doi={https://doi.org/10.1016/j.cageo.2022.105102},
+  doi={10.1016/j.cageo.2022.105102},
   publisher={Elsevier}
 }
 
