### 2024-08-21 71760d7
Confirm that issue #85 is OK
- add some doc, and typehints while passing over the logic
