+++ aurora/sandbox/triage_metadata.py
@@ -2,13 +2,14 @@
     This module contains various helper functions that were used to fix errors in metadata.
 """
 
+from mt_metadata.timeseries import Experiment
 from mt_metadata.timeseries.filters.helper_functions import MT2SI_ELECTRIC_FIELD_FILTER
 from mt_metadata.timeseries.filters.helper_functions import MT2SI_MAGNETIC_FIELD_FILTER
-
 from loguru import logger
+import mth5.groups
 
 
-def triage_mt_units_electric_field(experiment):
+def triage_mt_units_electric_field(experiment: Experiment) -> Experiment:
     """
     Updates an mth5 experiment with filter information
 
@@ -19,10 +20,13 @@ def triage_mt_units_electric_field(experiment):
 
     Parameters
     ----------
-    experiment ;
+    experiment: mt_metadata.timeseries.Experiment
+        mt_metadata representation of station XML
 
     Returns
     -------
+    experiment: mt_metadata.timeseries.Experiment
+        Updated experiment with additional filter to make electric field into MT Units (mV/km)
 
     """
     logger.info(
@@ -42,7 +46,7 @@ def triage_mt_units_electric_field(experiment):
     return experiment
 
 
-def triage_mt_units_magnetic_field(experiment):
+def triage_mt_units_magnetic_field(experiment: Experiment) -> Experiment:
     """
     Updates an mth5 experiment with filter information
 
@@ -52,10 +56,13 @@ def triage_mt_units_magnetic_field(experiment):
 
     Parameters
     ----------
-    experiment ;
+    experiment: mt_metadata.timeseries.Experiment
+        mt_metadata representation of station XML
 
     Returns
     -------
+    experiment: mt_metadata.timeseries.Experiment
+        Updated experiment with additional filter to make magnetic field into MT Units (nT)
 
     """
     logger.info(
@@ -75,11 +82,11 @@ def triage_mt_units_magnetic_field(experiment):
     return experiment
 
 
-def triage_missing_coil_hollister(experiment):
+def triage_missing_coil_hollister(experiment: Experiment) -> Experiment:
     """
     Fixes missing metadata for Hollister station
 
-    One off for hollister missing hy metadata for no reason I can tell
+    One-off for hollister missing hy metadata in NCEDC archive.
 
     Parameters
     ----------
@@ -89,27 +96,29 @@ def triage_missing_coil_hollister(experiment):
     -------
 
     """
-    survey = experiment.surveys[0]
-    stations = survey.stations
-    for station in stations:
-        if station.id == "SAO":
-            runs = station.runs
-            for run in runs:
-                channels = run.channels
-                logger.info(channels)
-                for channel in channels:
-                    logger.info(channel.id)
+    msg = "missing coil not fixed in metadata"
+    raise NotImplementedError(msg)
+    # survey = experiment.surveys[0]
+    # stations = survey.stations
+    # for station in stations:
+    #     if station.id == "SAO":
+    #         runs = station.runs
+    #         for run in runs:
+    #             channels = run.channels
+    #             logger.info(channels)
+    #             for channel in channels:
+    #                 logger.info(channel.id)
+    #
+    #     # station = stations[i_station]
+    #     runs = station.runs[0]
+    #     logger.info("help")
 
-        # station = stations[i_station]
-        runs = station.runs[0]
-        logger.info("help")
 
-
-def triage_run_id(expected_run_id, run_obj):
+def triage_run_id(expected_run_id: str, run_obj: mth5.groups.RunGroup) -> None:
     """
     Fixes metadata from an old version of MTH5.
 
-    This situation was encounterd during the Musgraves processing in 2023 HPC workshopl
+    This situation was encountered during the Musgraves processing for 2023 HPC workshop.
[…]
