### 2024-08-21 8e2d50c
Minor changes
- Update synthetic data ipynb to show two ways of storing FCs in mth5
- add some doc/typehinting to fourier_coefficients.py
