+++ aurora/pipelines/fourier_coefficients.py
@@ -69,47 +69,51 @@ from aurora.pipelines.time_series_helpers import calibrate_stft_obj
 from aurora.pipelines.time_series_helpers import prototype_decimate
 from aurora.pipelines.time_series_helpers import run_ts_to_stft_scipy
 from loguru import logger
+from mth5.mth5 import MTH5
 from mth5.utils.helpers import path_or_mth5_object
 from mt_metadata.transfer_functions.processing.fourier_coefficients import (
     Decimation as FCDecimation,
 )
-from typing import List, Union
+from typing import List, Optional, Union
 
 # =============================================================================
 GROUPBY_COLUMNS = ["survey", "station", "sample_rate"]
 
 
 def fc_decimations_creator(
-    initial_sample_rate, decimation_factors=None, max_levels=6, time_period=None
-):
+    initial_sample_rate: float,
+    decimation_factors: Optional[Union[list, None]] = None,
+    max_levels: Optional[int] = 6,
+    time_period: mt_metadata.timeseries.TimePeriod = None,
+) -> list:
     """
-    Based on the number of samples in the runts, we can compute the maximum number of valid decimation levels.
-    This would re-use code in processing summary ... or we could just decimate until we cant anymore?
 
-    You can provide something like: decimation_info = {0: 1.0, 1: 4.0, 2: 4.0, 3: 4.0}
+    Creates mt_metadata FCDecimation objects that parameterize Fourier coefficient decimation levels.
 
     Note 1:  This does not yet work through the assignment of which bands to keep.  Refer to
     mt_metadata.transfer_functions.processing.Processing.assign_bands() to see how this was done in the past
 
-    Args:
-        initial_sample_rate: float
-            Sample rate of the "level0" data -- usually the sample rate during field acquisition.
-        decimation_factors: list (or other iterable)
-            The decimation factors that will be applied at each FC decimation level
-        max_levels: int
-            The maximum number of decimation levels to allow
-        time_period:
+    Parameters
+    ----------
+    initial_sample_rate: float
+        Sample rate of the "level0" data -- usually the sample rate during field acquisition.
+    decimation_factors: list (or other iterable)
+        The decimation factors that will be applied at each FC decimation level
+    max_levels: int
+        The maximum number of decimation levels to allow
+    time_period:
 
-    Returns:
-        fc_decimations: list
-            Each element of the list is an object of type
-            mt_metadata.transfer_functions.processing.fourier_coefficients.Decimation,
-            (a.k.a. FCDecimation).
+    Returns
+    -------
+    fc_decimations: list
+        Each element of the list is an object of type
+        mt_metadata.transfer_functions.processing.fourier_coefficients.Decimation,
+        (a.k.a. FCDecimation).
 
-            The order of the list corresponds the order of the cascading decimation
-              - No decimation levels are omitted.
-              - This could be changed in future by using a dict instead of a list,
-              - e.g. decimation_factors = dict(zip(np.arange(max_levels), decimation_factors))
+        The order of the list corresponds the order of the cascading decimation
+          - No decimation levels are omitted.
+          - This could be changed in future by using a dict instead of a list,
+          - e.g. decimation_factors = dict(zip(np.arange(max_levels), decimation_factors))
 
     """
     if not decimation_factors:
@@ -122,19 +126,19 @@ def fc_decimations_creator(
     # See Note 1
     fc_decimations = []
     for i_dec_level, decimation_factor in enumerate(decimation_factors):
-        dd = FCDecimation()
-        dd.decimation_level = i_dec_level
-        dd.id = f"{i_dec_level}"
[…]
