### 2024-08-21 99b82c5
update actions versions to 3
- was 4 on codecov, but nothing has uploaded to simpeg codecov in 4 months
- was 2 on github actions, make 3 to be consistent
