+++ .github/workflows/tests.yml
@@ -22,7 +22,7 @@ jobs:
 #        python-version: ["3.10", ] 
 
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v3
 
     - name: Setup Miniconda
       uses: conda-incubator/setup-miniconda@v2.1.1
@@ -77,7 +77,7 @@ jobs:
         pytest -s -v --cov=./ --cov-report=xml --cov=aurora
 
     - name: Upload coverage reports to Codecov
-      uses: codecov/codecov-action@v4
+      uses: codecov/codecov-action@v3
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
         fail_ci_if_error: false
