+++ aurora/test_utils/synthetic/make_mth5_from_asc.py
@@ -1,495 +1,26 @@
 # -*- coding: utf-8 -*-
 """
-Created on Fri Jun 25 16:03:21 2021
 
-@author: jpeacock
-
-This module is concerned with creating mth5 files from the synthetic test data
- that originally came from EMTF -- test1.asc and test2.asc.  Each ascii file
- represents five channels of data sampled at 1Hz at a synthetic station.
-
-Mirroring the original ascii files are:
-data/test1.h5
-data/test2.h5
-data/test12rr.h5
-
-Also created are some files with the same data but other channel_nomenclature schemes:
-data/test12rr_LEMI34.h5
-data/test1_LEMI12.h5
-
-- 20231103: Added an 8Hz up-sampled version of test1.  No spectral content was added
- so the band between the old and new Nyquist frequencies is bogus.
-
-Notes: Work in progress -- this module is being migrated to MTH5.
+Notes: This module has been migrated to MTH5.
 
 """
-# import inspect
-import numpy as np
-import pandas as pd
-import pathlib
-import scipy.signal as ssig
 
-from aurora.test_utils.synthetic.paths import SyntheticTestPaths
-from aurora.test_utils.synthetic.station_config import make_filters
-from aurora.test_utils.synthetic.station_config import make_station_01
-from aurora.test_utils.synthetic.station_config import make_station_02
-from aurora.test_utils.synthetic.station_config import make_station_03
-from aurora.test_utils.synthetic.station_config import make_station_04
-from loguru import logger
-from mth5.mth5 import MTH5
-from mth5.timeseries import ChannelTS, RunTS
-from mt_metadata.transfer_functions.processing.aurora import ChannelNomenclature
-from mt_metadata.timeseries import Electric
-from mt_metadata.timeseries import Magnetic
-from mth5.utils.helpers import add_filters
-
-np.random.seed(0)
-
-synthetic_test_paths = SyntheticTestPaths()
-MTH5_PATH = synthetic_test_paths.mth5_path
-
-
-def create_run_ts_from_synthetic_run(run, df, channel_nomenclature="default") -> RunTS:
-    """
-    Loop over channels of synthetic data in df and make ChannelTS objects.
-
-    Parameters
-    ----------
-    run: aurora.test_utils.synthetic.station_config.SyntheticRun
-        One-off data structure with information mth5 needs to initialize
-        Specifically sample_rate, filters,
-    df : pandas.DataFrame
-        time series data in columns labelled from ["ex", "ey", "hx", "hy", "hz"]
-    channel_nomenclature : string
-        Keyword corresponding to channel nomenclature mapping in CHANNEL_MAPS variable
-        from channel_nomenclature.py module in mt_metadata.
-        Supported values include ['default', 'lemi12', 'lemi34', 'phoenix123']
-
-    Returns
-    -------
-    runts: RunTS
-        MTH5 run time series object, data and metadata bound into one.
-    """
-
-    channel_nomenclature_obj = ChannelNomenclature()
-    channel_nomenclature_obj.keyword = channel_nomenclature
-    ch_list = []
-    for col in df.columns:
-
-        data = df[col].values
-        if col in channel_nomenclature_obj.ex_ey:
-            channel_metadata = Electric()
-            channel_metadata.component = col
-            channel_metadata.units = "millivolts per kilometer"
-            chts = ChannelTS(
-                channel_type="electric",
-                data=data,
-                channel_metadata=channel_metadata.to_dict(),
-            )
-
-            # add metadata to the channel here
-            chts.channel_metadata.dipole_length = 50
-            if col == channel_nomenclature_obj.ey:
-                chts.channel_metadata.measurement_azimuth = 90.0
-
[…]
