### 2024-08-22 7844053
point synthetic data at mth5 data path
