### 2024-08-22 193feff
replace SyntheticTestPaths with mth5 version (#320)
