### 2024-08-22 dd8c766
set extra_fft_detrend to constant
- the test was passing but unneeded errors were occurring
- the value of detrend must be "linear" or "constant"
