+++ .github/workflows/tests.yml
@@ -22,7 +22,7 @@ jobs:
 #        python-version: ["3.10", ] 
 
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v4
 
     - name: Setup Miniconda
       uses: conda-incubator/setup-miniconda@v2.1.1
@@ -76,9 +76,17 @@ jobs:
         # pytest -s -v tests/test_general_helper_functions.py 
         pytest -s -v --cov=./ --cov-report=xml --cov=aurora
 
+# Commenting this old codecov - trying to fix #351
+#    - name: Upload coverage reports to Codecov
+#      uses: codecov/codecov-action@v3
+#      with:
+#        token: ${{ secrets.CODECOV_TOKEN }}
+#        fail_ci_if_error: false
+#        flags: tests
     - name: Upload coverage reports to Codecov
-      uses: codecov/codecov-action@v3
-      with:
+      uses: codecov/codecov-action@v4
+      env:
+        CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
         token: ${{ secrets.CODECOV_TOKEN }}
         fail_ci_if_error: false
         flags: tests
